Land model

Example Documentation

Canopy RT module

Leaf Bio Structure:

Land.CanopyRT.leafbio — Type

leafbio

Fields

N

Leaf structure parameter (-)

Cab

Chlorophyll a+b content (µg cm-2)

Car

Carotenoid content (µg cm-2)

Ant

Anthocynanin content (µg cm-2)

Cs

Senescent material fraction (-)

Cw

Equivalent water thickness (cm)

Cm

Dry matter content (dry leaf mass per unit area) (g cm-2)

Cx

Fractionation between Zeaxanthin and Violaxanthin in Car (1=all Zeaxanthin) (-)

ρ_LW

Broadband thermal reflectance (-)

τ_LW

Broadband thermal transmission (-)

fqe

Leaf fluorescence efficiency (Fo standard)

ρ_SW

shortwave leaf reflectance (-)

τ_SW

shortwave leaf transmission (-)

kChlrel

relative absorbtion by Chlorophyll+Car (-)

kChlrel_old

relative absorbtion by Chlorophyll (-)

Mb

Fluorescence excitation matrix backwards (-)

Mf

Fluorescence excitation matrix forwards (-)

Canopy Structure

Land.CanopyRT.struct_canopy — Type

struct_canopy

Fields

nlayers

number of canopy layers

LAI

Leaf Area Index

Ω

Clumping factor

clump_a

Structure factor a

clump_b

Structure factor b

leafwidth

Leaf width

hc

Vegetation height

LIDFa

Leaf Inclination

LIDFb

Variation in leaf inclination

hot

HotSpot parameter (still need to check!)

height

Canopy height (m)

z0m

Canopy roughness (m)

z0h

d

Canopy displacement height (m)

Cd

m/sqrt(s) turbulent transfer coefficient

lazitab

lidf

xl

dx

Canopy Radiation Structure

Land.CanopyRT.struct_canopyRadiation — Type

struct_canopyRadiation

Fields

intEout

integrated TOC outgoing flux [W m^-2]

incomingPAR

incident spectrally integrated total PAR [moles m^-2 s^-1]

incomingPAR_direct

incident spectrally integrated direct PAR [moles m^-2 s^-1]

incomingPAR_diffuse

incident spectrally integrated diffuse PAR [moles m^-2 s^-1]

RnSoil_diffuse

net radiation of shaded soil [W m^-2]

RnSoil_direct

net Short-wave radiation of sunlit soil [W m^-2]

RnSoil

net Short-wave radiation of soil (shaded + sunlit) [W m^-2]

RnSoilLW

net long-wave radiation of soil (shaded + sunlit) [W m^-2]

absPAR_shade

net PAR of shaded leaves [moles m^-2 s^-1]

absPAR_shadeCab

net PAR by Cab+Car of shaded leaves [moles m^-2 s^-1]

Lo

Short-wave TOC outgoing radiance in observation direction [mW m^-2 μm^-1 sr^-1]

Eout

Short-wave TOC outgoing radiation [mW m^-2 μm^-1]

alb_obs

Short-wave Albedo in viewing direction

alb_direct

Short-wave Albedo for direct incoming radiation

alb_diffuse

Short-wave Albedo for diffuse incoming radiation

E_up

upwelling diffuse short-wave radiation within canopy [mW m^-2 μm^-1]

E_down

downwelling diffuse short-wave radiation within canopy [mW m^-2 μm^-1]

netSW_sunlit

net absorbed direct radiation in each layer [mW m^-2 μm^-1]

netSW_shade

net absorbed diffuse radiation in each layer [mW m^-2 μm^-1]

intNetSW_sunlit

spectrally integrated net absorbed direct radiation in each layer [W m^-2)]

intNetSW_shade

spectrally integrated net absorbed diffuse radiation in each layer [W m^-2)]

intNetLW_sunlit

spectrally integrated net absorbed direct radiation in each layer [W m^-2)]

intNetLW_shade

spectrally integrated net absorbed diffuse radiation in each layer [W m^-2)]

absPAR_sun

net PAR of sunlit leaves [mol m^-2 s^-1]

absPAR_sunCab

net PAR by Cab+Car of sunlit leaves [mol m^-2 s^-1]

T_sun3D

Leaf temperature (sunlit) (K)

T_sun

Leaf temperature (sunlit) (K)

ϕ_sun

Fluorescence yield for sunlit leaves

T_shade

Leaf temperature (shaded) (K)

ϕ_shade

Fluorescence yield for shaded leaves

GPP_sun

GPP of sunlit leaves [μmol m^-2 s^-1]

GPP_shade

GPP of shaded leaves [μmol m^-2 s^-1]

gs_shade

gs of shaded leaves [mol m^-2 s^-1]

gs_sun

gs of sunlit leaves [mol m^-2 s^-1]

H_shade

Sensible Heat flux H of shaded leaves [W m-2]

H_sun

Sensible Heat flux H of sunlit leaves [W m-2]

LE_shade

Latent Heat flux LE of shaded leaves [W m-2]

LE_sun

Latent Heat flux LE of sunlit leaves [W m-2]

ψl_shade

Leaf water potential of shaded leaves [MPa]

ψl_sun

Leaf water potential of sunlit leaves [MPa]

Cc_shade

Cc of shaded leaves [µmol/mol]

Cc_sun

Cc of sunlit leaves [µmol/mol]

NPQ_shade

NPQ of shaded leaves [-]

NPQ_sun

NPQ of sunlit leaves [-]

SIF_hemi

Hemispheric total outgoing SIF flux [mW m^-2 μm^-1])

SIF_obs

Observer-direction outgoing SIF radiance (mW m^-2 μm^-1 sr^-1))

SIF_obs_sunlit

Observer-direction outgoing SIF radiance, sunlit leaves (mW m^-2 μm^-1 sr^-1))

SIF_obs_shaded

Observer-direction outgoing SIF radiance, shaded leaves (mW m^-2 μm^-1 sr^-1))

SIF_obs_scattered

Observer-direction outgoing SIF radiance, scattered (mW m^-2 μm^-1 sr^-1))

SIF_obs_soil

Observer-direction outgoing SIF radiance, soil-reflected (mW m^-2 μm^-1 sr^-1))

SIF_sum

Total SIF sum of layer sources [mW m^-2 μm^-1])

Canopy Optical Property Structure

Land.CanopyRT.struct_canopyOptProps — Type

struct_canopyOptProps

Fields

sdb

Solar -> Diffuse backscatter weight

sdf

Solar -> Diffuse forward scatter weight

dob

Diffuse -> Directional backscatter weight

dof

Diffuse -> Directional forward scatter weight

ddb

Diffuse -> Diffuse backscatter weight

ddf

Diffuse -> Diffuse forward scatter weight

ks

Solar beam extinction coefficient weight

ko

Outgoing beam extinction coefficient weight

bf

sob

weight of specular2directional backscatter coefficient

sof

weight of specular2directional forward coefficient

fs

conversion factor fs to compute irradiance on inclined leaf

absfs

abs(fs)

absfsfo

abs(fs*fo)

fsfo

fs*fo

fo

conversion factor fo for angle towards observer (not sun like fs)

cosΘ_l

Cosine of leaf azimuths

cos2Θ_l

cos of leaf azimuth sqared

Ps

Probability of directly viewing a leaf in solar direction

Po

Probability of directly viewing a leaf in viewing direction

Pso

Bi-directional probability of directly viewing a leaf (solar->canopy->viewing)

sigb

diffuse backscatter scattering coefficient for diffuse incidence

sigf

diffuse forward scattering coefficient for diffuse incidence

sb

diffuse backscatter scattering coefficient for specular incidence

sf

diffuse forward scattering coefficient for specular incidence

vb

directional backscatter scattering coefficient for diffuse incidence

vf

directional forward scattering coefficient for diffuse incidence

w

bidirectional scattering coefficent (directional-directional)

a

attenuation

Xsd

Effective layer transmittance (direct->diffuse)

Xdd

Effective layer transmittance (diffuse->diffuse)

R_sd

Effective layer reflectance (direct->diffuse)

R_dd

Effective layer reflectance (diffuse->diffuse)

Es_

Solar direct radiation per layer)

Photosynthesis module

Leaf Params:

Land.Photosynthesis.leaf_params — Type

leaf_params{FT<:AbstractFloat}
A structure, which contains all parameters for photosynthesis calculations of a single leaf

Fields

α

broadband shortwave albedo

ε

longwave emissivity

LMA

DRY leaf mass area [kg/m2]

RWC

leaf relative water content [-]

Cleaf

leaf specific heat capacity [J/kg/K]

Chloroplast_rel_volume

Chloroplast volume airspace [m^3 per pore]

PSII_frac

Fraction of absorbed light used by PSII ETR

Kf

Rate constant for fluorescence (const)

Kd

Rate constant for thermal dissipation

Kp

Rate constant for photochemistry (all reaction centers open)

Kn

NPQ rate constant

Kn_ss

maxPSII

max PSII yield

Fo

dark-adapted fluorescence yield Fo (Kp=max)

Fo′

light-adapted fluorescence yield Fo' (Kp=max)

Fm

dark adapted Fm yield (Kp=0)

Fm′

light adapted Fm' yield (Kp=0)

ϕs

steady-state (light-adapted) yield Ft (aka Fs)

φ

PSII yield

qQ

Photochemical quenching

qE

NPQ

Non-Photochemical quenching

dynamic_state

Dynamic Steady state [Bool]

effcon

Efficiency of photosynthesis light reactions[mol CO₂/mol e-]

CO2_per_electron

Total efficiency, incl. photorespiration [mol CO₂/mol e-]

Ja

Actual electron transport rate [μmol m-2 s-1]

o₂

O₂ mixing ratio [mol/mol]

T

Leaf temperature [K)]

T_old

Previous Leaf temperature [K]

esat

Saturation vapor pressure at leaf temperature [Pa]

desat

Derivative of saturation vapor pressure at leaf temperature wrt T [Pa/K]

θ_j

Curvature parameter:

gleaf

total leaf conductance [mmol H₂O /m2/s]

gm

Mesophyll conductance [mmol H₂O /m2/s]

gs

Stomatal conductance [mmol H₂O /m2/s]

gs_ss

Steady state Stomatal conductance [mmol H₂O /m2/s]

Ag

Gross Photosynthetic Rate [μmol/m2/s]

An

Net Photosynthetic Rate [μmol/m2/s]

Ac

Rubisco Limited Photosynthetic Rate [μmol/m2/s]

Aj

Light Limited Photosynthetic Rate [μmol/m2/s]

Ap

Product Limited Photosynthetic Rate [μmol/m2/s]

H

Sensible Heat Flux [W/m2]

LE

Latent Heat Flux [W/m2]

Rn

Net Radiation Balance [W/m2]

Sap

Sap flow [?]

APAR

Absorbed Photosynthetically Active Radiation [μmol/m2/s]

Kc

Michaelis-Menten constant for CO₂ [Pa]

Ko

Michaelis-Menten constant for O₂ [Pa]

Kpep

Michaelis-Menten constant for PEP carboxylase [Pa]

Γstar

CO₂ compensation point [Pa]

Vcmax25

Leaf maximum carboxylation rate at 25C [μmol/m2/s]

Vpmax25

Leaf maximum PEP carboxylation rate at 25C [μmol/m2/s]

Jmax25

Maximum electron transport rate at 25C [μmol/m2/s]

Rd25

Leaf respiration rate at 25C [μmol/m2/s]

Vcmax

Actual Leaf maximum carboxylation rate [μmol/m2/s]

Vpmax

Actual Leaf maximum PEP carboxylase rate [μmol/m2/s]

Jmax

Actual Maximum electron transport rate [μmol/m2/s]

Rd

Actual Leaf respiration rate [μmol/m2/s]

Cs

CO₂ concentration at leaf surface [μmol/mol]

Cc

CO₂ concentration in chloroplast `[μmol/mol]

RH

relative humidity at the surface of the leaf [-]

VPD

Vapor pressure difference across leaf interface [Pa]

Je

Electron transport rate [μmol/m2/s], light limited only

Je_pot

Potential Electron Transport Rate [μmol/m2/s]

psi_l

leaf water potential [Pa]

psi_l50

leaf water potential at 50% drop in conductivity [Pa]

kmax

maximum leaf-xylem conductivity [m/s]

kx

actual xylem conductivity [m/s]

kleaf

leaf level hydraulic conductivity [m/s]

ck

slope of Weibull curve [m/s/Pa]

height

tree height [m]

Ctree

tree capacitance [kg/m3/Pa]

use_colim

z0m

z0h

d

dleaf

LAI

Cd

Leaf Rate Limiting Steps:

Land.Photosynthesis.max_carboxylation_rate! — Function

max_carboxylation_rate!(model::VcmaxCLM, leaf)

Calculates the maximum carboxylation rate (Vcmax) at the leaf temperature as in CLM, given

model One VcmaxCLM

leaf One leaf_params

max_carboxylation_rate(model::VcmaxBernacchi, leaf)

Calculates the maximum carboxylation rate (Vcmax) at the leaf temperature (Bernacchi 2001 paper)

Land.Photosynthesis.max_electron_transport_rate! — Function

max_electron_transport_rate(model::JmaxCLM, leaf)

Calculates the potential max_electron transport rate (Jmax) at the leaf temperature

Land.Photosynthesis.michaelis_menten_constants! — Function

michaelis_menten_constants(model::JmaxCLM, leaf)

Calculates the leaf temperature adjusted Michaelis Menten constants

Missing docstring.

Missing docstring for Land.Photosynthesis.leaf_respiration!. Check Documenter's build log for details.