CanopyLayers

Structures

Land.CanopyLayers.Canopy4RT — Type

mutable struct Canopy4RT

A canopy struct for the radiation transfer module

Fields

nLayer

Number of canopy layers

LAI

Leaf Area Index

Ω

Clumping factor

clump_a

Structure factor a

clump_b

Structure factor b

leaf_width

Leaf width

hc

Vegetation height

LIDFa

Leaf Inclination

LIDFb

Variation in leaf inclination

hot

HotSpot parameter (still need to check!)

iLAI

LAI in the ith layer

height

Canopy height [m]

z0m

Canopy roughness [m]

z0h

Tree roughtnes [m]

d

Canopy displacement height [m]

Cd

m/sqrt(s) turbulent transfer coefficient

litab

List of mean inclination angles [°]

litab_bnd

List of inclination angle boundaries [°]

lazitab

List of mean azimuth angles [°]

cos_ttlo

Cosine of lazitab

cos_philo

Cosine of lazitab - psi (relative azimuth angle), update with time

cos_ttli

Cosine of litab

sin_ttli

Sine of litab

vol_scatt

Container for volome scatter function

lidf

Inclination angles weight distribution

xl

List of level location (level = layer + 1)

dx

1/nLayer

nAzi

Number of azimuth angles

nIncl

Number of inclination angles

Land.CanopyLayers.CanopyOpticals — Type

mutable struct CanopyOpticals{FT}

A struct for canopy optical properties

Fields

nAzi

Number of azimuth angles

nIncl

Number of inclination agles

nLayer

Number of canopy layers

nWL

Number of wave lengths

sdb

Solar -> Diffuse backscatter weight

sdf

Solar -> Diffuse forward scatter weight

dob

Diffuse -> Directional backscatter weight

dof

Diffuse -> Directional forward scatter weight

ddb

Diffuse -> Diffuse backscatter weight

ddf

Diffuse -> Diffuse forward scatter weight

ks

Solar beam extinction coefficient weight

ko

Outgoing beam extinction coefficient weight

bf

?

sob

Weight of specular2directional backscatter coefficient

sof

Weight of specular2directional forward coefficient

Ps

Probability of directly viewing a leaf in solar direction

Po

Probability of directly viewing a leaf in viewing direction

Pso

Bi-directional probability of directly viewing a leaf (solar->canopy->viewing)

fs

conversion factor fs to compute irradiance on inclined leaf

absfs

abs(fs)

absfsfo

abs(fs*fo)

fsfo

fs*fo

fo

conversion factor fo for angle towards observer (not sun like fs)

absfo

abs(fo)

cosΘ_l

Cosine of leaf azimuths

cos2Θ_l

cos of leaf azimuth sqared

sigb

diffuse backscatter scattering coefficient for diffuse incidence

sigf

diffuse forward scattering coefficient for diffuse incidence

sb

diffuse backscatter scattering coefficient for specular incidence

sf

diffuse forward scattering coefficient for specular incidence

vb

directional backscatter scattering coefficient for diffuse incidence

vf

directional forward scattering coefficient for diffuse incidence

w

bidirectional scattering coefficent (directional-directional)

a

attenuation

Xsd

Effective layer transmittance (direct->diffuse)

Xdd

Effective layer transmittance (diffuse->diffuse)

R_sd

Effective layer reflectance (direct->diffuse)

R_dd

Effective layer reflectance (diffuse->diffuse)

Es_

Solar direct radiation per layer)

Land.CanopyLayers.CanopyRads — Type

mutable struct CanopyRads{FT}

A struct for canopy radiation information

Fields

nAzi

Number of azimuth angles

nIncl

Number of inclination agles

nLayer

Number of canopy layers

nLevel

Number of canopy levels

nWL

Number of wave lengths

nWLF

Number of wave lengths for SIF

intEout

Integrated TOC outgoing flux [W m⁻²]

incomingPAR

Incident spectrally integrated total PAR [mol m⁻² s⁻¹]

incomingPAR_direct

Incident spectrally integrated direct PAR [mol m⁻² s⁻¹]

incomingPAR_diffuse

Incident spectrally integrated diffuse PAR [mol m⁻² s⁻¹]

RnSoil_diffuse

Net radiation of shaded soil [W m⁻²]

RnSoil_direct

Net Short-wave radiation of sunlit soil [W m⁻²]

RnSoil

Net Short-wave radiation of soil (shaded + sunlit) [W m⁻²]

RnSoilLW

Net long-wave radiation of soil (shaded + sunlit) [W m⁻²]

absPAR_shade

Net PAR of shaded leaves [mol m⁻² s⁻¹]

absPAR_shadeCab

Net PAR by Cab+Car of shaded leaves [moles m⁻² s⁻¹]

intNetSW_sunlit

Spectrally integrated net absorbed direct radiation in each layer [W m⁻²)]

intNetSW_shade

Spectrally integrated net absorbed diffuse radiation in each layer [W m⁻²)]

intNetLW_sunlit

Spectrally integrated net absorbed direct radiation in each layer [W m⁻²)]

intNetLW_shade

Spectrally integrated net absorbed diffuse radiation in each layer [W m⁻²)]

T_sun

Leaf temperature (sunlit) [K]

T_shade

Leaf temperature (shaded) [K]

ϕ_shade

Fluorescence yield for shaded leaves

H_shade

Sensible Heat flux H of shaded leaves [W m⁻²]

LE_shade

Latent Heat flux LE of shaded leaves [W m⁻²]

NPQ_shade

NPQ of shaded leaves

GPP_shade

GPP of shaded leaves [μmol m⁻² s⁻¹]

gs_shade

gs of shaded leaves [mol m⁻² s⁻¹]

ψl_shade

Leaf water potential of shaded leaves [MPa]

Cc_shade

Cc of shaded leaves [µmol/mol]

Pi_shade

internal CO₂ concentration of shaded leaves [µmol/mol]

Lo

Short-wave TOC outgoing radiance in observation direction [mW m⁻² nm⁻¹ sr⁻¹]

Eout

Short-wave TOC outgoing radiation [mW m⁻² nm⁻¹]

alb_obs

Short-wave Albedo in viewing direction

alb_direct

Short-wave Albedo for direct incoming radiation

alb_diffuse

Short-wave Albedo for diffuse incoming radiation

E_up

Upwelling diffuse short-wave radiation within canopy [mW m⁻² nm⁻¹]

E_down

Downwelling diffuse short-wave radiation within canopy [mW m⁻² nm⁻¹]

netSW_sunlit

Net absorbed direct radiation in each layer [mW m⁻² nm⁻¹]

netSW_shade

net absorbed diffuse radiation in each layer [mW m⁻² nm⁻¹]

absPAR_sun

net PAR of sunlit leaves [mol m⁻² s⁻¹]

absPAR_sunCab

net PAR by Cab+Car of sunlit leaves [mol m⁻² s⁻¹]

T_sun3D

Leaf temperature (sunlit) [K]

ϕ_sun

Fluorescence yield for sunlit leaves

H_sun

Sensible Heat flux H of sunlit leaves [W m⁻²]

LE_sun

Latent Heat flux LE of sunlit leaves [W m⁻²]

NPQ_sun

NPQ of sunlit leaves

GPP_sun

GPP of sunlit leaves [μmol m⁻² s⁻¹]

gs_sun

gs of sunlit leaves [mol m⁻² s⁻¹]

ψl_sun

Leaf water potential of sunlit leaves [MPa]

Cc_sun

Cc of sunlit leaves [µmol/mol]

Pi_sun

Internal CO₂ concentration of sunlit leaves [µmol/mol]

SIF_hemi

Hemispheric total outgoing SIF flux [mW m⁻² nm⁻¹])

SIF_obs

Observer-direction outgoing SIF radiance (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_sunlit

Observer-direction outgoing SIF radiance, sunlit leaves (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_shaded

Observer-direction outgoing SIF radiance, shaded leaves (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_scattered

Observer-direction outgoing SIF radiance, scattered (mW m⁻² nm⁻¹ sr⁻¹))

SIF_obs_soil

Observer-direction outgoing SIF radiance, soil-reflected (mW m⁻² nm⁻¹ sr⁻¹))

SIF_sum

Total SIF sum of layer sources [mW m⁻² nm⁻¹])

Land.CanopyLayers.IncomingRadiation — Type

mutable struct IncomingRadiation{FT}

Incoming radiation information

Fields

E_direct

Direct incoming radiation [mW m⁻² nm⁻¹]

E_diffuse

Diffuse incoming radiation [mW m⁻² nm⁻¹]

Land.CanopyLayers.LeafBios — Type

mutable struct LeafBios{FT}

A struct of leaf biological parameters

Fields

nWL

Number of wave length

nWLE

Number of wave length for excitation

nWLF

Number of wave length for SIF

N

Leaf structure parameter

Cab

Chlorophyll a+b content [µg cm⁻²]

Car

Carotenoid content [µg cm⁻²]

Ant

Anthocynanin content [µg cm⁻²]

Cs

Senescent material fraction

Cw

Equivalent water thickness [cm]

Cm

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

Cx

Fractionation between Zeaxanthin and Violaxanthin in Car (1=all Zeaxanthin) (-)

ρ_LW

Broadband thermal reflectance (-)

τ_LW

Broadband thermal transmission (-)

fqe

Leaf fluorescence efficiency (Fo standard)

ρ_SW

Shortwave leaf reflectance

τ_SW

Shortwave leaf transmission

kChlrel

Relative absorbtion by Chlorophyll+Car

kChlrel_old

Relative absorbtion by Chlorophyll

Mb

Fluorescence excitation matrix backwards

Mf

Fluorescence excitation matrix forwards

ndub

Doubling adding layers

Land.CanopyLayers.LeafOpticals — Type

mutable struct LeafOpticals{FT}

Struct for leaf optical properties

Fields

nr

Km

Kab

Kant

Kcar

Kw

KBrown

phi

KcaV

KcaZ

lambda

Wave length [nm], same as WL in WaveLengths`

Land.CanopyLayers.RTDimensions — Type

mutable struct RTDimensions

Struct that stores matrix dimension information

Fields

nAzi

Number of azimuth angles

nIncl

Number of inclination agles

nLayer

Number of canopy layers

nLevel

Number of canopy layer boundaries nLayer+1

nPAR

Number of PAR wave lengths

nWL

Number of wave lengths

nWLE

Number of wave length for excitation

nWLF

Number of wave lengths for SIF

Land.CanopyLayers.SoilOpticals — Type

mutable struct SoilOpticals{FT}

A struct of soil optical parameters

Fields

albedo_SW

Shortwave albedo

albedo_SW_SIF

Shortwave albedo for SIF

emsvty_SW

Shortwave Emissivity

albedo_LW

Longwave albedo

soil_skinT

Soil surface temperature [K]

Land.CanopyLayers.SolarAngles — Type

struct SolarAngles{FT}

Struct for observation and solar angles

Fields

tts

Solar Zenith Angle [degree]

tto

Viewing Zenith Angle in [degree]

psi

relative azimuth in [degree]

Land.CanopyLayers.WaveLengths — Type

mutable struct WaveLengths{FT}

Struct for pre-set wave length parameters

Fields

minwlPAR

Minimal WL for PAR [nm]

maxwlPAR

Maximal WL for PAR [nm]

minwle

Minimal WL for SIF excitation [nm]

maxwle

Maximal WL for SIF excitation [nm]

minwlf

Minimal WL for SIF emission/fluorescence [nm]

maxwlf

Maximal WL for SIF emission/fluorescence [nm]

sWL

Standard wave length [nm]

dWL

Differential wavelength

optis

Leaf optical parameter set

WL

Wave length [nm]

iWLE

Index of WLE in WL

iWLF

Index of WLF in WL

iPAR

index of wlPAR in WL

WLE

excitation wave length [nm]

WLF

Fluorescence wave length [nm]

WL_iPAR

Wave length for PAR

dWL_iPAR

Differential wave length for PAR

dWL_iWLE

Differential wave length for iWLE

nPAR

Length of WL_iPAR

nWL

Length of WL

nWLE

length of WLE

nWLF

length of WLF

Containers

Land.CanopyLayers.CFContainer — Type

mutable struct CFContainer{FT}

Container to speed canopy_fluxes! by pre-allocating arrays

Fields

abs_wave

absorbed energy from wave lengths

absfs_lidf

absfs' * lidf [nAzi]

E_all

wave length energy [same as dWL]

E_iPAR

wave length energy [same as iPAR]

lPs

lPs [nLayer]

kChlrel

kChlrel [same as iPAR]

PAR_diff

diffusive PAR [same as iPAR]

PAR_diffCab

diffusive PAR for photosynthesis [same as iPAR]

PAR_dir

direct PAR [same as iPAR]

PAR_dirCab

diffusive PAR for photosynthesis [same as iPAR]

Land.CanopyLayers.create_cf_container — Function

create_cf_container(FT, rt_dim::RTDimensions)

Create a CFContainer type struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.CGContainer — Type

mutable struct CGContainer{FT}

Container to speed canopy_geometry! by pre-allocating arrays

Fields

_Co

cos_ttli .* cos(tto) dim: nIncl

_Cs

cos_ttli .* cos(tts) dim: nIncl

_So

sin_ttli .* sin(tto) dim: nIncl

_Ss

sin_ttli .* sin(tts) dim: nIncl

_1s

maxtrix filled with 1 dim: (1, nAzi)

_2d

2D array to speed up _cds and _cdo dim: (nIncl, nAzi)

_cdo

Co * _1s .+ _So * cosphilo' dim: (nIncl, nAzi)

_cds

Cs * _1s .+ _Ss * costtlo' dim: (nIncl, nAzi)

Land.CanopyLayers.create_cg_container — Function

create_cg_container(FT, rt_dim::RTDimensions)

Create a CGContainer type struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.SFContainer — Type

mutable struct SFContainer{FT}

Container to speed SIF_fluxes! by pre-allocating arrays

Fields

M⁻_sun

M⁺_sun

wfEs

sfEs

sbEs

M⁺⁻

M⁺⁺

M⁻⁺

M⁻⁻

sun_dwl_iWlE

tmp_dwl_iWlE

ϕ_cosΘ_lidf

vfEplu_shade

vbEmin_shade

vfEplu_sun

vbEmin_sun

sigfEmin_shade

sigbEmin_shade

sigfEmin_sun

sigbEmin_sun

sigfEplu_shade

sigbEplu_shade

sigfEplu_sun

sigbEplu_sun

zeroB

tmp_1d_nWlF

tmp_1d_nLayer

dnorm

τ_dd

transmission of diffusive light?

ρ_dd

extinction of diffuse light?

Xdd

Rdd

Y

U

S⁻

S⁺

piLs

piLd

Fsmin

Fsplu

Fdmin

Fdplu

Femo

M⁺

M⁻

ϕ_cosΘ

F⁻

F⁺

net_diffuse

tmp_2d_nWlF_nLayer

tmp_2d_nWlF_nLayer_2

Land.CanopyLayers.create_sf_container — Function

create_sf_container(FT, rt_dim::RTDimensions)

Create a SFContainer type struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.SWContainer — Type

mutable struct SWContainer{FT}

Container to speed short_wave! by pre-allocating arrays

Fields

dnorm

dnorm?

piLo

pi * Lo

piLoc

pi * Lo from canopy

piLos

pi * Lo from soil

piLoc2

pi * Lo from canopy 2D matrix

ρ_dd

extinction of diffuse light?

ρ_sd

extinction of direct light?

τ_dd

transmission of diffusive light?

τ_sd

transmission of direct light?

Land.CanopyLayers.create_sw_container — Function

create_sw_container(FT, rt_dim::RTDimensions)

Create a CGContainer type struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.RTContainer — Type

mutable struct RTContainer{FT}

Collection of containers to speed up RT module

Fields

cf_con

CFContainer type container

cg_con

CGContainer type container

sf_con

SFContainer type container

sw_con

SWContainer type container

Land.CanopyLayers.create_rt_container — Function

create_rt_container(FT, rt_dim::RTDimensions)

Create an RTContainer, given

FT Floating number type

rt_dim RTDimensions type struct

Initialization of Structures

Land.CanopyLayers.create_canopy_opticals — Function

create_canopy_opticals( FT, rt_dim::RTDimensions)

Create a CanopyOpticals struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.create_canopy_rads — Function

create_canopy_rads(FT, rt_dim::RTDimensions)

Create a CanopyRads struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.create_canopy_rt — Function

create_canopy_rt(FT; nLayer::Intm LAI::Number)

Create Canopy4RT, given

FT Floating number type

nLayer Number of canopy layers

LAI Leaf area index

Land.CanopyLayers.create_incoming_radiation — Function

create_incoming_radiation(
            wls::WaveLengths{FT},
            wlfn::String
) where {FT<:AbstractFloat}

Create an AbstractIncomingRadiation struct, given

wls WaveLengths type struct

wlfn File that saves incoming wave information

Land.CanopyLayers.create_leaf_bios — Function

create_leaf_bios(FT, rt_dim::RTDimensions)

Create a LeafBios type struct, given

FT Floating number type

rt_dim RTDimensions type struct

Land.CanopyLayers.create_leaf_opticals — Function

create_leaf_opticals(sWL::Array{FT,1}, file::String)

Create an AbstractLeafOptiPara struct, given

sWL Standard wave length

opfn File that saves optical parameters

Land.CanopyLayers.create_rt_dims — Function

create_rt_dims(
            can::Canopy4RT{FT},
            wls::WaveLengths{FT}
) where {FT<:AbstractFloat}

Create RTDimensions, given

can Canopy4RT type struct

wls WaveLengths type struct

Land.CanopyLayers.create_soil_opticals — Function

create_soil_opticals(wls::WaveLengths{FT}) where {FT<:AbstractFloat}

Create SoilOpticals struct, given

wls WaveLengths type struct

Land.CanopyLayers.create_wave_length — Function

create_wave_length(FT)

Create WaveLengths type struct, given

FT Floating number type

Land.CanopyLayers.initialize_rt_module — Function

initialize_rt_module(FT; nLayer::Int, LAI::Number)

Initialize the RT module and return the sturctures, given

FT Floating number type

nLayer Number of canopy layers

LAI Leaf area index

This function initializes and returns

angles SolarAngles

can Canopy4RT

can_opt CanopyOpticals

can_rad CanopyRads

in_rad IncomingRadiation

leaves Array{LeafBios,1}

rt_con RTContainer

rt_dim RTDimensions

soil SoilOpticals

wls WaveLengths

Big Leaf Model

Land.CanopyLayers.big_leaf_partition — Function

big_leaf_partition(
            lai::FT,
            zenith::FT,
            r_all::FT
) where {FT<:AbstractFloat}

Partition the big-leaf canopy into sunlit and shaded layers, given

lai Leaf area index

zenith Zenith angle in degree

r_all Total radiation in [W m⁻²]

The function returns

ratio ratio of sunlit leaves out of all leaves

q_slm Mean sunlit layer PAR

q_shm Mean shaded layer PAR

e_sl Mean sunlit layer absorbed total energy

e_sh Mean shaded layer absorbed total energy

SCOPE Model

Land.CanopyLayers.canopy_fluxes! — Function

canopy_fluxes!(
            can::Canopy4RT{FT},
            can_opt::CanopyOpticals{FT},
            can_rad::CanopyRads{FT},
            in_rad::IncomingRadiation{FT},
            soil::SoilOpticals{FT},
            leaves::Array{LeafBios{FT},1},
            wls::WaveLengths{FT},
            rt_con::RTContainer{FT}
) where {FT<:AbstractFloat}

Computes a variety of integrated fluxes from the spectrally resolved computations in the short-wave Canopy RT (e.g. absorbed soil radiation, absorbed direct and diffuse PAR by layer (and angles for direct), net direct and diffuse energy balance per layer), given

can Canopy4RT type struct

can_opt CanopyOpticals type struct

can_rad CanopyRads type struct

in_rad IncomingRadiation type struct

soil SoilOpticals type struct

leaves Array of LeafBios type struct

wls WaveLengths type struct

rt_con RTContainer type container

Land.CanopyLayers.canopy_geometry! — Function

canopy_geometry!(
            can::Canopy4RT{FT},
            angles::SolarAngles{FT},
            can_opt::CanopyOpticals{FT},
            rt_con::RTContainer{FT}
) where {FT<:AbstractFloat}

Computes canopy optical properties (extinction coefficients for direct and diffuse light) based on the SAIL model. Most important input parameters are leaf inclination and azimuth distribution functions and sun-sensor geometry . Canopy clumping Ω is implemented as in Pinty et al (2015), given

can Canopy4RT type struct

angles SolarAngles type struct

can_opt CanopyOpticals type struct

rt_con RTContainer type container

Land.CanopyLayers.canopy_matrices! — Function

canopy_matrices!(
            leaves::Array{LeafBios{FT},1},
            can_opt::CanopyOpticals{FT}
) where {FT<:AbstractFloat}

Compute scattering coefficient matrices for direct and diffuse light given geometry dependent overall extinction coefficients and pigment dependent leaf reflectance and transmission (computed via fluspect). This function has to be called before short_wave! can be used.

leaves Array of LeafBios type struct

can_opt CanopyOpticals type struct

Land.CanopyLayers.clumping_factor! — Function

clumping_factor!(
            can::Canopy4RT{FT},
            angles::SolarAngles{FT}
) where {FT<:AbstractFloat}

Calculate the clumping factor, given

can Canopy4RT type struct

angles SolarAngles type struct

Land.CanopyLayers.diffusive_S — Function

diffusive_S(
            τ_dd::Array{FT},
            ρ_dd::Array{FT},
            S⁻::Array{FT},
            S⁺::Array{FT},
            boundary_top::Array{FT},
            boundary_bottom::Array{FT},
            rsoil::Array{FT}
) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport (used for thermal and SIF) given:

τ_dd A 2D Array with layer reflectances

ρ_dd A 2D Array with layer transmissions

S⁻ A 2D Array with layer source terms in the downwelling direction

S⁺ A 2D Array with layer source terms in the upwelling direction

boundary_top A 1D array with downwelling radiation at the top (top of canopy)

boundary_bottom A 1D array with upwnwelling radiation at the bottom (soil)

rsoil A 1D array with soil reflectance

Land.CanopyLayers.diffusive_S! — Function

diffusive_S!(
            sf_con::SFContainer{FT},
            soil::SoilOpticals{FT},
            rt_dim::RTDimensions
) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport (used for thermal and SIF), given

sf_con SFContainer type container

soil SoilOpticals type struct

rt_dim RTDimensions type struct

Land.CanopyLayers.fluspect! — Function

fluspect!(
            leaf::LeafBios{FT},
            wls::WaveLengths{FT}
) where {FT<:AbstractFloat}

Computes leaf optical properties (reflectance and transittance) based on pigment concentrations. Also computes Fluorescence excitation matrices. Mostly based on PROSPECT-D for leaf reflectance/transmission and FluSpec for fluorescence.

leaf LeafBios type struct

wls WaveLengths type struct

Land.CanopyLayers.short_wave! — Function

short_wave!(
            can::Canopy4RT{FT},
            can_opt::CanopyOpticals{FT},
            can_rad::CanopyRads{FT},
            in_rad::IncomingRadiation{FT},
            soil::SoilOpticals{FT},
            rt_con::RTContainer{FT}
) where {FT<:AbstractFloat}

Simulate the short wave radiation through the canopy, given

can Canopy4RT type struct

can_opt CanopyOpticals type struct

can_rad CanopyRads type struct

in_rad IncomingRadiation type struct

soil SoilOpticals type struct

rt_con RTContainer type container

Land.CanopyLayers.SIF_fluxes! — Function

SIF_fluxes!(
            leaves::Array{LeafBios{FT},1},
            can_opt::CanopyOpticals{FT},
            can_rad::CanopyRads{FT},
            can::Canopy4RT{FT},
            soil::SoilOpticals{FT},
            wls::WaveLengths{FT},
            rt_con::RTContainer{FT},
            rt_dim::RTDimensions
) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport for SIF radiation (calls [diffusive_S!] internally). Layer reflectance and transmission is computed from SW optical properties, layer sources from absorbed light and SIF efficiencies. Boundary conditions are zero SIF incoming from atmosphere or soil.

leaves Array of LeafBios type struct

can_opt CanopyOpticals type struct

can_rad CanopyRads type struct

can Canopy4RT type struct

soil SoilOpticals type struct

wls WaveLengths type struct

rt_con RTContainer type container

rt_dim RTDimensions type struct

Land.CanopyLayers.thermal_fluxes! — Function

thermal_fluxes!(
            leaves::Array{LeafBios{FT},1},
            can_opt::CanopyOpticals{FT},
            can_rad::CanopyRads{FT},
            can::Canopy4RT{FT},
            soil::SoilOpticals{FT},
            incLW::Array{FT},
            wls::WaveLengths{FT}
) where {FT<:AbstractFloat}

Computes 2-stream diffusive radiation transport for thermal radiation (calls [diffusive_S] internally). Layer reflectance and transmission is computed from LW optical properties, layer sources from temperature and Planck law, boundary conditions from the atmosphere and soil emissivity and temperature. Currently only uses Stefan Boltzmann law to compute spectrally integrated LW but can be easily adjusted to be spectrally resolved.

leaves Array of LeafBios type struct

can_opt CanopyOpticals type struct

can_rad CanopyRads type struct

can Canopy4RT type struct

soil SoilOpticals type struct

incLW A 1D array with incoming long-wave radiation

wls WaveLengths type struct

Utils

Land.CanopyLayers.calctav — Function

calctav(α::FT, nr::FT) where {FT<:AbstractFloat}

Computes transmission of isotropic radiation across an interface between two dielectrics (Stern F., 1964; Allen W.A., 1973)). From calctav.m in PROSPECT-D

α angle of incidence

nr Index of refraction

Land.CanopyLayers.dcum — Function

dcum(a::FT, b::FT, t::FT) where {FT<:AbstractFloat}

TODO Add function description

Land.CanopyLayers.dladgen — Function

dladgen(a::FT, b::FT, litab_bnd::Array{FT,2}) where {FT<:AbstractFloat}

TODO Calculate the freqency of WHAT?

Land.CanopyLayers.e2phot — Function

e2phot(λ::Array{FT}, E::Array{FT}) where {FT<:AbstractFloat}

Calculates the number of moles of photons, given

λ An array of wave length in [nm], converted to [m] by _FAC

E Joules of energy

Land.CanopyLayers.e2phot! — Function

e2phot!(
            λ::Array{FT,1},
            E::Array{FT,1},
            container::Array{FT,1}
) where {FT<:AbstractFloat}

Calculates the number of moles of photons, given

λ An array of wave length in [nm], converted to [m] by _FAC

E Joules of energy

container Container to avoid memory allocations

Land.CanopyLayers.expint — Function

expint(x::FT) where {FT<:AbstractFloat}

TODO add function description

Land.CanopyLayers.fast∫ — Function

fast∫(f::Vector{FT}, dx::Vector{FT}) where {FT<:AbstractFloat}

A fast way of integrating functions, given

f f(x) for each x

dx Delta x for each x

Land.CanopyLayers.fast∫! — Function

fast∫!(dx::Vector{FT}, f::Vector{FT}) where {FT<:AbstractFloat}

A fast way of integrating functions, given

f f(x) for each x

dx Delta x for each x

Note that f is a local container, and its values change in this operation.

Land.CanopyLayers.psofunction — Function

psofunction(
            K::FT,
            k::FT,
            Ω::FT,
            LAI::FT,
            q::FT,
            dso::FT,
            xl::FT
) where {FT<:AbstractFloat}

TODO explain the variables

Return the probability of observing a sunlit leaf at depth xl (pso, see eq 31 in vdT 2009), given

xl Leaf depth in the canopy

Land.CanopyLayers.volscatt! — Function

volscatt!(
            container::Array{FT,1},
            tts::FT,
            tto::FT,
            psi::FT,
            ttl::FT
) where {FT<:AbstractFloat}

Calculate interception parameters (chi_s and chi_s) and leaf reflectance multiplier (frho) and transmittance multiplier (ftau), given

container Array container for results

tts Solar zenith angle

tto Viewing zenith angle

psi Azimuth angle

ttl Leaf inclination angle