API

Plant Hydraulic System

The PlantHydraulics module provides two levels of hydraulics system: organ-level and plant-level. The organ-level hydraulic systems include Leaf, Root, and Stem (trunk and branch). The plant-level hydraulic system is can be any combination of the three organs (custimized definition may apply).

Leaf, Root, and Stem HS

Plant hydraulics is segmented to three organ-level systems/structs (LeafHydraulics, RootHydraulics, and StemHydraulics) subject to an Abstract type (AbstractHydraulicSystem). The major differences among the three structs are

LeafHydraulics has an extra-xylary component

RootHydraulics has a rhizosphere component

RootHydraulics and StemHydraulics have a gravity component

See the documentation for each struct for more details:

Land.PlantHydraulics.AbstractHydraulicSystem — Type

abstract type AbstractHydraulicSystem{FT}

Hierarchy of AbstractHydraulicSystem

LeafHydraulics

RootHydraulics

StemHydraulics

Land.PlantHydraulics.LeafHydraulics — Type

mutable struct LeafHydraulics{FT<:AbstractFloat}

A struct that contains leaf hydraulics information.

Fields

area

Leaf area [m²]

k_ox

Maximal extra-xylary hydraulic conductance [mol s⁻¹ MPa⁻¹ m⁻²]

k_sla

Maximal leaf hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

vc

Vulnerability curve

p_crt

Critical xylem pressure [MPa]

flow

Flow rate in the xylem [mol s⁻¹]

p_dos

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

p_leaf

Leaf end water pressure at air-water interface [MPa]

p_ups

Leaf xylem water pressure at the leaf base (upstream) [MPa]

k_element

List of leaf k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of leaf kr history per element

p_element

List of xylem water pressure [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_old

Temperature memory [K]

T_sap

Upstream sap temperature [K]

Land.PlantHydraulics.RootHydraulics — Type

mutable struct RootHydraulics{FT<:AbstractFloat}

A struct that contains root hydraulics information.

Fields

area

Root cross-section area [m²]

k_max

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

k_s

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

vc

Vulnerability curve

Δh

Root z difference [m]

k_rhiz

Rhizosphere conductance [mol s⁻¹ MPa⁻¹]

sh

Soil hydraulics

flow

Flow rate in the xylem [mol s⁻¹]

p_dos

Xylem water pressure at the downstream end of xylem [MPa]

p_rhiz

Xylem-rhizosphere interface water pressure [MPa]

p_ups

Soil matrix potential [MPa]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of kr history per element

p_element

List of xylem water pressure [MPa]

p_gravity

List of pressure drop caused by gravity [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_old

Temperature memory [K]

T_sap

Upstream sap temperature [K]

Land.PlantHydraulics.StemHydraulics — Type

mutable struct StemHydraulics{FT<:AbstractFloat}

A struct that contains stem hydraulics information.

Fields

area

Stem cross-section area [m²]

k_max

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

k_s

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

vc

Vulnerability curve

Δh

Stem height difference [m]

p_dos

Xylem water pressure at the downstream end of xylem [MPa]

p_ups

Xylem water pressure at the base (upstream) [MPa]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of kr history per element

p_element

List of xylem water pressure [MPa]

p_gravity

List of pressure drop caused by gravity [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_old

Temperature memory [K]

T_sap

Upstream sap temperature [K]

To initialize a hydraulics system, one needs to provide the floating type, for example:

using PlantHydraulics

FT = Float32;
hs_leaf = LeafHydraulics{FT}();
hs_root = RootHydraulics{FT}();
hs_stem = StemHydraulics{FT}();

Whole-plant HS

Plants differ in their structures, for example, some plants have a canopy far above the ground elevated by a trunk, some plants have a structured canopy supported by branch systems, and some plant has no trunk at all. To represent the structural differences, several types of plant hydraulics systems are pre-defined, and they are GrassLikeHS, PalmLikeHS, TreeLikeHS, and TreeSimple structs subject to a AbstractPlantHS type, where HS stands for hydraulic system. The major difference between the HSs are

GrassLikeHS has only mutiple root and canopy layers, no trunk or branch

PalmLikeHS has multiple root layers, a trunk, and multiple canopy layers, no branch system

TreeLikeHS has multiple root layers, a trunk, and multiple branch + canopy layers, and each branch corresponds to a canopy layer

TreeSimple has one root, one stem, and one leaf for testing purpose

See the documentation for each struct for more details:

Land.PlantHydraulics.AbstractPlantHS — Type

abstract type AbstractPlantHS{FT}

Hierachy of AbstractPlantHS

GrassLikeHS

PalmLikeHS

TreeLikeHS

Land.PlantHydraulics.GrassLikeHS — Type

mutable struct GrassLikeHS{FT<:AbstractFloat}

A plant hydraulic system like a grass, which contains multiple root layers, and multiple canopy layers. No trunk or branch system applies.

Fields

n_root

Root Layers

n_canopy

Canopy Layers

roots

Roots system

leaves

Leaves

root_index_in_soil

Corresponding soil layer per root layer

canopy_index_in_air

Corresponding air layer per canopy layer

container_k

Conductances for each layer at given flow

container_p

Pressure for each layer at given flow

container_q

Flow rate

Land.PlantHydraulics.PalmLikeHS — Type

mutable struct PalmLikeHS{FT<:AbstractFloat}

A plant hydraulic system like a palm, which contains multiple root layers, one trunk, and multiple canopy layers. No branch system applies.

Fields

n_root

Root Layers

n_canopy

Canopy Layers

roots

Roots system

trunk

Trunk

leaves

Leaves

root_index_in_soil

Corresponding soil layer per root layer

canopy_index_in_air

Corresponding air layer per canopy layer

container_k

Conductances for each layer at given flow

container_p

Pressure for each layer at given flow

container_q

Flow rate

Land.PlantHydraulics.TreeLikeHS — Type

mutable struct TreeLikeHS{FT<:AbstractFloat}

A plant hydraulic system like a tree, which contains multiple root layers, one trunk, and multiple branch and canopy layers.

Fields

n_root

Root Layers

n_canopy

Canopy Layers

roots

Roots system

trunk

Trunk

branch

Branch system

leaves

Leaves

root_index_in_soil

Corresponding soil layer per root layer

canopy_index_in_air

Corresponding air layer per canopy layer

container_k

Conductances for each layer at given flow

container_p

Pressure for each layer at given flow

container_q

Flow rate

Land.PlantHydraulics.TreeSimple — Type

mutable struct TreeSimple{FT<:AbstractFloat}

A plant hydraulic system with one root, one stem, and one leaf for testing purpose

Fields

root

Root

stem

Stem

leaf

Leaf

krs

Relative hydraulic conductance

To ease the initialization of a plant hydraulics system, a few customized functions are provided for quick initialization. More importantly, modifications to each field in the struct are always allowed. The quick functions are create_grass_like_hs, create_palm_like_hs , and create_tree_like_hs:

Land.PlantHydraulics.create_grass_like_hs — Function

create_grass_like_hs(z_root::FT, z_trunk::FT, z_canopy::FT, soil_bounds::Array{FT,1}, air_bounds::Array{FT,1}) where {FT<:AbstractFloat}

Create a GrassLikeHS, given

z_root Maximal root depth (negative value)

z_canopy Maximal canopy height (positive value)

soil_bounds Array of soil layer boundaries starting from 0

air_bounds Array of air layer boundaries starting from 0

Land.PlantHydraulics.create_palm_like_hs — Function

create_palm_like_hs(z_root::FT, z_trunk::FT, z_canopy::FT, soil_bounds::Array{FT,1}, air_bounds::Array{FT,1}) where {FT<:AbstractFloat}

Create a PalmLikeHS, given

z_root Maximal root depth (negative value)

z_trunk Maximal trunk height (positive value)

z_canopy Maximal canopy height (positive value)

soil_bounds Array of soil layer boundaries starting from 0

air_bounds Array of air layer boundaries starting from 0

Land.PlantHydraulics.create_tree_like_hs — Function

create_tree_like_hs(z_root::FT, z_trunk::FT, z_canopy::FT, soil_bounds::Array{FT,1}, air_bounds::Array{FT,1}) where {FT<:AbstractFloat}

Create a TreeLikeHS, given

z_root Maximal root depth (negative value)

z_trunk Maximal trunk height (positive value)

z_canopy Maximal canopy height (positive value)

soil_bounds Array of soil layer boundaries starting from 0

air_bounds Array of air layer boundaries starting from 0

What these functions do are to determine how many root layers and branch/canopy layers to add based on the tree information and environmental settings. To determine number of root layers, rooting depth and the soil layer information are required. The z_root is the maximal root depth in negative number, and soil_bounds is the boundaries of soil layers staring from 0. For example, for a soil_bounds of [0.0, -1.0, -2.0, -3.0, -4.0], a z_root of -1 gives 1 root layer, and a z_root of -1.5 or -2.0 gives 2 root layers. The z_trunk, z_canopy, and air_bounds determine how many canopy layers to add. For example, for a air_bounds of [0.0, 1.0, 2.0, 3.0, 4.0, 5.0 ... 20.0, 21.0, 22.0], a z_trunk of 5.0 z_canopy of 7.0 give 2 canopy layers, and a z_trunk of 5.5 z_canopy of 7.0 give 2 canopy layers. Also, the root_index_in_soil and canopy_index_in_air indicate which soil or air layer the root or canopy layer corresponds with, respectively. For instance, a index of 7 means that the canopy layer should use the 7th air layer.

To initialize a whole-plant hydraulic system, checkout the example below:

using PlantHydraulics

FT = Float32;
grass = create_grass_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
palm  =  create_palm_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
tree  =  create_tree_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
treet = TreeSimple{FT}();

Xylem hydraulic conductance

Plants transport water through xylem conduits (vessels in most angiosperms, trachieds in most gymnosperms). With the ascent of sap along the hydraulic system, water pressure in the conduits is typically negative. The negative xylem water pressure tends to pull air from surrounding tisses or the atmosphere into the xylem conduits, resulting in xylem cavitation. The air bubbles in cavitated conduits block water flow, and thus results in decline of water transport capability (measured by xylem hydraulic conductance).

Typically, the correlation between xylem water pressure ($P \leq 0$) and hydraulic conductance ($k$) is expressed by a Weibull function for WeibullSingle type correlation:

\[k = k_\text{max} \cdot \exp \left( -\left( \dfrac{-P}{B} \right)^C \right)\]

where $k_\text{max}$ is the maximal hydraulic conductance, and $B$ and $C$ are the Weibull parameters. This correlation is also known as vulnerability curve (VC) to drought stress. Sometimes, plants exhibit a segmented VC, for example, the fibers may transport water as well and are much more resistant to drought than vessels. Thus, a dual Weibull function is presented for WeibullDual type correlation ($P \leq 0$):

\[k = k_\text{max} \cdot \left\{ f_1 \cdot \exp \left[ -\left( \dfrac{-P}{B_1} \right)^{C_1} \right] + (1 - f_1) \cdot \exp \left[ -\left( \dfrac{-P}{B_2} \right)^{C_2} \right] \right\}\]

The VC formulations are abstractized as

Land.PlantHydraulics.AbstractXylemVC — Type

abstract type AbstractXylemVC{FT}

Hierachy of AbstractXylemVC

WeibullSingle

WeibullDual

Land.PlantHydraulics.WeibullDual — Type

mutable struct WeibullDual{FT}

Struct that contains dual Weibull function parameters.

Fields

b1

B of first part [MPa]

c1

C of first part

f1

F of first part

b2

B of second part [MPa]

c2

C of second part

f2

F of second part

Land.PlantHydraulics.WeibullSingle — Type

mutable struct WeibullSingle{FT}

Struct that contains single Weibull function parameters.

Fields

b

B of first part [MPa]

c

C of first part

The function to call is

Land.PlantHydraulics.xylem_k_ratio — Function

xylem_k_ratio(vc::WeibullSingle{FT}, p_25::FT, vis::FT) where {FT<:AbstractFloat}
xylem_k_ratio(vc::  WeibullDual{FT}, p_25::FT, vis::FT) where {FT<:AbstractFloat}

Returns the relative hydraulic conductance, given

vc Xylem vulnerability curve

p Xylem pressure at 298.15 K in [MPa]

p_25 Equivalent xylem pressure at 298.15 K in [MPa]

vis Relative viscosity. If missing, vis = 1.

Note it here that xylem_k_ratio(vc, p) calculate the k without making temperature corrections, but the xylem_k_ratio(vc, p_25, vis) makes correction over the viscosity (the higher the viscosity, the lower the k). Also, p_25 means that the pressure has been corrected to 298.15 K for surface tension (the higher the surface tension, the more resistant the xylem).

Meanwhile, there is a function to call to calculate the critical pressure, beyond which leaf will decicate. The critical pressure is calculated as the pressure at which $k$ is 0.001 of $k_\text{max}$ for WeibullSingle (for WeibullDual, each segment need to reach 0.001). The functions is

Land.PlantHydraulics.xylem_p_crit — Function

xylem_p_crit(vc::WeibullSingle{FT}, f_st::FT) where {FT<:AbstractFloat}
xylem_p_crit(vc::WeibullDual{FT}, f_st::FT) where {FT<:AbstractFloat}

Returns the relative hydraulic conductance, given

vc Xylem vulnerability curve

st Relative surface tension. If missing, vis = 1.

Examples:

using PlantHydraulics

FT = Float32;
vc_1 = WeibullSingle{FT}();
vc_2 = WeibullDual{FT}();

k_1 = xylem_k_ratio(vc_1, -1.0);
k_2 = xylem_k_ratio(vc_2, -1.0);
k_3 = xylem_k_ratio(vc_1, -1.0, 1.2);
k_4 = xylem_k_ratio(vc_2, -1.0, 1.2);

Rhizosphere hydraulic conductance

As mentioned above, there is a rhizosphere component in the root hydraulic system, and thus one needs to compute the pressure dtop along the rhizosphere. The soil properties are classified to BrooksCorey and VanGenuchten types subjected to AbstractSoilVC:

Land.PlantHydraulics.AbstractSoilVC — Type

abstract type AbstractSoilVC{FT}

Hierachy of AbstractSoilVC:

BrooksCorey

VanGenuchten

Land.PlantHydraulics.BrooksCorey — Type

mutable struct BrooksCorey{FT<:AbstractFloat}

Brooks Corey soil parameters

Fields

stype

Soil type

b

Soil b

ϕs

ϕ at saturation [MPa]

Θs

Soil water content (Θ) at saturation

Θr

Residual soil water content

Land.PlantHydraulics.VanGenuchten — Type

mutable struct VanGenuchten{FT<:AbstractFloat}

Van Gunechten soil parameters

Fields

stype

Soil type

α

Soil α is related to the inverse of the air entry suction, α > 0

n

Soil n is Measure of the pore-size distribution

m

Soil m = 1 - 1/n

Θs

Soil water content (Θ) at saturation

Θr

Residual soil water content

Correlations among soil relative water content, relative hydraulic conductance, and soil matrix potential are

Land.PlantHydraulics.soil_erwc — Function

soil_erwc(sh::BrooksCorey{FT}, p_25::FT) where {FT<:AbstractFloat}
soil_erwc(sh::VanGenuchten{FT}, p_25::FT) where {FT<:AbstractFloat}

Returns the Effective relative water content of the soil $\frac{Θs - Θr}{Θs - Θ}$, given

sh BrooksCorey or VanGenuchten type soil hydraulics

p_25 Matrix water potential equivalent to 25 degree C, with surface tension

correction

Land.PlantHydraulics.soil_rwc — Function

soil_rwc(sh::BrooksCorey{FT}, p_25::FT) where {FT<:AbstractFloat}
soil_rwc(sh::VanGenuchten{FT}, p_25::FT) where {FT<:AbstractFloat}

Returns the relative soil water content, given

sh BrooksCorey or VanGenuchten type soil hydraulics

p_25 Matrix water potential equivalent to 25 degree C, with surface tension

correction

Land.PlantHydraulics.soil_swc — Function

soil_swc(sh::BrooksCorey{FT}, p_25::FT) where {FT<:AbstractFloat}
soil_swc(sh::VanGenuchten{FT}, p_25::FT) where {FT<:AbstractFloat}

Returns the soil water content, given

sh BrooksCorey or VanGenuchten type soil hydraulics

p_25 Matrix water potential equivalent to 25 degree C, with surface tension

correction

Land.PlantHydraulics.soil_k_ratio_erwc — Function

soil_k_ratio_erwc(sh::BrooksCorey{FT}, erwc::FT) where {FT<:AbstractFloat}
soil_k_ratio_erwc(sh::VanGenuchten{FT}, erwc::FT) where {FT<:AbstractFloat}

Return the soil k relative to maximal k, given

sh BrooksCorey or VanGenuchten type soil hydraulics

erwc Effective relative soil water content ($\frac{Θs - Θr}{Θs - Θ}$)

Land.PlantHydraulics.soil_k_ratio_rwc — Function

soil_k_ratio_rwc(sh::BrooksCorey{FT}, rwc::FT) where {FT<:AbstractFloat}
soil_k_ratio_rwc(sh::VanGenuchten{FT}, rwc::FT) where {FT<:AbstractFloat}

Return the soil k relative to maximal k, given

sh BrooksCorey or VanGenuchten type soil hydraulics

rwc Relative soil water content

Land.PlantHydraulics.soil_k_ratio_swc — Function

soil_k_ratio_swc(sh::BrooksCorey{FT}, swc::FT) where {FT<:AbstractFloat}
soil_k_ratio_swc(sh::VanGenuchten{FT}, swc::FT) where {FT<:AbstractFloat}

Return the soil k relative to maximal k, given

sh BrooksCorey or VanGenuchten type soil hydraulics

swc Relative soil water content

Land.PlantHydraulics.soil_k_ratio_p25 — Function

soil_k_ratio_p25(sh::AbstractSoilVC{FT}, p_25::FT) where {FT<:AbstractFloat}
soil_k_ratio_p25(sh::VanGenuchten{FT}, p_25::FT) where {FT<:AbstractFloat}

Return the soil k relative to maximal k, given

sh AbstractSoilVC type soil hydraulics

p_25 Matrix water potential equivalent to 25 degree C, with surface tension

Land.PlantHydraulics.soil_p_25_erwc — Function

soil_p_25_erwc(sh::BrooksCorey{FT}, erwc::FT) where {FT<:AbstractFloat}
soil_p_25_erwc(sh::VanGenuchten{FT}, erwc::FT) where {FT<:AbstractFloat}

Returns the Relative water content of the soil, given

sh BrooksCorey or VanGenuchten type soil hydraulics

erwc Effectibe relative soil water content ($\frac{Θs - Θr}{Θs - Θ}$)

Note that this function returns the matrix potential of water, not water potential. Also, the potential is that at 25 degree C, not yet been corrected for temperature effect on surface tension.

Land.PlantHydraulics.soil_p_25_rwc — Function

soil_p_25_rwc(sh::BrooksCorey{FT}, rwc::FT) where {FT<:AbstractFloat}
soil_p_25_rwc(sh::VanGenuchten{FT}, rwc::FT) where {FT<:AbstractFloat}

Returns the Relative water content of the soil, given

sh BrooksCorey or VanGenuchten type soil hydraulics

rwc Relative soil water content

Note that this function returns the matrix potential of water, not water potential. Also, the potential is that at 25 degree C, not yet been corrected for temperature effect on surface tension.

Land.PlantHydraulics.soil_p_25_swc — Function

soil_p_25_swc(sh::BrooksCorey{FT}, rwc::FT) where {FT<:AbstractFloat}
soil_p_25_swc(sh::VanGenuchten{FT}, rwc::FT) where {FT<:AbstractFloat}

Returns the Relative water content of the soil, given

sh BrooksCorey or VanGenuchten type soil hydraulics

swc Soil water content

Note that this function returns the matrix potential of water, not water potential. Also, the potential is that at 25 degree C, not yet been corrected for temperature effect on surface tension.

Pressure and Flow

The PlantHydraulics module is designed to run numerically for the following reasons:

Weibull function is cannot be integrated

The VC is segmented, i.e., if $P > 0$, $k = k_\text{max}$ (implemented in xylem_k_ratio)

Once xylem cavitation occurs, it cannot be easily recovered unless $P > 0$, and thus there is a drought legacy effect. This is why there are a few fields in the LeafHydraulics, RootHydraulics, and StemHydraulics structs to store the drought history information.

Temperature may change along the flow path. The f_st and f_vis in the structs help deal with these effects.

Function xylem_p_from_flow calculates the xylem end pressure for an organ-level hysraulic system. As mentioned above, the RootHydraulics and StemHydraulics has a gravity component, and the RootHydraulics has a rhizosphere component. Also be aware that xylem_p_from_flow accounts for temperature effects on surface tension and viscosity.

Land.PlantHydraulics.xylem_p_from_flow — Function

xylem_p_from_flow(leaf::LeafHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(root::RootHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(stem::StemHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(tree::GrassLikeHS{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(tree::PalmLikeHS{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(tree::TreeLikeHS{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(tree::TreeSimple{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(tree::TreeSimple{FT}, f_sl::FT, f_sh::FT, r_sl::FT) where {FT<:AbstractFloat}

Return the xylen end pressure(s) from flow rate(s), given

leaf LeafHydraulics type struct

root RootHydraulics type struct

stem StemHydraulics type struct

tree AbstractPlantHS type struct

flow Flow rate (per leaf area for LeafHydraulics)

f_sl Flow rate to sunlit leaves

f_sh Flow rate to shaded leaves

r_sl Fraction of sunlit leaves

Note, gravity is accounted for in root and stem; rhizosphere conductance is accounted for in root; extra-xylary conductance is not accounted for in leaf here because it calculates xylem end pressure.

Noe that function xylem_p_from_flow does not update the pressure profiles or history in the xylem. To update these profiles, use hydraulic_p_profile!, and to remove these legacy profiles, use inititialize_legacy!:

Land.PlantHydraulics.hydraulic_p_profile! — Function

hydraulic_p_profile!(leaf::LeafHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
hydraulic_p_profile!(root::RootHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
hydraulic_p_profile!(stem::StemHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
hydraulic_p_profile!(tree::TreeSimple{FT}, p_soil::FT, flow::FT) where {FT<:AbstractFloat}
hydraulic_p_profile!(tree::TreeSimple{FT}, p_soil::FT, f_sl::FT, f_sh::FT, r_sl::FT) where {FT<:AbstractFloat}

Update the pressure profile, given

leaf LeafHydraulics type struct

root RootHydraulics type struct

stem StemHydraulics type struct

tree TreeSimple type struct

p_soil Soil water potential

flow Flow rate (per leaf area for LeafHydraulics)

f_sl Flow rate to sunlit leaves

f_sh Flow rate to shaded leaves

r_sl Fraction of sunlit leaves

Land.PlantHydraulics.inititialize_legacy! — Function

inititialize_legacy!(hs::AbstractHydraulicSystem{FT}) where {FT<:AbstractFloat}
inititialize_legacy!(tree::TreeSimple{FT}) where {FT<:AbstractFloat}

Initialize the drought legacy effects in the xylem, given

hs [AbstractHydraulicSystem] type struct

tree TreeSimple type struct

Examples:

using PlantHydraulics

FT = Float32;
leaf = LeafHydraulics{FT}();
p = xylem_p_from_flow(leaf, FT(0.01));
@show leaf.p_element;
hydraulic_p_profile!(leaf, FT(0.01));
@show leaf.p_element;

Root Hydraulics

Function xylem_p_from_flow works for the case of only 1 root layer if one needs the plant base xylem water pressure. However, when there are multiple root layers, xylem_p_from_flow does not apply. In this case, iterations are required to calculate the xylem end pressure for each root layers, and then make sure all root layers have the same xylem end pressure. A few functions are provided to realize this.

Function root_q_from_pressure uses Root Solving method to calculate the flow rate through the RootHydraulics struct that yields the given xylem end pressure. The ini in the function is optional. However, using the flow rate from last instant when pressure does not differ much will speed up the calculation.

Land.PlantHydraulics.root_q_from_pressure — Function

root_q_from_pressure(root::RootHydraulics{FT}, pressure::FT, ini::FT) where {FT<:AbstractFloat}

Calculate the flow rate from a given tree base pressure, given

root RootHydraulics type struct

pressure Given tree base pressure in [MPa]

ini Initial guess

In the plant hydraulic module design, flow rate is computed for each canopy layer, and thus computing flow rate for each root layer is required for a multiple layered root system. One feasible way is to do iterations using root_q_from_pressure function, i.e., iterate the xylem end pressure til the total flow rate equals the given value. However, this method is too inefficient. What I did is

Calculate the xylem end pressure and whole root layer conductance from the initial flow rate;

Calculate the mean xylem end pressure, and tune the flow rates in all root layers using the difference from mean pressure and root conductance;

Sum up the new flow rates, and calculate the difference with given total flow rate;

Use the calculated condutcance to weight out the differences.

The functions provided by PlantHydraulics module are

Land.PlantHydraulics.root_pk_from_flow — Function

root_pk_from_flow(root::RootHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}

Return root xylem end pressure and root hydraulic conductance (reverse of summed resistance), given

root RootHydraulics struct

flow Given flow rates in the root layer

Land.PlantHydraulics.recalculate_roots_flow! — Function

rrecalculate_roots_flow!(roots::Array{RootHydraulics{FT},1}, ks::Array{FT,1}, ps::Array{FT,1}, qs::Array{FT,1}, flow::FT) where {FT<:AbstractFloat}

Recalculate the flow rates in the root from the pressure and conductance profiles in each root at non-steady state, given

roots Array of RootHydraulics structs

ks Container for conductance in each root layer

ps Container for end xylem pressure in each layer

qs Container for flow rate in each layer

flow Total flow rate

However, the steps above are only 1 iteration, and can only be used for the non-steady state version of model. For the steady-state flow rates, function roots_flow! does thw work. What the function does is to iterate recalculate_roots_flow! till the difference among the calculated end pressures is small enough. I also emphasize that to speed up the code, 3 containers are added to the AbstractPlantHS structs, and they are container_k, container_p, and container_q.

Land.PlantHydraulics.roots_flow! — Function

root_qs_p_from_q(roots::Array{RootHydraulics{FT},1}, ks::Array{FT,1}, ps::Array{FT,1}, qs::Array{FT,1}, flow::FT) where {FT<:AbstractFloat}

Solve for the flow rates in each root layer and root pressure at steady state, given

roots Array of RootHydraulics structs

ks Container for conductance in each root layer

ps Container for end xylem pressure in each layer

qs Container for flow rate in each layer

flow Total flow rate

Example:

using PlantHydraulics

FT = Float32;
palm  =  create_palm_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
roots_flow!(palm.roots, palm.container_k, palm.container_p, palm.container_q, FT(1));

Leaf Hydraulics

The stomatal models often require plant hydraulics either as a correction factor (in empirical stomatal models) or as the risk term (in optimal stomatal models). To facilitate the calculations, a few specific functions are provided.

Function leaf_xylem_risk returns the risk in xylem hydraulic function based on the most downstream end of the xylem. The risk of plant hydraulic system is not only on current system, but also potential new growth (plants don't want to risk new growth either). Thus, function leaf_xylem_risk evaluates the risk from the xylem pressure calculated from current system (with drought history), and then compute the risk from the pressure (the severer the srought history, the higher the risk):

Land.PlantHydraulics.leaf_xylem_risk — Function

leaf_xylem_risk(hs::LeafHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}

Evaluate the hydraulic risk at the end of leaf xylem, given

hs LeafHydraulics type struct

flow Flow rate (per leaf area)

Note that function leaf_xylem_risk can work on its own without having other organ-level components. For example, by changing the p_ups of a LeafHydraulics, one can simulate the case of drought without caring about other hydraulic systems. Same for function leaf_e_crit below. However, these functions are only useful for sensitivity analysis or when p_ups in the LeafHydraulics is accurate.

Examples

using PlantHydraulics

FT = Float32;
leaf = LeafHydraulics{FT}();
risk = leaf_xylem_risk(leaf, FT(0.01));
@show risk;
leaf.p_ups = FT(-1.0);
risk = leaf_xylem_risk(leaf, FT(0.01));
@show risk;

Function leaf_e_crit calculates critical leaf transpiration rate, beyond which leaf will desicate. Function leaf_e_crit accounts for drought legacy effect by design, and the more severe the drought history, the lower the e_crit. Again, ini in the function is also optional, but a good guess will speed up the calculations.

Land.PlantHydraulics.leaf_e_crit — Function

leaf_e_crit(hs::LeafHydraulics{FT}, ini::FT) where {FT<:AbstractFloat}

Calculate the critical flow rate (K ≈ 0), given

hs LeafHydraulics type struct

ini Initial guess

Note, for the safety of no NaN, update e_crit when ΔP >= -0.01

Examples

using PlantHydraulics

FT = Float32;
leaf = LeafHydraulics{FT}();
risk = leaf_e_crit(leaf);
@show risk;
leaf.p_ups = FT(-1.0);
risk = leaf_e_crit(leaf);
@show risk;

Whole-plant Hydraulics

Though leaf_xylem_risk and leaf_e_crit can work on their own, the functions only evaluate the risks on leaf level. The more realistic case is that when leaf transpiration rate increases, p_ups in the LeafHydraulics gets more negative. Thus, the leaf_xylem_risk and leaf_e_crit tends to underestimate the risk and overestimate the critical flow rate. To overcome this problem, whole-plant level plant hydraulics are provided.

Function xylem_p_from_flow calculates the leaf xylem end pressure for a whole-plant struct using these steps:

calculate the plant base pressure from a given total flow rate

calculate the trunk end pressure (if present)

calculate the branch end pressure (if present)

calculate the leaf end pressure (if present)

Accordingly, there is a function tree_e_crit to calculate the critical flow rate for the whole plant. Be aware that Plant-level function xylem_p_from_flow and tree_e_crit only applies to the case of only one canopy layer (or big-leaf model). As to the case of multiple canopy layer, more functions are pending.

Land.PlantHydraulics.tree_e_crit — Function

tree_e_crit(tree::PalmLikeHS{FT}, ini::FT) where {FT<:AbstractFloat}

Calculate the critical flow rate for the tree per LA, given

tree GrassLikeHS or PalmLikeHS or TreeLikeHS type hydraulics system

ini Initial guess of the critical flow rate

Note that the organ level or whole-plant level conductances are different from xylem hydraulic conductance at a given xylem slice. Also, simply computing the conductance as the flow rate divided by pressre drop is not accurate because of the gravity. In the PlantHydraulics module, the organ and whole-plant level conductances are computed by firstly adding up all the resistances in each element, and then computing the relative loss of conductance. The function available to use is

Land.PlantHydraulics.plant_conductances! — Function

plant_conductances!(tree::TreeSimple{FT}) where {FT<:AbstractFloat}

Update plant total hydraulic information in each element and whole plant, given

tree TreeSimple type struct

Temperature effects

Plant hydraulic properties changes with temperature, due to its impacts on surface tension and viscosity. As to surface tension, when temperature increases, air-water interface surface tension decreases, meaning that with the same curvature, the capillary pressure provided decreases. As a result, soil matrix potential becomes less nagative (good for plants!), but the conduit resistance to cavitation decreases (bad for plants!). As to viscosity, when temperature increases, viscosity decreases, meaning that pressure drop decreases at the same flow rate (good for plants!). To account for these effects, we provided vc_temperature_effects! function:

Land.PlantHydraulics.vc_temperature_effects! — Function

vc_temperature_effects!(hs::AbstractHydraulicSystem{FT}, T::FT) where {FT<:AbstractFloat}
vc_temperature_effects!(hs::AbstractHydraulicSystem{FT}) where {FT<:AbstractFloat}
vc_temperature_effects!(tree::TreeSimple{FT}) where {FT<:AbstractFloat}

Update temperature effetcs on hydralic system, given

hs AbstractHydraulicSystem type struct

T Given temperature

tree TreeSimple type struct

Keep in mind that, leaf critical pressure changes due to surface tension, though the impact can be neglected for minor temperature change. However, soil water content is still computed using the equivalent matrix potential at 25 Celcius because water content is only related to the air-water interface curvature.