StomtaModels API

Stomatal model schemes

The StomataModels module relies mainly on Photosynthesis and PlantHydraulics modules to predict stomatal behavior from plant physiology. This module has both empirical and optimal stomatal models. These stomatal models are abstractized to an abstract AbstractStomatalModel, which further has subtypes EmpiricalStomatalModel and OptimizationStomatalModel.

Land.StomataModels.AbstractStomatalModel — Type

type AbstractStomatalModel

Hierarchy of the AbstractStomatalModel:

EmpiricalStomatalModel

OptimizationStomatalModel

Land.StomataModels.EmpiricalStomatalModel — Type

type EmpiricalStomatalModel

Hierarchy of the EmpiricalStomatalModel:

ESMBallBerry

ESMGentine

ESMLeuning

ESMMedlyn

Land.StomataModels.OptimizationStomatalModel — Type

type OptimizationStomatalModel

Hierarchy of the OptimizationStomatalModel:

OSMEller

OSMSperry

OSMWang

OSMWAP

OSMWAPMod

Currently, the StomataModels module has four empirical model schemes, and they are

Land.StomataModels.ESMBallBerry — Type

struct ESMBallBerry{FT}

An empirical model parameter set type for Ball-Berry type model. The equation used for Ball-Berry type model is

\[gs = g0 + g1 ⋅ RH ⋅ \dfrac{A}{Cs}\]

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [unitless]

Land.StomataModels.ESMGentine — Type

struct ESMGentine{FT}

An empirical model parameter set type for Gentine type model. The equation used for Gentine type model is

\[gs = g0 + g1 ⋅ \dfrac{k_{leaf}}{k_{max}} ⋅ \dfrac{A}{Ca}.\]

Note it that the Gentine model does not require for a β function to tune the soil drought response, but the use of k_leaf also does not permit post-drought stomatal response unless k_leaf can be recovered.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [unitless]

Land.StomataModels.ESMLeuning — Type

struct ESMLeuning{FT}

An empirical model parameter set type for Leuning type model. The equation used for Leuning type model is

\[gs = g0 + g1 ⋅ \dfrac{A}{Cs - Γ^{*}} ⋅ \dfrac{1}{1 + \dfrac{VPD}{d0}}\]

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [unitless]

d0

fitting parameter of d/d0 below the fraction, same unit as vpd [Pa]

Land.StomataModels.ESMMedlyn — Type

struct ESMMedlyn{FT}

An empirical model parameter set type for Medlyn type model. The equation used in Medlyn type model is

\[gs = g0 + 1.6 ⋅ \left( 1 + \dfrac{g1}{\sqrt{VPD}} \right) ⋅ \dfrac{A}{Ca}\]

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [Pa⁽⁵⁾]

All the empirical models rely on beta functions to make corrections over stomatal conductance to account for the stomatal closure with drier soil. We have the following prescribed beta function types, and they are:

Land.StomataModels.AbstractBetaFunction — Type

abstract type AbstractBetaFunction{FT}

Hierachy of AbstractBetaFunction:

AbstractBetaG

BetaGLinearPleaf

BetaGLinearPsoil

BetaGLinearSWC

AbstractBetaV

BetaVLinearPleaf

BetaVLinearPsoil

BetaVLinearSWC

Some beta functions make correction over the g1 parameter as in the empitical models, and they are:

Land.StomataModels.AbstractBetaG — Type

abstract type AbstractBetaG{FT}

Hierachy of AbstractBetaG:

BetaGLinearPleaf

BetaGLinearPsoil

BetaGLinearSWC

Land.StomataModels.BetaGLinearPleaf — Type

mutable struct BetaGLinearPleaf{FT}

Linear β function for g1 based on soil water potential.

Fields

p_max

Upper bound of Pleaf [MPa]

p_min

Lower bound of Pleaf [MPa]

Land.StomataModels.BetaGLinearPsoil — Type

mutable struct BetaGLinearPsoil{FT}

Linear β function for g1 based on soil water potential.

Fields

p_max

Upper bound of Psoil [MPa]

p_min

Lower bound of Psoil [MPa]

Land.StomataModels.BetaGLinearSWC — Type

mutable struct BetaGLinearSWC{FT}

Linear β function for g1 based on soil water content.

Fields

swc_max

Upper bound of SWC

swc_min

Lower bound of SWC

Some beta functions make correction over the photosynthetic capacity as in the Photosynthesis module, and they are:

Land.StomataModels.AbstractBetaV — Type

abstract type AbstractBetaV{FT}

Hierachy of AbstractBetaV:

BetaVLinearPleaf

BetaVLinearPsoil

BetaVLinearSWC

Land.StomataModels.BetaVLinearPleaf — Type

mutable struct BetaVLinearPleaf{FT}

Linear β function for g1 based on soil water potential.

Fields

p_max

Upper bound of Pleaf [MPa]

p_min

Lower bound of Pleaf [MPa]

Land.StomataModels.BetaVLinearPsoil — Type

mutable struct BetaVLinearPsoil{FT}

Linear β function for Vcmax based on soil water potential.

Fields

p_max

Upper bound of Psoil [MPa]

p_min

Lower bound of Psoil [MPa]

Land.StomataModels.BetaVLinearSWC — Type

mutable struct BetaVLinearSWC{FT}

Linear β function for Vcmax based on soil water content.

Fields

swc_max

Upper bound of SWC

swc_min

Lower bound of SWC

The beta functions are generalized with

Land.StomataModels.β_factor — Function

β_factor(bt::AbstractBetaFunction{FT},
         pl::FT,
         ps::FT,
         swc::FT) where {FT<:AbstractFloat}

Calculate the β correction factor, given

bt AbstractBetaFunction type struct

pl Leaf water potential [MPa]

ps Soil water potential [MPa]

swc Soil water content

The StomataModels module also contains five optimization model schemes:

Land.StomataModels.OSMEller — Type

struct OSMEller

An optimization model parameter set type for Eller model. The equation used for Eller model is

\[\dfrac{∂Θ}{∂E} = -\dfrac{∂K}{∂E} ⋅ \dfrac{A}{K}\]

where K is $\dfrac{∂E}{∂P}$.

Fields

Land.StomataModels.OSMSperry — Type

struct OSMSperry

An optimization model parameter set type for Sperry model. The equation used for Sperry model is

\[\dfrac{∂Θ}{∂E} = -\dfrac{∂K}{∂E} ⋅ \dfrac{A_{max}}{K_{max}}\]

where K is $\dfrac{∂E}{∂P}$.

Fields

Land.StomataModels.OSMWang — Type

struct OSMWang

An optimization model parameter set type for Eller type model. The equation used for Wang model is

\[\dfrac{∂Θ}{∂E} = \dfrac{A}{E_{crit} - E}\]

Fields

Land.StomataModels.OSMWAP — Type

struct OSMWAP{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model. The equation used for Wolf-Anderegg-Pacala model is

\[\dfrac{∂Θ}{∂E} = \dfrac{2aP + b}{K}\]

where K is ∂P/∂E.

Fields

a

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻²]

b

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻¹]

Land.StomataModels.OSMWAPMod — Type

struct OSMWAPMod{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model, modified by adding a photosynthesis component while set b and c = 0. The equation used for modified Wolf-Anderegg-Pacala model is

\[\dfrac{∂Θ}{∂E} = \dfrac{aAP}{K}\]

where P is absolute value of leaf xylem pressure.

Fields

a

Quadratic equation parameter [mol mol⁻¹ MPa⁻¹]

CanopyLayer

The StomataModels module is designed for multi-layer canopies, and each canopy has multiple leaves. The stomatal behaviors are modeled per layer basis, and the layer may contain any number of leaves starting from 1. Photosynthesis-related information is stored in CanopyLayer struct, but be aware that the leaves have uniform photosynthetic parameters and temperature (conductances are different in response to light environment).

Land.StomataModels.CanopyLayer — Type

struct CanopyLayer{FT}

Struct to store leaf information (multi-dimensional).

Fields

ps

leaf photosynthesis system

LA

Total leaf area [m²]

LAI

Total leaf area index in the layer

n_leaf

H

Sensible Heat Flux [W m⁻²]

LE

Latent Heat Flux [W m⁻²]

Rn

Net Radiation Balance [W m⁻²]

LV

Latent Heat of evaporation [J mol⁻¹]

T

Temperature [K]

T_old

Old temperature [K]

width

Leaf width [m]

g_bc

Boundary layer conductance to CO₂ [mol m⁻² s⁻¹]

g_bh

Boundary layer conductance to heat [mol m⁻² s⁻¹]

g_bw

Boundary layer conductance to H₂O [mol m⁻² s⁻¹]

g_lc

Leaf diffusive conductance to water CO₂ [mol m⁻² s⁻¹]

g_lw

Leaf diffusive conductance to water H₂O [mol m⁻² s⁻¹]

g_m

Mesophyll conductance for CO₂ [mol m⁻² s⁻¹]

g_sc

Stomatal conductance to water CO₂ [mol m⁻² s⁻¹]

g_sw

Stomatal conductance to water H₂O [mol m⁻² s⁻¹]

g_ias_c

Gias correction constant

g_ias_e

Gias correction exponent

g_max

Maximal leaf diffusive conductance [mol m⁻² s⁻¹]

g_max25

Maximal leaf diffusive conductance at 298.15 K [mol m⁻² s⁻¹]

g_min

Minimal leaf diffusive conductance [mol m⁻² s⁻¹]

g_min25

Minimal leaf diffusive conductance at 298.15 K [mol m⁻² s⁻¹]

p_i

Leaf internal CO₂ partial pressure [Pa]

p_s

Leaf surface CO₂ partial pressure [Pa]

p_sat

Leaf saturation vapor pressure [Pa]

Ac

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

Aj

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

Ag

Gross photosynthetic rate [μmol m⁻² s⁻¹]

An

Net photosynthetic rate [μmol m⁻² s⁻¹]

Ap

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

J

Electron transport [μmol m⁻² s⁻¹]

J_pot

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

CO₂_per_electron

Total efficiency, incl. photorespiration [mol CO₂ mol⁻¹ e-]

Fm′

light adapted yield (Kp=0)

Fo′

light-adapted fluorescence yield in the dark (Kp=max)

Ja

Actual electron transport rate [μmol m⁻² s⁻¹]

NPQ

Non-Photochemical quenching

qQ

Photochemical quenching

qE

energy quenching

φ

PSII yield

ϕs

Steady-state (light-adapted) yield (aka Fs)

Fm

dark adapted yield (Kp=0)

Fo

dark-adapted fluorescence yield (Kp=max)

APAR

Absorbed photosynthetic active radiation [μmol m⁻² s⁻¹]

LAIx

Leaf area fractions

a_max

Maximal photosynthetic rate [μmol m⁻² s⁻¹]

e

Flow rate [mol m⁻² s⁻¹]

ec

Critical flow rate [mol m⁻² s⁻¹]

kr_max

Maximal hydraulic conductance ratio

p_ups

Base xylem pressre [MPa]

p_old

Base xylem pressre memory [MPa]

Stomatal conductance

For empirical stomatal models, the stomatal conductance is computed as the intercept of two functions: an empirical function that describe stomatal responses to the physiological and environmental cues and an function that follows the diffusion nature of H₂O and CO₂. The abstractized function for the empirical correlation is

Land.StomataModels.empirical_gsw_from_model — Function

empirical_gsw_from_model(model::EmpiricalStomatalModel{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, β::FT) where {FT<:AbstractFloat}
empirical_gsw_from_model(model::EmpiricalStomatalModel{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}, β::FT) where {FT<:AbstractFloat}
empirical_gsw_from_model(model::EmpiricalStomatalModel{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}, β::FT, ind::Int) where {FT<:AbstractFloat}

Steady state gsw from empirical approach given

model EmpiricalStomatalModel type empirical model parameter set

leaf [Leaf] type struct

canopyi CanopyLayer type struct

envir [AirLayer] type struct

β Correction factor over the g1 part of an empirical model

ind Nth leaf in the canopy layer

For optimization stomatal models, the stomatal conductance is computed as the point where the marginal carbon gains equals the marginal carbon risk. The marginal carbon gain and risk are generally numerically computed by marginally increasing transpiration rate.

This module uses ConstrainedRootSolver module to iterate through the two functions to find the solution. The aim is to find the stomatal conductance when the envir_diff! function equals 0. The envir_diff! returns the diference between real and model-predicted conductances for empirical stomatal models, and the difference between marginal carbon gain and risk for optimization stomatal models.

Land.StomataModels.envir_diff! — Function

envir_diff!(x::FT,
            photo_set::AbstractPhotoModelParaSet{FT},
            canopyi::CanopyLayer{FT},
            hs::LeafHydraulics{FT},
            psoil::FT,
            swc::FT,
            envir::AirLayer{FT},
            sm::OptimizationStomatalModel{FT},
            bt::AbstractBetaFunction{FT},
            ind::Int) where {FT<:AbstractFloat}
envir_diff!(x::FT,
            photo_set::AbstractPhotoModelParaSet{FT},
            canopyi::CanopyLayer{FT},
            hs::LeafHydraulics{FT},
            envir::AirLayer{FT},
            sm::AbstractStomatalModel{FT},
            ind::Int) where {FT<:AbstractFloat}

Calculate the difference to be minimized for a given

x Assumed leaf diffusive conductance

photo_set[C3ParaSet] or [C4ParaSet] type parameter set

canopyiCanopyLayer type struct

hs Leaf hydraulic system

psoil Soil water potential [MPa]

swc Soil water content

envir[AirLayer] type struct

sm EmpiricalStomatalModel or OptimizationStomatalModel

bt AbstractBetaFunction type struct

ind Nth leaf in the canopy layer

The former function works for Ball-Berry, Leuning, and Medlyn models, the latter works for Gentine and all the optimization based models.

In the envir_diff! function, leaf photosynthetic rates is modeled using update_leaf_from_glc!, which calculates the gas exchange rates from a known total leaf diffusive conductance.

Land.StomataModels.update_leaf_from_glc! — Function

update_leaf_from_glc!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}, ind::Int, glc::FT) where {FT<:AbstractFloat}
update_leaf_from_glc!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Update Nth leaf photosynthesis, given

photo_set [C3ParaSet] or [C4ParaSet] type parameter set

canopyi CanopyLayer type struct

envir [AirLayer] type struct

ind Nth leaf

glc Given leaf diffusive conductance

Note that this function does not make the gsw control, so it is not guaranteed that gsw is within the physiological range. Thus, gsw control should be made outside this function. This function is supposed to be used in the optimal stomatl conductance models only, because optimal conductance can be outside the physiological stomatal conductance range. Thus, using this function for other purposes need to be cautious. In this case, it is recommended to use update_leaf_from_gsw!.

However, these functions do not force stomatal conductance to stay in its ranges. For example, the stomatal conductance solution is set to be zero if light is lower than the compensation point. In this case, the envir_diff! function has to be used along with a control function to guarantee realistic stomatal conductance.

Land.StomataModels.leaf_gsw_control! — Function

leaf_gsw_control!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}, ind::Int) where {FT<:AbstractFloat}
leaf_gsw_control!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

make sure g_sw is in its physiological range limited by diffusion, given

photo_set [C3ParaSet] or [C4ParaSet] type parameter set

canopyi CanopyLayer type struct

envir [AirLayer] type struct

ind Nth leaf

Note that this function is meant to use jointly with updateleaffrom_glc! when computing optimal stomtal conductance.

To facilitate the use of the StomataModels module, an abstractized function is provided for conveniently obtaining stomatal conductance from given environmental conditions.

Land.StomataModels.leaf_photo_from_envir! — Function

leaf_photo_from_envir!(
            photo_set::AbstractPhotoModelParaSet{FT},
            canopyi::CanopyLayer{FT},
            hs::LeafHydraulics{FT},
            psoil::FT,
            swc::FT,
            envir::AirLayer{FT},
            sm::EmpiricalStomatalModel{FT},
            bt::AbstractBetaFunction{FT}) where {FT<:AbstractFloat}
leaf_photo_from_envir!(
            photo_set::AbstractPhotoModelParaSet{FT},
            canopyi::CanopyLayer{FT},
            hs::LeafHydraulics{FT},
            psoil::FT,
            swc::FT,
            envir::AirLayer{FT},
            sm::EmpiricalStomatalModel{FT},
            bt::AbstractBetaFunction{FT},
            ind::Int) where {FT<:AbstractFloat}
leaf_photo_from_envir!(
            photo_set::AbstractPhotoModelParaSet{FT},
            canopyi::CanopyLayer{FT},
            hs::LeafHydraulics{FT},
            envir::AirLayer{FT},
            sm::AbstractStomatalModel{FT}) where {FT<:AbstractFloat}
leaf_photo_from_envir!(
            photo_set::AbstractPhotoModelParaSet{FT},
            canopyi::CanopyLayer{FT},
            hs::LeafHydraulics{FT},
            envir::AirLayer{FT},
            sm::AbstractStomatalModel{FT},
            ind::Int) where {FT<:AbstractFloat}

Calculate steady state gsw and photosynthesis from empirical approach, given

photo_set [C3ParaSet] or [C4ParaSet] type parameter set

canopyi CanopyLayer type struct

hs Leaf hydraulic system

envir [AirLayer] type struct

sm EmpiricalStomatalModel or OptimizationStomatalModel

bt AbstractBetaFunction type struct

ind Nth leaf in canopyi

The former function works for Ball-Berry, Leuning, and Medlyn models, the latter works for Gentine and all the optimization based models.

To speed up the calculations, leaf physiological parameters are updated only if the environmental conditions changes. For example, PAR (photosyntheis active radiation) is constant when we iterate envir_diff!, and the electron transport is only updated once. Similar for the cases of leaf temperature and soil moisture. This kind of functions used in the present module are

Land.StomataModels.update_leaf_TP! — Function

update_leaf_TP!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, hs::LeafHydraulics{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Update leaf physiological parameters if temperature or pressure changes in the daytime, given

photo_set [C3ParaSet] or [C4ParaSet] type parameter set

canopyi CanopyLayer type struct

hs Leaf hydraulic system

envir [AirLayer] type struct

Land.StomataModels.update_leaf_AK! — Function

update_leaf_AK!(
        photo_set::AbstractPhotoModelParaSet{FT},
        canopyi::CanopyLayer{FT},
        hs::LeafHydraulics{FT},
        envir::AirLayer{FT}) where {FT<:AbstractFloat}

Update leaf maximal A and K for Sperry model, given

photo_set [C3ParaSet] or [C4ParaSet] type parameter set

canopyi CanopyLayer type struct

envir [AirLayer] type struct

I'd like to emphasize it here that the leaf_photo_from_envir! function only applies to the case of constant leaf temperature because leaf energy budget is not calculated, and thus leaf_photo_from_envir! is only applicable to (1) known leaf temperature, and (2) prognostically modeling the non-steady state stomatal behaviors. As to the steady state case, leaf energy budget has to be considered. For the prognotic stomatal conductance, it is recommended to use update_leaf_from_gsw! function.

Land.StomataModels.update_leaf_from_gsw! — Function

update_leaf_from_gsw!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}, ind::Int, gsw::FT) where {FT<:AbstractFloat}
update_leaf_from_gsw!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}, ind::Int) where {FT<:AbstractFloat}
update_leaf_from_gsw!(photo_set::AbstractPhotoModelParaSet{FT}, canopyi::CanopyLayer{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Update Nth leaf photosynthesis, given

photo_set [C3ParaSet] or [C4ParaSet] type parameter set

canopyi CanopyLayer type struct

envir [AirLayer] type struct

ind Nth leaf

gsw Given stomatal conductance to H₂O

Note that this function makes the gsw control so that gsw is within the physiological range.

Note it here that stomtal conductance is controlled in this function, and thus no additional control like leaf_gsw_control! is required if update_leaf_from_gsw! is used.