EmeraldConstants

Universal Constants

EmeraldConstants.UniversalConstants — Type

mutable struct UniversalConstants

Structure to save universal constants.

Fields

AVOGADRO::Float64

: Avogadro's number [molecule mol⁻¹]

CP_D::Float64

: Isobaric specific heat of dry air [J kg⁻¹ K⁻¹]

CP_I::Float64

: Isobaric specific heat of ice water [J kg⁻¹ K⁻¹]

CP_L::Float64

: Isobaric specific heat of liquid water [J kg⁻¹ K⁻¹]

CP_V::Float64

: Isobaric specific heat of water vapor [J kg⁻¹ K⁻¹]

F_O₂::Float64

: O₂ fraction in air [-]

GAS_R::Float64

: Universal gas constant [J mol⁻¹ K⁻¹]

GRAVITY::Float64

: Gravity of the Earth [m s⁻²]

H_PLANCK::Float64

: Planck constant [m² kg s⁻¹]

K_BOLTZMANN::Float64

: Boltzmann constant [m² kg s⁻² K⁻¹]

K_STEFAN::Float64

: Stefan-Boltzmann constant [W m⁻² K⁻⁴]

K_VON_KARMAN::Float64

: Von Karman constant [-]

LH_V₀::Float64

: Latent heat vaporization at T₀ [K kg⁻¹]

LIGHT_SPEED::Float64

: Light speed in vacuum [m s⁻¹]

M_DRYAIR::Float64

: Molar mass of dry air [kg mol⁻¹]

M_H₂O::Float64

: Molar mass of water [kg mol⁻¹]

P_ATM::Float64

: Mean atmospheric pressure at sea level [Pa]

PRESS_TRIPLE::Float64

: Water vapor pressure at triple temperature [Pa]

T₀::Float64

: Freezing temperature of water [K]

T_TRIPLE::Float64

: Triple temperature of water [K]

YEAR_D::Float64

: Mean number of days per year [day]

Λ_THERMAL_H₂O::Float64

: Thermal conductivity of water [W m⁻¹ K⁻¹]

ρ_H₂O::Float64

: Density of liquid water [kg m⁻³]

EmeraldConstants.AVOGADRO — Function

Avogadro's number [molecule mol⁻¹]

EmeraldConstants.CP_D — Function

Isobaric specific heat of dry air [J kg⁻¹ K⁻¹]

EmeraldConstants.CP_D_MOL — Function

Isobaric specific heat of dry air per mole [J mol⁻¹ K⁻¹]

EmeraldConstants.CP_I — Function

Isobaric specific heat of ice water [J kg⁻¹ K⁻¹]

EmeraldConstants.CP_I_MOL — Function

Isobaric specific heat of ice water per mole [J mol⁻¹ K⁻¹]

EmeraldConstants.CP_L — Function

Isobaric specific heat of liquid water [J kg⁻¹ K⁻¹]

EmeraldConstants.CP_L_MOL — Function

Isobaric specific heat of liquid water per mole [J mol⁻¹ K⁻¹]

EmeraldConstants.CP_V — Function

Isobaric specific heat of water vapor [J kg⁻¹ K⁻¹]

EmeraldConstants.CP_V_MOL — Function

Isobaric specific heat of water vapor per mole [J mol⁻¹ K⁻¹]

EmeraldConstants.F_O₂ — Function

O₂ fraction in air [-]

EmeraldConstants.GAS_R — Function

Universal gas constant [J mol⁻¹ K⁻¹]

EmeraldConstants.GRAVITY — Function

Gravity of the Earth [m s⁻²]

EmeraldConstants.H_PLANCK — Function

Planck constant [m² kg s⁻¹]

EmeraldConstants.K_BOLTZMANN — Function

Boltzmann constant [m² kg s⁻² K⁻¹]

EmeraldConstants.K_STEFAN — Function

Stefan-Boltzmann constant [W m⁻² K⁻⁴]

EmeraldConstants.K_VON_KARMAN — Function

Von Karman constant [-]

EmeraldConstants.LH_V₀ — Function

Latent heat vaporization at T₀ [K kg⁻¹]

EmeraldConstants.LIGHT_SPEED — Function

Light speed in vacuum [m s⁻¹]

EmeraldConstants.M_DRYAIR — Function

Molar mass of dry air [kg mol⁻¹]

EmeraldConstants.M_H₂O — Function

Molar mass of water [kg mol⁻¹]

EmeraldConstants.P_ATM — Function

Mean atmospheric pressure at sea level [Pa]

EmeraldConstants.PRESS_TRIPLE — Function

Water vapor pressure at triple temperature [Pa]

EmeraldConstants.R_V — Function

Gas constant water vapor [J kg⁻¹ K⁻¹]

EmeraldConstants.RT₂₅ — Function

Gas constant times 298.15 K [J mol⁻¹]

EmeraldConstants.T₀ — Function

Freezing temperature of water [K]

EmeraldConstants.T₂₅ — Function

Kelvin temperature at 25 Celcius [K]

EmeraldConstants.T_TRIPLE — Function

Triple temperature of water [K]

EmeraldConstants.V_H₂O — Function

Molar volume of liqiud water

EmeraldConstants.YEAR_D — Function

Mean number of days per year [day]

EmeraldConstants.Λ_THERMAL_H₂O — Function

Thermal conductivity of water [W m⁻¹ K⁻¹]

EmeraldConstants.ρ_H₂O — Function

Density of liquid water [kg m⁻³]

EmeraldConstants.ρg_MPa — Function

Density of water times gravity [MPa m⁻¹]