API

Vulnerability curve

ClimaCache.BrooksCorey — Method

BrooksCorey{FT}(vg::VanGenuchten{FT}) where {FT<:AbstractFloat}

A constructor for BrooksCorey to create BrooksCorey type soil from VanGenuchten type, given

vg VanGenuchten type soil water retention curve

SoilHydraulics.soil_θ — Function

soil_θ(bc::BrooksCorey{FT}, ψ_25::FT) where {FT<:AbstractFloat}
soil_θ(vg::VanGenuchten{FT}, ψ_25::FT) where {FT<:AbstractFloat}

Return the soil water content, given

bc or vg BrooksCorey or VanGenuchten type structure

ψ_25 Soil metric potential corrected to 25 Celcius

SoilHydraulics.soil_ψ_25 — Function

soil_ψ_25(bc::BrooksCorey{FT}, θ::FT; oversaturation::Bool = false) where {FT<:AbstractFloat}
soil_ψ_25(vg::VanGenuchten{FT}, θ::FT; oversaturation::Bool = false) where {FT<:AbstractFloat}

Return the soil metric potential, given

bc or vg BrooksCorey or VanGenuchten type structure

θ Soil volumetric water content (absolute value)

oversaturation If true, allow for soil water oversaturation

SoilHydraulics.relative_hydraulic_conductance — Function

relative_hydraulic_conductance(bc::BrooksCorey{FT}, θ::FT) where {FT<:AbstractFloat}
relative_hydraulic_conductance(bc::BrooksCorey{FT}, ψ::Bool, ψ_25::FT) where {FT<:AbstractFloat}
relative_hydraulic_conductance(vg::VanGenuchten{FT}, θ::FT) where {FT<:AbstractFloat}
relative_hydraulic_conductance(vg::VanGenuchten{FT}, ψ::Bool, ψ_25::FT) where {FT<:AbstractFloat}

Return the relative hydraulic conductance of the soil, given

bc or vg BrooksCorey or VanGenuchten type structure

θ Soil volumetric water content (absolute value)

ψ Bool to indicate that next parameter is potential

ψ_25 Soil metric potential at a reference temperature of 25 °C

Soil energy+water budget

SoilHydraulics.root_sink — Function

root_sink(root::Root{FT}) where {FT<:AbstractFloat}

Return root water update, given

root Root type struct that may contain non- and steady state flow

SoilHydraulics.soil_budget! — Function

This function have two major features:

Compute the marginal change of soil water content and energy

Update soil water content and energy without over-saturating or draining the soil

SoilHydraulics.soil_budget! — Method

soil_budget!(spac::Union{MonoMLGrassSPAC{FT}, MonoMLPalmSPAC{FT}, MonoMLTreeSPAC{FT}}) where {FT<:AbstractFloat}

Update the marginal increase of soil water content and energy per layer, given

spac MonoMLGrassSPAC, MonoMLPalmSPAC, or MonoMLTreeSPAC SPAC

SoilHydraulics.soil_budget! — Method

soil_budget!(spac::Union{MonoMLGrassSPAC{FT}, MonoMLPalmSPAC{FT}, MonoMLTreeSPAC{FT}}, δt::FT) where {FT<:AbstractFloat}

Run soil water and energy budget, given

spac MonoMLGrassSPAC, MonoMLPalmSPAC, or MonoMLTreeSPAC SPAC

δt Time step