Leaf module in the LAND model

Types and parameter sets

Leaf boundary layer

Land.Leaf.AbstractLeafBLParaSet — Type

AbstractLeafBLParaSet

A parameter set that stores boundary layer parameters

Land.Leaf.LeafBLParaSetFixed — Type

struct LeafBLParaSetFixed{FT, resistance}

A leaf boundary layer with fixed resistance, namely ensuring surface concentration

Land.Leaf.LeafBLParaSetGentine — Type

struct LeafBLParaSetGentine

Gentine's boundary layer scheme that computes Monin Obhukow length and resistances across the leaf to Ca scale

Land.Leaf.BLFixed — Function

Create a fixed boundary layer with 0 resistance

Leaf photosynthesis

Land.Leaf.update_leaf_TD! — Function

update_leaf_TD!(paraset::C3ParaSet, leaf::LeafParams{FT})
update_leaf_TD!(paraset::C4ParaSet, leaf::LeafParams{FT})

Update leaf photosynthesis-related temperature dependencies, given

paraset A C3ParaSet or C4ParaSet type parameter set

leaf A LeafParams struct

Land.Leaf.electron_transport_rate! — Function

electron_transport_rate!(leaf::LeafParams{FT}, APAR::FT)

Update the electron transport variables in the leaf struct, given

paraset A C3ParaSet type parameter set

leaf A LeafParams struct

APAR Absorbed PAR

Land.Leaf.rubisco_limited_rate! — Function

rubisco_limited_rate!(paraset::C3ParaSet, leaf::LeafParams{FT})
rubisco_limited_rate!(paraset::C4ParaSet, leaf::LeafParams{FT})

Update the Rubisco-limited photosynthesis in the leaf struct, given

paraset A C3ParaSet or C4ParaSet type parameter set

leaf A LeafParams struct

Land.Leaf.light_limited_rate! — Function

light_limited_rate!(paraset::C3ParaSet, leaf::LeafParams{FT}, APAR::FT)
light_limited_rate!(paraset::C4ParaSet, leaf::LeafParams{FT}, APAR::FT)

Update the Electron Transport-limited photosynthesis in the leaf struct, given

paraset A C3ParaSet or C4ParaSet type parameter set

leaf A LeafParams struct

APAR Absorbed PAR

Land.Leaf.product_limited_rate! — Function

product_limited_rate!(paraset::C3ParaSet, leaf::LeafParams{FT})
product_limited_rate!(paraset::C4ParaSet, leaf::LeafParams{FT})

Update the Product-limited photosynthesis in the leaf struct, given

paraset A C3ParaSet or C4ParaSet type parameter set

leaf A LeafParams struct

Land.Leaf.leaf_fluorescence! — Function

leaf_fluorescence!(paraset::FluoParaSet, leaf::LeafParams{FT})

Compute fluorescence yield, Kn, and Kp for leaf, given

paraset A FluoParaSet type parameter set

leaf A LeafParams struct