Photosynthesis module for the LAND model

Types and parameter sets

Temperature dependency

Land.Photosynthesis.AbstractTDParameterSet — Type

AbstractTDParameterSet

Hierarchy of the AbstractTDParameterSet:

AbstractArrheniusTDParameterSet

AbstractArrheniusPeakTDParameterSet

Land.Photosynthesis.AbstractArrheniusTDParameterSet — Type

AbstractArrheniusTDParameterSet

A parameter set type that uses Arrhenius correction using correction = exp( c - ΔHa/(RT) ).

Land.Photosynthesis.AbstractArrheniusPeakTDParameterSet — Type

AbstractArrheniusPeakTDParameterSet

A parameter set type that uses Arrhenius peak correction using

    C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )]
    f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ]
    f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ]
    correction = f_above /f_below

Land.Photosynthesis.ArrheniusTD — Type

struct ArrheniusTD{FT,val_25,ΔHa}

An AbstractArrheniusTDParameterSet type struct that takes the following to initialize

FT Floating type

val_25 Uncorrected value at 298.15 K

ΔHa Heat of corrosion reaction

Fields

VAL_25

Uncorrected value at 298.15 K

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

Land.Photosynthesis.ArrheniusPeakTD — Type

struct ArrheniusPeakTD{FT,val_25,ΔHa}

An AbstractArrheniusPeakTDParameterSet type struct that takes the following to initialize

FT Floating type

ΔHa Heat of corrosion reaction

ΔHd Dimerization enthalpy

ΔSv Entropy

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Land.Photosynthesis.KcTDBernacchi — Function

Create Kc TD from Bernacchi's data

Land.Photosynthesis.KoTDBernacchi — Function

Create Ko TD from Bernacchi's data

Land.Photosynthesis.RespirationTDBernacchi — Function

Create Respiration TD from Bernacchi's data

Land.Photosynthesis.VcmaxTDBernacchi — Function

Create Vcmax TD from Bernacchi's data

Land.Photosynthesis.VomaxTDBernacchi — Function

Create Vomax TD from Bernacchi's data

Land.Photosynthesis.ΓStarTDBernacchi — Function

Create Γ* TD from Bernacchi's data

Land.Photosynthesis.KpepTDBoyd — Function

Create Kpep TD from Boyd's data

Land.Photosynthesis.VpmaxTDBoyd — Function

Create Vpmax TD from Boyd's data

Land.Photosynthesis.JmaxTDLeuning — Function

Create Jmax TD from Leuning's data

Land.Photosynthesis.VcmaxTDLeuning — Function

Create Vcmax TD from Leuning's data

Land.Photosynthesis.KcTDCLM — Function

Create Kc TD

Land.Photosynthesis.KoTDCLM — Function

Create Ko TD

Land.Photosynthesis.KpepTDCLM — Function

Create Kpep TD

Land.Photosynthesis.ΓStarTDCLM — Function

Create Γ* TD

Land.Photosynthesis.JmaxTDBernacchi — Function

Create Jmax TD

Land.Photosynthesis.JmaxTDCLM — Function

Create Jmax TD

Land.Photosynthesis.RespirationTDCLM — Function

Create Respiration TD

Land.Photosynthesis.VcmaxTDCLM — Function

Create Vcmax TD

Land.Photosynthesis.VtoRCollatz — Function

A constant of 0.01

Land.Photosynthesis.VtoRDefault — Function

A constant of 0.015

Photosynthesis model

Land.Photosynthesis.AbstractPhotoModelParaSet — Type

AbstractPhotoModelParaSet

A parameter set that stores TD parasets and constants

Land.Photosynthesis.C3ParaSet — Type

struct C3Paraset{FT, JTD, KcTD, KoTD, ReTD, VcTD, ΓsTD, V2R, eff_1, eff_2}

An AbstractPhotoModelParaSet type of parameter sets for different temperature dependencies.

Fields

JT

Jmax temperature dependency

KcT

Kc temperature dependency

KoT

Ko temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

ΓsT

Γ_star temperature dependency

VR

Vcmax25 and respiration correlation

Eff_1

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

Eff_2

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

Land.Photosynthesis.C4ParaSet — Type

struct C4ParaSet{FT, KpTD, ReTD, VcTD, VpTD, V2R}

An AbstractPhotoModelParaSet type of parameter sets for different temperature dependencies.

Fields

KpT

Kpep temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

VpT

Vpmax temperature dependency

VR

Vcmax25 and respiration correlation

Land.Photosynthesis.C3Bernacchi — Function

Bernacchi set for C3 photosynthesis

Land.Photosynthesis.C3CLM — Function

CLM set for C3 photosynthesis

Land.Photosynthesis.C4CLM — Function

CLM set for C4 photosynthesis

Fluorescence

Land.Photosynthesis.AbstractFluoModelParaSet — Type

AbstractFluoModelParaSet

A parameter set that stores fluorescence model parameters

Land.Photosynthesis.FluoParaSet — Type

struct FluoParaSet{FT, kn1, kn2, kn3}

A AbstractFluoModelParaSet type paramter set.

Fields

Kn1

Fluorescence model coefficient

Kn2

Fluorescence model coefficient

Kn3

Fluorescence model coefficient

Land.Photosynthesis.FluorescenceFlexas — Function

Create Fluorescence parameter set using Flexas's data

Photosynthesis model

Land.Photosynthesis.arrhenius_correction — Function

arrhenius_correction(td_set::ArrheniusTD, T::FT)
arrhenius_correction(td_set::ArrheniusPeakTD, T::FT)

A correction factor based on arrhenius's fitting procedure, given

td_set An ArrheniusTD or ArrheniusPeakTD type struct

T Leaf temperature in [K]

The equation used for ArrheniusTD is correction = exp( c - ΔHa/(R*T_leaf) ).

The equations used for ArrheniusPeakTD are

    C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )]
    f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ]
    f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ]
    correction = f_above /f_below

Land.Photosynthesis.get_jmax — Function

get_jmax(td_set::AbstractJmaxTD, j25::FT, T::FT)

Maximal electron transport rate at leaf temperature, given

td_set AbstractJmax parameter set

j25 Maximal electron transport rate at 298.15 K

T Leaf temperature

Land.Photosynthesis.get_j — Function

get_j(jmax::FT, par::FT, curvature::FT, qy::FT)

Electron transport rate j, given

jmax Maximal eclectron transport @ leaf temperature (not 298.15 K)

par Absorbed PAR (photosynthetic active radiation)

curvature Curvature parameter (default at 0.9)

qy Quantum yield of electron, qy = maxPSII * PSII_frac, maxPSII is maximal PSII yield (default at 4.0/4.9), PSII_frac is Fraction of absorbed light used by PSII ETR (default at 0.5)

Land.Photosynthesis.get_kc — Function

get_kc(td_set::AbstractKc, T::FT)

Kc at leaf temperature, given

td_set One AbstractKc type that store temperature correction information

T Leaf temperature

Land.Photosynthesis.get_ko — Function

get_ko(td_set::AbstractKo, T::FT)

Ko at leaf temperature, given

td_set One AbstractKo type that store temperature correction information

T Leaf temperature

Land.Photosynthesis.get_kpep — Function

get_kpep(td_set::AbstractKpep, T::FT)

Kpep at leaf temperature, given

td_set One AbstractKpep type that store temperature correction information

T Leaf temperature

Land.Photosynthesis.get_r — Function

get_r(td_set::AbstractRespirationTD, r25::FT, T::FT)

Leaf respiration rate r_leaf at leaf temperature, given

td_set One AbstractRespirationTD type that store temperature correction information

r25 Leaf respiration rate at 298.15 K (25 Celcius)

T Leaf temperature

Land.Photosynthesis.get_vmax — Function

get_vmax(v25::FT, td_set::AbstractVmaxTD, T::FT)

Maximal electron transport rate at leaf temperature, given

td_set AbstractVmaxTD parameter set

v25 Maximal carboxylation rate at 298.15 K

T Leaf temperature

Land.Photosynthesis.get_Γ_star — Function

get_Γ_star(td_set::AbstractΓStarTD, T::FT)

Γ_star at leaf temperature, given

td_set One AbstractΓStarTD type that store temperature correction information

T Leaf temperature

Land.Photosynthesis.an_ag_r_from_pi — Function

an_ag_r_from_pi(paraset::C3ParaSet,
                    p_i::FT,
                    v25::FT,
                    j25::FT,
                    par::FT,
                   p_O₂::FT,
                      T::FT,
                    r25::FT,
              curvature::FT,
                     qy::FT)

Gross photosynthetic rate limited by carboxylation with ATP limitation, given

paraset A C3ParaSet type parameter sets that stores temperature denpendencies

p_i Leaf internal CO₂ partial pressure

v25 Maximal carboxylation rate at 298.15 K

j25 Maximal electron transport rate

par Absorbed photosynthetic active radiation

p_O₂ O₂ partial pressure

T Leaf temperature

r25 Respiration rate at 298.15 K

curvature Curvature parameter to calculate J

qy Quantum yield of electron

The equations used are from Farquhar et al. (1980) "A biochemical model of photosynthetic CO₂ assimilation in leaves of C3 species."

an_ag_r_from_pi(paraset::C4ParaSet,
                    p_i::FT,
                    v25::FT,
                    p25::FT,
                    par::FT,
                      T::FT,
                    r25::FT,
                     qy::FT)

Gross photosynthetic rate limited by carboxylation, given

paraset A C4ParaSet type parameter sets that stores temperature denpendencies

p_i Leaf internal CO₂ partial pressure

v25 Maximal carboxylation rate at 298.15 K

p25 Maximal PEP carboxylation rate at 298.15 K

par Absorbed photosynthetic active radiation

T Leaf temperature

r25 Respiration rate at 298.15 K

qy Quantum yield of electron

The model is adapted from Collatz et al. (1992) "Coupled photosynthesis-stomatal conductance model for leaves of C4 plants."

Land.Photosynthesis.an_ag_r_pi_from_gsc — Function

an_ag_r_pi_from_gsc(paraset::AbstractPhotoModelParaSet,
                        gsc::FT,
                        v25::FT,
                        j25::FT,
                        p25::FT,
                        p_a::FT,
                          T::FT,
                        par::FT,
                      p_atm::FT,
                       p_O₂::FT,
                        r25::FT,
                  curvature::FT,
                         qy::FT)

Net photosynthetic rate tar_an, gross photosynthetic rate tar_ag, respiration rate tar_r, and leaf internal CO₂ partial pressure tar_p, given

paraset A C3ParaSet or C4ParaSet type parameter set

gsc Leaf diffusive conductance to CO₂

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plant only

p_a Atmospheric CO₂ partial pressure

T Leaf temperature

par Photosynthetic active radiation

p_atm Atmospheric pressure

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

curvature Curvature parameter to calculate J

qy Quantum yield of electron

For C3 plants, with ATP limitation. Some parameters are useless in C3 photosynthesis, like p25.

For C4 plants. Some parameters are useless for C4 photosynthesis, like j25, p_O₂, and curvature.