Plant module in the LAND model

Plant structure

Land.Plant.RootLayer — Type

RootLayer{FT<:AbstractFloat}

A RootLayer type which contains root hydraulics information, including rhizosphere conductance.

Fields

z_hi

Upper root depth [m]

z_lo

Lower root depth [m]

r_layer

Average horizental root length [m]

f_layer

Fraction of biomass and k_max

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance, default at 2700.0 Kg h⁻¹ MPa⁻¹ m⁻² basal area [mol s⁻¹ MPa⁻¹]

k_s

Maximal hydraulic conductivity per cross section basal area per root depth [mol s⁻¹ MPa⁻¹ m⁻²]

k_rhiz

Maximal rhizosphere conductance, default at 9.72e12 Kg h⁻¹ MPa⁻¹ mm⁻² basal area [mol s⁻¹ MPa⁻¹]

p_ups

Soil water potential (upstream) [MPa]

soil_a

Soil texture parameter

soil_m

Soil texture parameter, 1 - 1/n

soil_n

Soil texture parameter

soil_msc

Maximal soil volumatric water content, vary with rock fraction

soil_rwc

Relative soil volumatric water content

soil_vwc

Soil volumatric water content

t_soil

Soil temperature [K]

soil_type

Soil texture

p_rhiz

Water potential at the root-rhizosphere interface [MPa]

p_dos

Xylem water pressure at the tree base (downstream) [MPa]

q

Flow rate in the root layer [mol s⁻¹]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹]

p_element

List of xylem water pressure per element [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of water temperature per element [K]

z_element

List of height per element to account for gravity [m]

Land.Plant.create_root_list — Function

create_root_list(FT, n, z_lo)

The function creates a root system with n root layers evenly in the soil. This function need to work with another function (pending) to quantify the root distribution.

FT Floating type for the RootLayer struct

n Total number of root layers in the root system

z_lo Maximal root depth

Land.Plant.Stem — Type

Stem{FT<:AbstractFloat}

A Stem type which contains hydraulic information, including drought history profile. The Stem type is used for creating a trunk and a list of branches.

Fields

r_layer

Horizental radius of the layer [m]

z_hi

Upper height from the ground (trunk-branch joint for trunk, stem-leaf joint for branch) [m]

z_lo

Lower height from the ground (0 for trunk, trunk-branch joint for branch) [m]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance, = 5400.0 * 8/3 Kg h⁻¹ MPa⁻¹ m⁻² basal area [mol s⁻¹ MPa⁻¹]

k_s

Maximal hydraulic conductivity per cross section area per tree height [mol s⁻¹ MPa⁻¹ m⁻²]

p_ups

Xylem water pressure at the tree basa (upstream) [MPa]

p_dos

Xylem water pressure at the trunk-stem or stem-leaf joint (downstream) [MPa]

q

Flow rate in the xylem [mol s⁻¹]

k_element

List of stem k_max per element [mol s⁻¹ MPa⁻¹]

p_element

List of stem xylem pressure per element [MPa]

p_history

List of stem xylem pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of stem temperature for each element [K]

z_element

List of stem height per element to account for gravity [m]

Land.Plant.create_branch_list — Function

create_branch_list(FT, n, z_lo, z_hi)

This function creates a list of stem structs as branch system for the tree, given

FT Floating type for the Stem struct

n Total number of canopy layers, one branch for a canopy layer

z_lo Minimal canopy distance from the ground

z_hi Maximal canopy distance from the ground

Each branch (Stem) is responsible for a canopy layer. The branch index from 1:n is from lower to higher canopy.

Land.Plant.Leaf — Type

Leaf{FT<:AbstractFloat}

A Leaf type which contains leaf hydraulics information.

Fields

angle_locat

direction from the tree trunk to the leaf, 0 for east, 270 for south, and 180 for west [°]

angle_incli

direction of leave, 0 for flat and 90 for vertical [°]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_sla

Maximal leaf hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

p_ups

Leaf xylem water pressure (different from leaf water potential) at the leaf base (upstream) [MPa]

p_dos

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

k_element

List of leaf k_max per element (mol s⁻¹ MPa⁻¹)

p_element

List of xylem water pressure per element [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of xylem water temperature per element [K]

z_element

List of leaf element height change to account for gravity, default = 0 [m]

Land.Plant.CanopyLayer — Type

CanopyLayer{FT<:AbstractFloat, n_leaf}

A CanopyLayer type, which constains environemntal conditions and leaf-level fluxes for n_leaf Leaf. The n_leaf is the sum of n_Ari * n_Incli sunlit leaves and 1 shaded leaves.

Fields

f_layer

Leaf area fraction in the canopy

f_view

Fraction of the sunlit leaves

la

Leaf area in the layer [m²]

width

Leaf width [m]

p_a

Atmospheric CO₂ partial pressure in the layer [Pa]

p_atm

Atmospheric pressure [Pa]

p_O₂

Atmospheric O₂ partial presssure [Pa]

p_H₂O

Atmospheric H₂O partial pressure [Pa]

t_air

Air temperature [K]

wind

Wind speed [m s⁻¹]

g_ias_c

Mesophyll conductance correction factor: multiplier

g_ias_e

Mesophyll conductance correction factor: exponent

g_max

Maximal leaf diffusive conductance for H₂O at 298.15 K [mol m⁻² s⁻¹]

g_min

Minimal leaf diffusive conductance for H₂O at 298.15 K [mol m⁻² s⁻¹]

gs_nssf

Non-steady state factor for optimization approach (use it by multiplying the ∂A/∂E - ∂Θ/∂E)

gs_empi

Non-steady state factor for empirical apporach (used by multiplying the gswmod - gswcurr)

j_max

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

r_25

Leaf respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_max

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

p_max

Maximal PEP carboxylation rate at 298.15 K (only for C4 plants) [μmol m⁻² s⁻¹]

curvature

Curvature ratio for J

qy

Quntum yield for electron, using aPAR

leaf_list

List of Leaf

ag_list

List of gross photosynthetic rate for n_leaf leaves [μmol m⁻² s⁻¹]

an_list

List of net photosynthetic rate [μmol m⁻² s⁻¹]

am_list

List if maximal A at the given scenario [μmol m⁻² s⁻¹]

d_list

List of leaf-to-air vapor pressure deficit [unitless]

e_list

List of flow rate per leaf area [mol m⁻² s⁻¹]

ec_list

List of critical flow rate (when leaf xylem pressure induces desiccation) per leaf area [mol m⁻² s⁻¹]

gsc_list

List of effective leaf diffusive conductance for CO₂ [mol m⁻² s⁻¹]

gsw_empi

List of empirical leaf diffusive conductance for H₂O [mol m⁻² s⁻¹]

gsw_list

List of leaf diffusive conductance for H₂O [mol m⁻² s⁻¹]

km_list

List of maximal hydraulic conductance [mol m⁻² s⁻¹]

la_list

List of leaf area per leaf [m²]

par_list

List of PAR for each leaf [μmol m⁻² s⁻¹]

pi_list

List of leaf internal CO₂ partial pressure [Pa]

q_list

List of flow rate [mol s⁻¹]

r_list

List of leaf respiration rate [μmol m⁻² s⁻¹]

t_list

List of leaf temperature [K]

Land.Plant.Tree — Type

Tree{FT<:AbstractFloat}

A Tree type which includes

root_no number of root layers

a trunk

canopy_no branches

leaf_no canopy layers

Fields

age

Age [year]

ba

Basal area [m²]

ga

Ground area [m²]

h

Tree height [m]

photo_type

Photosynthesis type C3/C4/CAM

photo_para_set

Photosynthesis model parameter set, type undefined

stomata_scheme

Optimization model option, type undefined

steady_state

Use steady-state or non-steady-state stomatal behavior

n_root

Number of root layers

root_frac

List of root fraction

root_zs

List of root z lower boundary

root_list

List of RootLayer

n_canopy

Number of canopy layers

trunk

Trunk using type Stem

branch_list

List of Stem

n_leaf

Number of leaves in each canopy layer

canopy_list

List of CanopyLayer

Stomatal models

Land.Plant.AbstractStomatalModel — Type

AbstractStomatalModel

The AbstractStomatalModel type includes AbstractEmpiricalStomatalModel and AbstractOptimizationStomatalModel

Land.Plant.AbstractEmpiricalStomatalModel — Type

AbstractEmpiricalStomatalModel

The AbstractEmpiricalStomatalModel type include different empirical structs

Land.Plant.ESMBallBerry — Type

ESMBallBerry{FT}

An empirical model parameter set type for Ball-Berry type model. The equation used for Ball-Berry type model is gs = g0 + g1 * RH * A/Ci. Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [unitless]

Land.Plant.ESMGentine — Type

ESMGentine{FT}

An empirical model parameter set type for Gentine type model. The equation used for Gentine type model is gs = g0 + g1 * k_leaf/k_max * A/Ci. Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required. Note it that the Gentine model does not require for a β function to tune the soil drought response, but the use of k_leaf also does not permit post-drought stomatal response unless k_leaf can be recovered.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [unitless]

Land.Plant.ESMLeuning — Type

ESMLeuning{FT}

An empirical model parameter set type for Leuning type model. The equation used for Leuning type model is gs = g0 + g1 * A/(Ci-Γ_star) * 1/(1+VPD/d0). Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [unitless]

d0

fitting parameter of d/d0 below the fraction, same unit as vpd [Pa]

Land.Plant.ESMMedlyn — Type

ESMMedlyn{FT}

An empirical model parameter set type for Medlyn type model. The equation used in Medlyn type model is gs = g0 + (1+g1/sqrt(VPD)) * A/Ci. Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [Pa⁽⁵⁾]

Land.Plant.AbstractOptimizationStomatalModel — Type

AbstractOptimizationStomatalModel

The AbstractOptimizationStomatalModel includes different optimization structs

Land.Plant.OSMEller — Type

OSMEller

An optimization model parameter set type for Eller model. The equation used for Eller model is ∂Θ/∂E = -∂K/∂E * A/K, where K is ∂E/∂P.

Fields

Land.Plant.OSMSperry — Type

OSMSperry

An optimization model parameter set type for Sperry model. The equation used for Sperry model is ∂Θ/∂E = -∂K/∂E * Amax/Kmax, where K is ∂E/∂P.

Fields

Land.Plant.OSMWang — Type

OSMWang

An optimization model parameter set type for Eller type model. The equation used for Eller model is ∂Θ/∂E = A / (E_crit - E).

Fields

Land.Plant.OSMWAP — Type

OSMWAP{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model. The equation used for Wolf-Anderegg-Pacala model is ∂Θ/∂E = (2aP+b) / K, where K is ∂P/∂E.

Fields

a

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻²]

b

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻¹]

Land.Plant.OSMWAPMod — Type

OSMWAPMod{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model, modified by adding a photosynthesis component while set b and c = 0. The equation used for modified Wolf-Anderegg-Pacala model is ∂Θ/∂E = a * P, where P is absolute value of leaf xylem pressure.

Fields

a

Quadratic equation parameter [mol mol⁻¹ MPa⁻¹]

Pant hydraulics

Land.Plant.get_p_base_q_list_from_q — Function

get_p_base_q_list_from_q(tree, flow)

Tree base pressyre p_base and a list of flow rate in each root layer q_list, given

tree A Tree type

flow Flow rate for the whole tree [mol s⁻¹]

This function returns the p_base from total transpiration rate using Newton Raphson. A warning will display if total iterations is beyond 50 times.

Land.Plant.get_q_layer_from_p_base — Function

get_q_layer_from_p_base(root_layer, p_base)

Flow rate in a root layer q_layer, given

root_layer One RootLayer in Tree

p_base Xylem water pressure at the tree base

This function use Newton Raphson combined with Bi-section to get q_layer. A warning message will display if total iterations exceed 50 times.

Land.Plant.get_struct_p_end_from_q — Function

get_struct_p_end_from_q(root_layer::RootLayer, flow::FT; p_ini::FT=FT(Inf))
get_struct_p_end_from_q(stem::Stem, flow::FT; p_ini::FT=FT(Inf))
get_struct_p_end_from_q(leaf::Leaf, flow::FT; p_ini::FT=FT(Inf))

End pressure p_end (in MPa), including impacts from rhizosphere conductance and gravity, given

root_layer One RootLayer in Tree

stem Stem as Trunk in Tree or Stem as side branch in Tree

leaf Leaf in CanopyLayer in Tree

flow Flow rate (in mol s⁻¹) in the given RootLayer

p_ini Upstream soil water potential if p_ini is given, otherwise (Inf) root_layer.p_ups will be used

Land.Plant.update_struct_from_q! — Function

update_struct_from_q!(root_layer::RootLayer, flow::FT; p_ini::FT=FT(Inf))
update_struct_from_q!(stem::Stem, flow::FT, p_ini::FT=FT(Inf))
update_struct_from_q!(leaf::Leaf, flow::FT, p_ini::FT=FT(Inf))

Update hydraulic information considering the impact from rhizosphere conductance and gravity, given

root_layer One RootLayer in Tree

stem Stem as trunk in a Tree or as side branch in Tree

leaf Leaf type in CanopyLayer in a Tree

flow Flow rate in the given root layer

p_ini Upstream soil water potential if p_ini is given, otherwise (Inf) root_layer.p_ups will be used

Land.Plant.update_leaf_ak_max! — Function

update_leaf_ak_max!(tree::Tree)

Update leaf maximal A_net allowed at the given environmental conditions, given -tree A Tree type struct

This function is meant to work with Sperry et al. (2017) model, which requires a_max in the ∂Θ∂E function.

Land.Plant.update_tree_e_crit! — Function

update_tree_e_crit!(tree::Tree; displaying::Bool=false)

Update the critical transpiration rate information, given

tree A Tree type

displaying If true, messages of how e_crit evolves step-by-step will display

This function update e_crit for each layer (sunlit and shaded). The e_crit for each leaf is either increased or decreased so that p_leaf is in the range -Inf to -20.0 MPa. The e_crit increases by 1e-5 mol m⁻² s⁻¹ if p_leaf is less negative than -20 MPa. The e_crit decreases by 1e-6 mol m⁻² s⁻¹ if p_leaf is -Inf.

Optimality stomatal model

Land.Plant.get_marginal_gain — Function

get_marginal_gain(canopyi::CanopyLayer, indx::Number, photo_para_set::AbstractPhotoModelParaSet)
get_marginal_gain(canopyi::CanopyLayer, photo_para_set::AbstractPhotoModelParaSet)

Marginal water use efficiency ∂A/∂E for a leaf in a canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

photo_para_set An AbstractPhotoModelParaSet type parameter set

Land.Plant.get_marginal_penalty — Function

get_marginal_penalty(canopyi::CanopyLayer, indx::Number, scheme::AbstractOptimizationStomatalModel)
get_marginal_penalty(canopyi::CanopyLayer, scheme::AbstractOptimizationStomatalModel)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

scheme An AbstractOptimizationStomatalModel type optimizaiton stomatal model scheme

OSMEller scheme uses the Eller et al. (2018) model. ∂Θ/∂E = A / K * dK/dE, unit [mol mol⁻¹]. To make the Eller model suitable for multiple layer canopy, K at the leaf xylem end is used. Also, to make the model capable of handling drought history, loss of conductance is evaluated assuming no drought legacy from a more negative p_leaf.

OSMSperry uses the Sperry et al. (2017) model. ∂Θ/∂E = A_max / K_max * dK/dE. To make the Sperry model suitable for multiple layer canopy, K at the leaf xylem end is used. Also, to make the model capable of handling drought history, loss of conductance is evaluated assuming no drought legacy from a more negative p_leaf.

OSMWang function uses the Wang et al. (2020) model. ∂Θ/∂E = A / (e_crit - e_leaf).

OSMWAP uses the Anderegg et al. (2018) model. ∂Θ/∂E = (2*a*P + b) / K.

OSMWAPMod uses modified Anderegg et al. (2018) model. ∂Θ/∂E = a * A * P / K.

Empirical stomatal model

Land.Plant.get_empirical_gsw_from_model — Function

get_empirical_gsw_from_model(model::AbstractEmpiricalStomatalModel,
                             an::FT,
                             p_atm::FT,
                             p_i::FT,
                             rh::FT,
                             vpd::FT,
                             Γ_star::FT,
                             beta::FT)
get_empirical_gsw_from_model(model::AbstractEmpiricalStomatalModel,
                             an::Array,
                             p_atm::FT,
                             p_i::Array,
                             rh::FT,
                             vpd::Array,
                             Γ_star::Array,
                             beta::Array)
get_empirical_gsw_from_model(model::ESMGentine,
                             an::FT,
                             p_atm::FT,
                             p_i::FT,
                             beta::FT)
get_empirical_gsw_from_model(model::ESMGentine,
                             an::Array,
                             p_atm::FT,
                             p_i::Array,
                             beta::Array)

Steady stage gsw from empirical approach given

model An AbstractEmpiricalStomatalModel type empirical model parameter set

an (Array of) Net photosynthetic rate [μmol m⁻² s⁻¹]

p_atm Atmospheric pressure

p_i (Array of) Leaf internal CO₂ partial pressure

rh Relative humidity

vpd (Array of) Vapor pressure deficit in the air

Γ_star (Array of) CO₂ compensation point with the absence of dark respiration

beta (Array of) Correction factor over the g1 part of an empirical model

Stomatal dynamics

Land.Plant.update_tree_with_time! — Function

update_tree_with_time!(tree::Tree,
                       Δt::FT,
                       scheme::AbstractOptimizationStomatalModel;
                       updating::Bool=false)
update_tree_with_time!(tree::Tree,
                       Δt::FT,
                       scheme::ESMGentine;
                       updating::Bool=false)
update_tree_with_time!(tree::Tree,
                       Δt::FT,
                       scheme::AbstractEmpiricalStomatalModel;
                       updating::Bool=false)

Update the tree flux information and hydraulic parameters, given

tree One Tree type struct

Δt Non-steady state $g_\text{sw}$ time in [s]

scheme An AbstractOptimizationStomatalModel type optimization stomatal model scheme

updating Hydraulic update option. If true, plant hydraulic parameters will be updated for the Tree

The function updates the non-steady state stomatal conductance by

\[\Delta g_\text{sw} = \text{factor} \cdot \left( \dfrac{∂A}{∂E} - \dfrac{∂Θ}{∂E} \right) \cdot Δt\]

for optimization stomatal models. The $\dfrac{∂A}{∂E}$ is the same for every stomatal control model, but $\dfrac{∂Θ}{∂E}$ differs among models.

The function updates the stomatal conductance via a Δgsw = factor * (gs_ss - gs_nss) for empirical stomatal models like Ball-Berry, Gentine, Leuning, and Medlyn models. The gs_mod stands for the gs calculated from model, and gs_nss stands for gs at non-steady state.