PHOTOSYNTHESIS MODELS MODULE

Temperature Dependency

Jmax

Land.PhotosynthesisModels.get_jmax — Function

get_jmax(para_set::AbstractJmaxTD, j25::FT, t_leaf::FT)

Maximal electron transport rate at leaf temperature, given

para_set AbstractJmax parameter set

j25 Maximal electron transport rate at 298.15 K

t_leaf Leaf temperature

Land.PhotosynthesisModels.get_j — Function

get_j(jmax::FT, par::FT, curvature::FT, qy::FT)

Electron transport rate j, given

jmax Maximal eclectron transport @ leaf temperature (not 298.15 K)

par Absorbed PAR (photosynthetic active radiation)

curvature Curvature parameter (default at 0.9)

qy Quantum yield of electron, qy = maxPSII * PSII_frac, maxPSII is maximal PSII yield (default at 4.0/4.9), PSII_frac is Fraction of absorbed light used by PSII ETR (default at 0.5)

Land.PhotosynthesisModels.JmaxTDBernacchi — Type

JmaxTDBernacchi{FT} <:AbstractJmaxTD

A non-mutable AbstractJmaxTD type JmaxTDBernacchi that stores information for Jmax temperature correction. The equation used for temperature correction is Arrhenius equation (See Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from "".

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Land.PhotosynthesisModels.JmaxTDCLM — Type

JmaxTDCLM{FT} <:AbstractJmaxTD

A non-mutable AbstractJmaxTD type JmaxTDCLM that stores information for Jmax temperature correction. The equation used for temperature correction is Arrhenius equation (See Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from "".

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Land.PhotosynthesisModels.JmaxTDLeuning — Type

JmaxTDLeuning{FT} <:AbstractJmaxTD

A non-mutable AbstractJmaxTD type JmaxTDLeuning that stores information for Jmax temperature correction. The equation used for temperature correction is Arrhenius equation (See Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from Leuning (2002).

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Kc

Land.PhotosynthesisModels.get_kc — Function

get_kc(para_set::AbstractKc, t_leaf::FT)

Kc at leaf temperature, given

para_set One AbstractKc type that store temperature correction information

t_leaf Leaf temperature

Land.PhotosynthesisModels.KcTDBernacchi — Type

KcTDBernacchi{FT} <: AbstractKcTD

A non-mutable AbstractKcTD type KcTDBernacchi that stores information for Kc temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from Bernacchi et al. (2001) "Improved temperature response functions for models of Rubisco‐limited photosynthesis".

Fields

Kc

Kc at 298.15 K (404.9 ppm) [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.KcTDCLM — Type

KcTDCLM{FT} <: AbstractKcTD

A non-mutable AbstractKcTD type KcTDCLM that stores information for Kc temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from "".

Fields

Kc

Kc at 298.15 K [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Ko

Land.PhotosynthesisModels.get_ko — Function

get_ko(para_set::AbstractKo, t_leaf::FT)

Ko at leaf temperature, given

para_set One AbstractKo type that store temperature correction information

t_leaf Leaf temperature

Land.PhotosynthesisModels.KoTDBernacchi — Type

KoBernacchi{FT} <: AbstractKoTD

A non-mutable AbstractKoTD type KoTDBernacchi that stores information for Ko temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from Bernacchi et al. (2001) "Improved temperature response functions for models of Rubisco‐limited photosynthesis".

Fields

Ko

Ko at 298.15 K (278.4 ‰) [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.KoTDCLM — Type

KoTDCLM{FT} <: AbstractKoTD

A non-mutable AbstractKoTD type KoTDCLM that stores information for Ko temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from "".

Fields

Ko

Ko at 298.15 K [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Kpep

Land.PhotosynthesisModels.get_kpep — Function

get_kpep(para_set::AbstractKpep, t_leaf::FT)

Kpep at leaf temperature, given

para_set One AbstractKpep type that store temperature correction information

t_leaf Leaf temperature

Land.PhotosynthesisModels.KpepTDBoyd — Type

KpepTDBoyd{FT} <: AbstractKpepTD

A non-mutable AbstractKpepTD type KpepTDBoyd that stores information for Kpep temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from Boyd et al. (2001) "Temperature responses of C4 photosynthesis: biochemical analysis of Rubisco, phosphoenolpyruvate carboxylase, and carbonic anhydrase in Setaria viridis."

Fields

Kpep

Kpep at 298.15 K [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.KpepTDCLM — Type

KpepTDCLM{FT} <: AbstractKpepTD

A non-mutable AbstractKpepTD type KpepTDCLM that stores information for Kpep temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from ""

Fields

Kpep

Kpep at 298.15 K [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Respiration

Land.PhotosynthesisModels.get_r — Function

get_r(para_set::AbstractRespirationTD, r25::FT, t_leaf::FT)

Leaf respiration rate r_leaf at leaf temperature, given

para_set One AbstractRespirationTD type that store temperature correction information

r25 Leaf respiration rate at 298.15 K (25 Celcius)

t_leaf Leaf temperature

Land.PhotosynthesisModels.RespirationTDBernacchi — Type

RespirationTDBernacchi{FT} <: AbstractRespirationTD

A non-mutable AbstractRespirationTD type RespirationTDBernacchi that stores information for Respiration temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from Bernacchi et al. (2001) "Improved temperature response functions for models of Rubisco‐limited photosynthesis".

Fields

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.RespirationTDCLM — Type

RespirationTDCLM{FT} <: AbstractRespirationTD

A non-mutable AbstractRespirationTD type RespirationTDCLM that stores information for Respiration temperature correction. The equation used for temperature correction is Arrhenius equation (see Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from "".

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Vmax

Land.PhotosynthesisModels.get_vmax — Function

get_vmax(v25::FT, para_set::AbstractVmaxTD, t_leaf::FT)

Maximal electron transport rate at leaf temperature, given

para_set AbstractVmaxTD parameter set

v25 Maximal carboxylation rate at 298.15 K

t_leaf Leaf temperature

Land.PhotosynthesisModels.VcmaxTDBernacchi — Type

VcmaxTDBernacchi{FT} <: AbstractVcmaxTD

A non-mutable AbstractVcmaxTD type VcmaxTDBernacchi that stores information for Vcmax temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from Bernacchi et al. (2001) "Improved temperature response functions for models of Rubisco‐limited photosynthesis".

Fields

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.VcmaxTDCLM — Type

VcmaxTDCLM{FT} <:AbstractVcmaxTD

A non-mutable AbstractVcmaxTD type VcmaxTDCLM that stores information for Vcmax temperature correction. The equation used for temperature correction is Arrhenius equation (see Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from "".

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Land.PhotosynthesisModels.VcmaxTDLeuning — Type

VcmaxTDLeuning{FT} <:AbstractVcmaxTD

A non-mutable AbstractVcmaxTD type VcmaxTDLeuning that stores information for Vcmax temperature correction. The equation used for temperature correction is Arrhenius equation (see Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from leuning (2002).

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Land.PhotosynthesisModels.VomaxTDBernacchi — Type

VomaxTDBernacchi{FT} <: AbstractVomaxTD

A non-mutable AbstractVomaxTD type VomaxTDBernacchi that stores information for Vomax temperature correction. The equation used for temperature correction is correction = exp( c - ΔHa/(R*T_leaf) ). The data source for the constants can be referred from Bernacchi et al. (2001) "Improved temperature response functions for models of Rubisco‐limited photosynthesis".

Fields

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.VpmaxTDBoyd — Type

VpmaxTDBoyd{FT} <: AbstractVpmaxTD

A non-mutable AbstractVpmaxTD type VpmaxTDBoyd that stores information for Vomax temperature correction. The equation used for temperature correction is Arrhenius peak correction (see Leuning (2002) "Leuning, R. "Temperature dependence of two parameters in a photosynthesis model"). The data source for the constants can be referred from Boyd et al. (2001) "Temperature responses of C4 photosynthesis: biochemical analysis of Rubisco, phosphoenolpyruvate carboxylase, and carbonic anhydrase in Setaria viridis".

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Γ*

Land.PhotosynthesisModels.get_Γ_star — Function

get_Γ_star(paraset::AbstractΓStarTD, t_leaf::FT)

Γ_star at leaf temperature, given

para_set One AbstractΓStarTD type that store temperature correction information

t_leaf Leaf temperature

Land.PhotosynthesisModels.ΓStarTDBernacchi — Type

ΓStarTDBernacchi{FT} <: AbstractΓStarTD

A non-mutable AbstractΓStarTD type ΓStarTDBernacchi that stores information for Γstar temperature correction. The equation used for temperature correction is `correction = exp( c - ΔHa/(R*Tleaf) )`. The data source for the constants can be referred from Bernacchi et al. (2001) "Improved temperature response functions for models of Rubisco‐limited photosynthesis".

Fields

Γ_star

Γ_star at 298.15 K (42.75 ppm) [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Land.PhotosynthesisModels.ΓStarTDCLM — Type

ΓStarTDCLM{FT} <: AbstractΓStarTD

A non-mutable AbstractΓStarTD type ΓStarTDCLM that stores information for Γstar temperature correction. The equation used for temperature correction is `correction = exp( c - ΔHa/(R*Tleaf) )`. The data source for the constants can be referred "".

Fields

Γ_star

Γ_star at 298.15 K [Pa]

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K-25

Arrhenius Correction

Land.PhotosynthesisModels.arrhenius_correction — Function

arrhenius_correction(para_set::KcTDBernacchi{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A KcTDBernacchi type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::KcTDCLM{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A KcTDCLM type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::KoTDBernacchi{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A KoTDBernacchi type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::KoTDCLM{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A KoTDCLM type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::KpepTDBoyd{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A KpepTDBoyd type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::KpepTDCLM{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A KpepTDCLM type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::RespirationTDBernacchi{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A RespirationTDBernacchi type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::VcmaxTDBernacchi{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A VcmaxTDBernacchi type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::VomaxTDBernacchi{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A VomaxTDBernacchi type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::ΓStarTDBernacchi{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A ΓStarTDBernacchi type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::ΓStarTDCLM{FT}, t_leaf::FT)

A correction factor based on arrhenius's fitting procedure, given

para_set A ΓStarTDCLM type parameter set

t_leaf Leaf temperature in [K]

The equation used is correction = exp( c - ΔHa/(R*T_leaf) ).

arrhenius_correction(para_set::JmaxTDBernacchi{FT}, t_leaf::FT)

Arrhenius correction over Jmax, given

para_set A JmaxTDBernacchi type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

arrhenius_correction(para_set::JmaxTDCLM{FT}, t_leaf::FT)

Arrhenius correction over Jmax, given

para_set A JmaxTDCLM type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

arrhenius_correction(para_set::JmaxTDLeuning{FT}, t_leaf::FT)

Arrhenius correction over Jmax, given

para_set A JmaxTDLeuning type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

arrhenius_correction(para_set::RespirationTDCLM{FT}, t_leaf::FT)

Arrhenius correction over Respiration, given

para_set A RespirationTDCLM type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

arrhenius_correction(para_set::VcmaxTDCLM{FT}, t_leaf::FT)

Arrhenius correction over Vcmax, given

para_set A VcmaxTDCLM type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

arrhenius_correction(para_set::VcmaxTDLeuning{FT}, t_leaf::FT)

Arrhenius correction over Vcmax, given

para_set A VcmaxTDLeuning type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

arrhenius_correction(para_set::VpmaxTDBoyd{FT}, t_leaf::FT)

Arrhenius correction over Vpmax, given

para_set A VpmaxTDBoyd type parameter set

t_leaf Leaf temperature

The equations used are f_above = C * exp [ ( Ha /RT_0 ) * ( 1 - T_0/T_1 ) ], f_below = 1 + exp [ ( Sv*T_1 - Hd ) / ( RT_1) ], C = 1 + exp[(Sv*T_0 - Hd )/(RT_0 )], and correction = f_above /f_below

Parameter Sets

Respiration~Vcmax

Land.PhotosynthesisModels.VtoRCollatz — Type

VtoRCollatz{FT} <: AbstractVtoR

A non-mutable AbstractVtoR type VtoRCollatz that stores information for Respiration Vcmax25 correction. The equation used is r25 = v25 * correction. The data source for the constants can be referred from Collatz et al. (1991) "Physiological and environmental regulation of stomatal conductance, photosynthesis and transpiration: a model that includes a laminar boundary layer."

Fields

v_to_r

Ratio between r25 and v25

Land.PhotosynthesisModels.VtoRDefault — Type

VtoRDefault{FT} <: AbstractVtoR

A non-mutable AbstractVtoR type VtoRDefault that stores information for Respiration Vcmax25 correction. The equation used is r25 = v25 * correction. Correction default at 0.015

Fields

v_to_r

Ratio between r25 and v25

Preset Photosynthesis Types

Land.PhotosynthesisModels.C3VcJBernacchi — Type

C3VcJBernacchi{FT} <: C3ParaSetVcJ

A C3ParaSetVcJ type of parameter sets for different temperature dependencies.

Fields

JT

Jmax temperature dependency

KcT

Kc temperature dependency

KoT

Ko temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

ΓsT

Γ_star temperature dependency

VR

Vcmax25 and respiration correlation

Land.PhotosynthesisModels.C3VcVpJBernacchi — Type

C3VcVpJBernacchi{FT} <: C3ParaSetVcVpJ

A C3ParaSetVcVpJ type of parameter sets for different temperature dependencies.

Fields

JT

Jmax temperature dependency

KcT

Kc temperature dependency

KoT

Ko temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

ΓsT

Γ_star temperature dependency

VR

Vcmax25 and respiration correlation

Land.PhotosynthesisModels.C4VcVpJBoyd — Type

C4VcVpJBoyd{FT} <: C4ParaSetVcVpJ

A C4ParaSetVcVpJ type of parameter sets for different temperature dependencies.

Fields

KpT

Kpep temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

VpT

Vpmax temperature dependency

VR

Vcmax25 and respiration correlation

Land.PhotosynthesisModels.C4VcVpJCLM — Type

C4VcVpJCLM{FT} <: C4ParaSetVcVpJ

A C4ParaSetVcVpJ type of parameter sets for different temperature dependencies.

Fields

KpT

Kpep temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

VpT

Vpmax temperature dependency

VR

Vcmax25 and respiration correlation

Procedure

Calculate Photosynthesis from CO2

Land.PhotosynthesisModels.get_an_ag_r_from_pi — Function

get_an_ag_r_from_pi(model::C3ParaSetVcJ, p_i::FT, v25::FT, j25::FT, par::FT, p_O₂::FT, t_leaf::FT, r25::FT, curvature::FT, qy::FT)

Gross photosynthetic rate limited by carboxylation without ATP limitation, given

model A C3ParaSetVcJ type parameter sets that stores temperature denpendencies

p_i Leaf internal CO₂ partial pressure

v25 Maximal carboxylation rate at 298.15 K

j25 Maximal electron transport rate

par Absorbed photosynthetic active radiation

p_O₂ O₂ partial pressure

t_leaf Leaf temperature

r25 Respiration rate at 298.15 K

curvature Curvature parameter to calculate J

qy Quantum yield of electron

The equations used are from Farquhar et al. (1980) "A biochemical model of photosynthetic CO2 assimilation in leaves of C3 species."

get_an_ag_r_from_pi(model::C3ParaSetVcVpJ, p_i::FT, v25::FT, j25::FT, par::FT, p_O₂::FT, t_leaf::FT, r25::FT, curvature::FT, qy::FT)

Gross photosynthetic rate limited by carboxylation with ATP limitation, given

model A C3ParaSetVcVpJ type parameter sets that stores temperature denpendencies

p_i Leaf internal CO₂ partial pressure

v25 Maximal carboxylation rate at 298.15 K

j25 Maximal electron transport rate

par Absorbed photosynthetic active radiation

p_O₂ O₂ partial pressure

t_leaf Leaf temperature

r25 Respiration rate at 298.15 K

curvature Curvature parameter to calculate J

qy Quantum yield of electron

The equations used are from Farquhar et al. (1980) "A biochemical model of photosynthetic CO2 assimilation in leaves of C3 species."

get_an_ag_r_from_pi(model::C4ParaSetVcVpJ, p_i::FT, v25::FT, p25::FT, par::FT, t_leaf::FT, r25::FT, qy::FT)

Gross photosynthetic rate limited by carboxylation, given

model A C4ParaSetVcVpJ type parameter sets that stores temperature denpendencies

p_i Leaf internal CO₂ partial pressure

v25 Maximal carboxylation rate at 298.15 K

p25 Maximal PEP carboxylation rate at 298.15 K

par Absorbed photosynthetic active radiation

t_leaf Leaf temperature

r25 Respiration rate at 298.15 K

qy Quantum yield of electron

The model is adapted from Collatz et al. (1992) "Coupled photosynthesis-stomatal conductance model for leaves of C4 plants."

Calculate Photosynthesis from gsc

Land.PhotosynthesisModels.get_an_ag_r_pi_from_gsc — Function

get_an_ag_r_pi_from_gsc(model::C3ParaSet, gsc::FT, v25::FT, j25::FT, p25::FT, p_a::FT, t_leaf::FT, par::FT, p_atm::FT, p_O₂::FT, r25::FT, curvature::FT, qy::FT)

Net photosynthetic rate tar_an, gross photosynthetic rate tar_ag, respiration rate tar_r, and leaf internal CO₂ partial pressure tar_p, given

model A C3ParaSet type parameter set

gsc Leaf diffusive conductance to CO₂

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plant only

p_a Atmospheric CO₂ partial pressure

t_leaf Leaf temperature

par Photosynthetic active radiation

p_atm Atmospheric pressure

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

curvature Curvature parameter to calculate J

qy Quantum yield of electron

For C3 plants, w/wo ATP limitation. Some parameters are useless in C3 photosynthesis, like p25.

get_an_ag_r_pi_from_gsc(model::C4ParaSet, gsc::FT, v25::FT, j25::FT, p25::FT, p_a::FT, t_leaf::FT, par::FT, p_atm::FT, p_O₂::FT, r25::FT, curvature::FT, qy::FT)

Net photosynthetic rate tar_an, gross photosynthetic rate tar_ag, respiration rate tar_r, and leaf internal CO₂ partial pressure tar_p, given

model A C4ParaSet type parameter set

gsc Leaf diffusive conductance to CO₂

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plant only

p_a Atmospheric CO₂ partial pressure

t_leaf Leaf temperature

par Photosynthetic active radiation

p_atm Atmospheric pressure

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

curvature Curvature parameter to calculate J

qy Quantum yield of electron

For C4 plants. Some parameters are useless for C4 photosynthesis, like j25, p_O₂, and curvature.

Land.PhotosynthesisModels.get_an_ag_r_pi_from_gsc_list — Function

get_an_ag_r_pi_from_gsc_list(model::C3ParaSet, gsc_list::Array{FT,1}, v25::FT, j25::FT, p25::FT, p_a::FT, t_leaf_list::Array{FT,1}, par_list::Array{FT,1}, p_atm::FT, p_O₂::FT, r25::FT, curvature::FT, qy::FT)

List of net photosynthetic rate list_an, gross photosynthetic rate list_ag, respiration rate list_re, and leaf internal CO₂ partial pressure list_pi, given

model A C3ParaSet type parameter set

gsc_list List of leaf diffusive conductance to CO₂

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plants only

p_a Atmospheric CO₂ partial pressure

t_leaf_list List of leaf temperature

par_list List of photosynthetic active radiation

p_atm Atmospheric pressure

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

curvature Curvature parameter to calculate J

qy Quantum yield of electron

Some parameters are useless here for C3 photosynthesis model (e.g., p25), but they are given for more convenient modeling.

get_an_ag_r_pi_from_gsc_list(model::C4ParaSet, gsc_list::Array{FT,1}, v25::FT, j25::FT, p25::FT, p_a::FT, t_leaf_list::Array{FT,1}, par_list::Array{FT,1}, p_atm::FT, p_O₂::FT, r25::FT, curvature::FT, qy::FT)

List of net photosynthetic rate list_an, gross photosynthetic rate list_ag, respiration rate list_re, and leaf internal CO₂ partial pressure list_pi, given

model A C4ParaSet type parameter set

gsc_list List of leaf diffusive conductance to CO₂

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plants only

p_a Atmospheric CO₂ partial pressure

t_leaf_list List of leaf temperature

par_list List of photosynthetic active radiation

p_atm Atmospheric pressure

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

curvature Curvature parameter to calculate J

qy Quantum yield of electron

Some parameters are useless here for C4 photosynthesis model (e.g., p_O₂), but they are given for more convenient modeling.