PLANT MODULE

Plant Structure

Root

Land.Plant.RootLayer — Type

RootLayer{FT<:AbstractFloat}

A RootLayer type which contains root hydraulics information, including rhizosphere conductance.

Fields

z_hi

Upper root depth [m]

z_lo

Lower root depth [m]

r_layer

Average horizental root length [m]

f_layer

Fraction of biomass and k_max

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance, default at 2700.0 Kg h⁻¹ MPa⁻¹ m⁻² basal area [mol s⁻¹ MPa⁻¹]

k_s

Maximal hydraulic conductivity per cross section basal area per root depth [mol s⁻¹ MPa⁻¹ m⁻²]

k_rhiz

Maximal rhizosphere conductance, default at 9.72e12 Kg h⁻¹ MPa⁻¹ mm⁻² basal area [mol s⁻¹ MPa⁻¹]

p_ups

Soil water potential (upstream) [MPa]

soil_a

Soil texture parameter

soil_m

Soil texture parameter, 1 - 1/n

soil_n

Soil texture parameter

soil_msc

Maximal soil volumatric water content, vary with rock fraction

soil_rwc

Relative soil volumatric water content

soil_vwc

Soil volumatric water content

t_soil

Soil temperature [K]

soil_type

Soil texture

p_rhiz

Water potential at the root-rhizosphere interface [MPa]

p_dos

Xylem water pressure at the tree base (downstream) [MPa]

q

Flow rate in the root layer [mol s⁻¹]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹]

p_element

List of xylem water pressure per element [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of water temperature per element [K]

z_element

List of height per element to account for gravity [m]

Land.Plant.create_root_list — Function

create_root_list(FT, n, z_lo)

The function creates a root system with n root layers evenly in the soil. This function need to work with another function (pending) to quantify the root distribution.

FT Floating type for the RootLayer struct

n Total number of root layers in the root system

z_lo Maximal root depth

Trunk and Branch

Land.Plant.Stem — Type

Stem{FT<:AbstractFloat}

A Stem type which contains hydraulic information, including drought history profile. The Stem type is used for creating a trunk and a list of branches.

Fields

r_layer

Horizental radius of the layer [m]

z_hi

Upper height from the ground (trunk-branch joint for trunk, stem-leaf joint for branch) [m]

z_lo

Lower height from the ground (0 for trunk, trunk-branch joint for branch) [m]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance, = 5400.0 * 8/3 Kg h⁻¹ MPa⁻¹ m⁻² basal area [mol s⁻¹ MPa⁻¹]

k_s

Maximal hydraulic conductivity per cross section area per tree height [mol s⁻¹ MPa⁻¹ m⁻²]

p_ups

Xylem water pressure at the tree basa (upstream) [MPa]

p_dos

Xylem water pressure at the trunk-stem or stem-leaf joint (downstream) [MPa]

q

Flow rate in the xylem [mol s⁻¹]

k_element

List of stem k_max per element [mol s⁻¹ MPa⁻¹]

p_element

List of stem xylem pressure per element [MPa]

p_history

List of stem xylem pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of stem temperature for each element [K]

z_element

List of stem height per element to account for gravity [m]

Land.Plant.create_branch_list — Function

create_branch_list(FT, n, z_lo, z_hi)

This function creates a list of stem structs as branch system for the tree, given

FT Floating type for the Stem struct

n Total number of canopy layers, one branch for a canopy layer

z_lo Minimal canopy distance from the ground

z_hi Maximal canopy distance from the ground

Each branch (Stem) is responsible for a canopy layer. The branch index from 1:n is from lower to higher canopy.

Canopy and Leaf

Land.Plant.Leaf — Type

Leaf{FT<:AbstractFloat}

A Leaf type which contains leaf hydraulics information.

Fields

angle_locat

direction from the tree trunk to the leaf, 0 for east, 270 for south, and 180 for west [°]

angle_incli

direction of leave, 0 for flat and 90 for vertical [°]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_sla

Maximal leaf hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

p_ups

Leaf xylem water pressure (different from leaf water potential) at the leaf base (upstream) [MPa]

p_dos

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

k_element

List of leaf k_max per element (mol s⁻¹ MPa⁻¹)

p_element

List of xylem water pressure per element [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of xylem water temperature per element [K]

z_element

List of leaf element height change to account for gravity, default = 0 [m]

Land.Plant.CanopyLayer — Type

CanopyLayer{FT<:AbstractFloat, n_leaf}

A CanopyLayer type, which constains environemntal conditions and leaf-level fluxes for n_leaf Leaf. The n_leaf is the sum of n_Ari * n_Incli sunlit leaves and 1 shaded leaves.

Fields

f_layer

Leaf area fraction in the canopy

f_view

Fraction of the sunlit leaves

la

Leaf area in the layer [m²]

width

Leaf width [m]

p_a

Atmospheric CO₂ partial pressure in the layer [Pa]

p_atm

Atmospheric pressure [Pa]

p_O₂

Atmospheric O₂ partial presssure [Pa]

p_H₂O

Atmospheric H₂O partial pressure [Pa]

t_air

Air temperature [K]

wind

Wind speed [m s⁻¹]

g_ias_c

Mesophyll conductance correction factor: multiplier

g_ias_e

Mesophyll conductance correction factor: exponent

g_max

Maximal leaf diffusive conductance for H₂O at 298.15 K [mol m⁻² s⁻¹]

gs_nssf

Non-steady state factor for optimization approach (use it by multiplying the ∂A/∂E - ∂Θ/∂E)

gs_empi

Non-steady state factor for empirical apporach (used by multiplying the gswmod - gswcurr)

j_max

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

r_25

Leaf respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_max

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

p_max

Maximal PEP carboxylation rate at 298.15 K (only for C4 plants) [μmol m⁻² s⁻¹]

curvature

Curvature ratio for J

qy

Quntum yield for electron, using aPAR

leaf_list

List of Leaf

ag_list

List of gross photosynthetic rate for n_leaf leaves [μmol m⁻² s⁻¹]

an_list

List of net photosynthetic rate [μmol m⁻² s⁻¹]

am_list

List if maximal A at the given scenario [μmol m⁻² s⁻¹]

d_list

List of leaf-to-air vapor pressure deficit [unitless]

e_list

List of flow rate per leaf area [mol m⁻² s⁻¹]

ec_list

List of critical flow rate (when leaf xylem pressure induces desiccation) per leaf area [mol m⁻² s⁻¹]

gsc_list

List of effective leaf diffusive conductance for CO₂ [mol m⁻² s⁻¹]

gsw_empi

List of empirical leaf diffusive conductance for H₂O [mol m⁻² s⁻¹]

gsw_list

List of leaf diffusive conductance for H₂O [mol m⁻² s⁻¹]

km_list

List of maximal hydraulic conductance [mol m⁻² s⁻¹]

la_list

List of leaf area per leaf [m²]

par_list

List of PAR for each leaf

pi_list

List of leaf internal CO₂ partial pressure [Pa]

q_list

List of flow rate [mol s⁻¹]

r_list

List of leaf respiration rate [μmol m⁻² s⁻¹]

t_list

List of leaf temperature [K]

Tree

Land.Plant.Tree — Type

Tree{FT<:AbstractFloat}

A Tree type which includes

root_no number of root layers

a trunk

canopy_no branches

leaf_no canopy layers

Fields

age

Age [year]

ba

Basal area [m²]

ga

Ground area [m²]

h

Tree height [m]

photo_type

Photosynthesis type C3/C4/CAM

photo_para_set

Photosynthesis model parameter set

stomata_scheme

Optimization model option

n_root

Number of root layers

root_frac

List of root fraction

root_zs

List of root z lower boundary

root_list

List of RootLayer

n_canopy

Number of canopy layers

trunk

Trunk using type Stem

branch_list

List of Stem

n_leaf

Number of leaves in each canopy layer

canopy_list

List of CanopyLayer

Plant Hydraulics

Land.Plant.get_struct_p_end_from_q — Function

get_struct_from_q(root_layer, flow; p_ini)

End pressure p_end (in MPa) for RootLayer, including impacts from rhizosphere conductance and gravity, given

root_layer One RootLayer in Tree

flow Flow rate (in mol s⁻¹) in the given RootLayer

p_ini Upstream soil water potential if p_ini is given, otherwise (Inf) root_layer.p_ups will be used

get_struct_from_q(stem, flow; p_ini)

End pressure p_end (in MPa) for stem Stem, including impact from gravity, given

stem Stem as Trunk in Tree or Stem as side branch in Tree

flow Flow rate (in mol s⁻¹) in the given Stem

p_ini Upstream xylem pressure if p_ini is given, otherwise (Inf) stem.p_ups will be used

get_struct_from_q(leaf, flow; p_ini)

End pressure p_end (in MPa) for leaf leaf, including impact from gravity, given

leaf Leaf in CanopyLayer in Tree

flow Flow rate per leaf area (mol m⁻² s⁻¹) in the given Leaf

p_ini Upstream xylem pressure if p_ini is given, otherwise (Inf) leaf.p_ups will be used

Land.Plant.get_q_layer_from_p_base — Function

get_q_layer_from_p_base(root_layer, p_base)

Flow rate in a root layer q_layer, given

root_layer One RootLayer in Tree

p_base Xylem water pressure at the tree base

This function use Newton Raphson combined with Bi-section to get q_layer. A warning message will display if total iterations exceed 50 times.

Land.Plant.get_p_base_q_list_from_q — Function

get_p_base_q_list_from_q(tree, flow)

Tree base pressyre p_base and a list of flow rate in each root layer q_list, given

tree A Tree type

flow Flow rate for the whole tree [mol s⁻¹]

This function returns the p_base from total transpiration rate using Newton Raphson. A warning will display if total iterations is beyond 50 times.

Land.Plant.update_struct_from_q! — Function

update_struct_from_q!(root_layer, flow; p_ini)

Update root hydraulic information considering the impact from rhizosphere conductance and gravity, given

root_layer One RootLayer in Tree

flow Flow rate in the given root layer

p_ini Upstream soil water potential if p_ini is given, otherwise (Inf) root_layer.p_ups will be used

update_struct_from_q!(stem, flow; p_ini)

Update the stem hydraulic information considering the impact from gravity, given

stem Stem as trunk in a Tree or as side branch in Tree

flow::FT Flow rate in the given stem

p_ini::FT Upstream xylem pressure if p_ini is given, otherwise (Inf) stem.p_ups will be used

update_struct_from_q!(leaf, flow; p_ini)

Update the leaf hydraulic information considering the impact from gravity, given

leaf Leaf type in CanopyLayer in a Tree

flow Flow rate per leaf area in the given leaf

p_ini Upstream xylem pressure if p_ini is given, otherwise (Inf) leaf.p_ups will be used

Land.Plant.update_tree_e_crit! — Function

update_tree_e_crit!(tree; displaying)

Update the critical transpiration rate information, given

tree A Tree type

displaying If true, messages of how e_crit evolves step-by-step will display

This function update e_crit for each layer (sunlit and shaded). The e_crit for each leaf is either increased or decreased so that p_leaf is in the range -Inf to -20.0 MPa. The e_crit increases by 1e-6 mol m⁻² s⁻¹ if p_leaf is less negative than -20 MPa. The e_crit decreases by 1e-7 mol m⁻² s⁻¹ if p_leaf is -Inf.

Stomatal Control Options

Land.Plant.ESMBallBerry — Type

ESMBallBerry{FT}

An empirical model parameter set type for Ball-Berry type model. The equation used for Ball-Berry type model is gs = g0 + g1 * RH * A/Ci. Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [unitless]

Land.Plant.ESMGentine — Type

ESMGentine{FT}

An empirical model parameter set type for Gentine type model. The equation used for Gentine type model is gs = g0 + g1 * k_leaf/k_max * A/Ci. Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required. Note it that the Gentine model does not require for a β function to tune the soil drought response, but the use of k_leaf also does not permit post-drought stomatal response unless k_leaf can be recovered.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [unitless]

Land.Plant.ESMLeuning — Type

ESMLeuning{FT}

An empirical model parameter set type for Leuning type model. The equation used for Leuning type model is gs = g0 + g1 * A/(Ci-Γ_star) * 1/(1+VPD/d0). Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [unitless]

d0

fitting parameter of d/d0 below the fraction, same unit as vpd [Pa]

Land.Plant.ESMMedlyn — Type

ESMMedlyn{FT}

An empirical model parameter set type for Medlyn type model. The equation used in Medlyn type model is gs = g0 + (1+g1/sqrt(VPD)) * A/Ci. Note it that these empirical model often takes Ci or Cc as input, so a conversion from Pa to ppm is required.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹], should this one be temperature-dependent?

g1

slope of conductance-photosynthesis correlation [Pa⁽⁵⁾]

Land.Plant.OSMEller — Type

OSMEller

An optimization model parameter set type for Eller model. The equation used for Eller model is ∂Θ/∂E = -∂K/∂E * A/K, where K is ∂E/∂P.

Fields

Land.Plant.OSMSperry — Type

OSMSperry

An optimization model parameter set type for Sperry model. The equation used for Sperry model is ∂Θ/∂E = -∂K/∂E * Amax/Kmax, where K is ∂E/∂P.

Fields

Land.Plant.OSMWang — Type

OSMWang

An optimization model parameter set type for Eller type model. The equation used for Eller model is ∂Θ/∂E = A / (E_crit - E).

Fields

Land.Plant.OSMWAP — Type

OSMWAP{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model. The equation used for Wolf-Anderegg-Pacala model is ∂Θ/∂E = (2aP+b) / K, where K is ∂P/∂E.

Fields

a

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻²]

b

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻¹]

Land.Plant.OSMWAPMod — Type

OSMWAPMod{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model, modified by adding a photosynthesis component while set b and c = 0. The equation used for modified Wolf-Anderegg-Pacala model is ∂Θ/∂E = a * P, where P is absolute value of leaf xylem pressure.

Fields

a

Quadratic equation parameter [mol mol⁻¹ MPa⁻¹]

Optimization Stomatal Dynamics

Land.Plant.get_marginal_gain — Function

get_marginal_gain(canopyi, indx, photo_para_set)

Marginal water use efficiency ∂A/∂E for a leaf in a canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

get_marginal_gain(canopyi, indx)

A list of marginal water use efficiency for all the leaves, given

canopyi A CanopyLayer in a Tree

Land.Plant.update_leaf_ak_max! — Function

update_leaf_ak_max!(tree::Tree)

Update leaf maximal A_net allowed at the given environmental conditions, given -tree A Tree type struct

This function is meant to work with Sperry et al. (2017) model, which requires a_max in the ∂Θ∂E function.

Land.Plant.get_marginal_penalty — Function

get_marginal_penalty(canopyi, indx, scheme::OSMEller)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

scheme optimizaiton stomatal model scheme OSMEller

This function uses the Eller et al. (2018) model. ∂Θ/∂E = A / K * dK/dE, unit [mol mol⁻¹]. To make the Eller model suitable for multiple layer canopy, K at the leaf xylem end is used. Also, to make the model capable of handling drought history, loss of conductance is evaluated assuming no drought legacy from a more negative p_leaf.

get_marginal_penalty(canopyi, scheme::OSMEller)

List of marginal water penalty for all the leaves, given

canopyi A CanopyLayer in a Tree

scheme optimizaiton stomatal model scheme OSMEller

This function uses the Eller et al. (2018) model. ∂Θ/∂E = A / K * dK/dE, unit [mol mol⁻¹]. To make the Eller model suitable for multiple layer canopy, K at the leaf xylem end is used. Also, to make the model capable of handling drought history, loss of conductance is evaluated assuming no drought legacy from a more negative p_leaf.

get_marginal_penalty(canopyi, indx, scheme::OSMSperry)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

scheme optimizaiton stomatal model scheme OSMSperry

This function uses the Sperry et al. (2017) model. ∂Θ/∂E = A_max / K_max * dK/dE, unit [mol mol⁻¹]. To make the Sperry model suitable for multiple layer canopy, K at the leaf xylem end is used. Also, to make the model capable of handling drought history, loss of conductance is evaluated assuming no drought legacy from a more negative p_leaf.

get_marginal_penalty(canopyi, scheme::OSMSperry)

List of marginal water penalty for all the leaves, given

canopyi A CanopyLayer in a Tree

scheme optimizaiton stomatal model scheme OSMSperry

This function uses the Sperry et al. (2017) model. ∂Θ/∂E = A_max / K_max * dK/dE, unit [mol mol⁻¹]. To make the Sperry model suitable for multiple layer canopy, K at the leaf xylem end is used. Also, to make the model capable of handling drought history, loss of conductance is evaluated assuming no drought legacy from a more negative p_leaf.

get_marginal_penalty(canopyi, indx, scheme::OSMWang)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

scheme optimizaiton stomatal model scheme OSMWang

This function uses the Wang et al. (2020) model. ∂Θ/∂E = A / (e_crit - e_leaf), unit [mol mol⁻¹].

get_marginal_penalty(canopyi, scheme::OSMWang)

List of marginal water penalty for all the leaves, given

canopyi A CanopyLayer in a Tree

scheme optimizaiton stomatal model scheme OSMWang

This function uses the Wang et al. (2020) model. ∂Θ/∂E = A / (e_crit - e_leaf), unit [mol mol⁻¹].

get_marginal_penalty(canopyi, indx, scheme::OSMWAP)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

scheme optimizaiton stomatal model scheme OSMWAP

This function uses the Anderegg et al. (2018) model. ∂Θ/∂E = (2*a*P + b) / K, unit [mol mol⁻¹].

get_marginal_penalty(canopyi, scheme::OSMWAP)

List of marginal water penalty for all the leaves, given

canopyi A CanopyLayer in a Tree

scheme optimizaiton stomatal model scheme OSMWAP

This function uses the Anderegg et al. (2018) model. ∂Θ/∂E = (2*a*P + b*P) / K, unit [mol mol⁻¹].

get_marginal_penalty(canopyi, indx, scheme::OSMWAPMod)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in a Tree

indx indx'th Leaf in the canopyi.leaf_list

scheme optimizaiton stomatal model scheme OSMWAPMod

This function is modified from Anderegg et al. (2018) model. ∂Θ/∂E = a * A * P / K, unit [mol mol⁻¹].

get_marginal_penalty(canopyi, scheme::OSMWAPMod)

List of marginal water penalty for all the leaves, given

canopyi A CanopyLayer in a Tree

scheme optimizaiton stomatal model scheme OSMWAPMod

This function is modified from Anderegg et al. (2018) model. ∂Θ/∂E = a * A * P / K, unit [mol mol⁻¹].

Land.Plant.update_tree_with_time! — Function

update_tree_with_time!(tree::Tree, Δt::FT, scheme::AbstractOptimizationStomatalModel; updating::Bool=false)

Update the tree flux information and hydraulic parameters, given

tree One Tree type

Δt Δt of non-steady state gs

scheme An AbstractOptimizationStomatalModel type optimization stomatal model scheme

updating If true, plant hydraulic parameters will be updated for the Tree

The function updates the stomatal conductance via a Δgsw = factor * (∂A/∂E - ∂Θ/∂E) for optimization stomatal models. The ∂A/∂E is the same for every stomatal control model, but ∂Θ/∂E differs among models.

update_tree_with_time!(tree::Tree, Δt::FT, scheme::ESMGentine; updating::Bool=false)

Update the tree flux information and hydraulic parameters, given

tree One Tree type

Δt Δt of non-steady state gs

scheme An ESMGentine type empirical stomatal model scheme

updating If true, plant hydraulic parameters will be updated for the Tree

The function updates the stomatal conductance via a Δgsw = factor * (gs_ss - gs_nss) for empirical stomatal models like Ball-Berry, Leuning, and Medlyn models. The gs_ss stands for the gs at steady state, and gs_nss stands for gs at non-steady state.

update_tree_with_time!(tree::Tree, Δt::FT, scheme::AbstractEmpiricalStomatalModel; updating::Bool=false)

Update the tree flux information and hydraulic parameters, given

tree One Tree type

Δt Δt of non-steady state gs

scheme An AbstractEmpiricalStomatalModel type empirical stomatal model scheme

updating If true, plant hydraulic parameters will be updated for the Tree

The function updates the stomatal conductance via a Δgsw = factor * (gs_ss - gs_nss) for empirical stomatal models like Ball-Berry, Leuning, and Medlyn models. The gs_ss stands for the gs at steady state, and gs_nss stands for gs at non-steady state.

Empirical Stomatal Dynamics

Land.Plant.get_empirical_gsw_from_model — Function

get_empirical_gsw_from_model(model::ESMBallBerry, an::FT, p_atm::FT, p_i::FT, rh::FT, vpd::FT, Γ_star::FT)

Steady stage gsw from empirical approach given

model An ESMBallBerry empirical model type parameter set

an Net photosynthetic rate [μmol m⁻² s⁻¹]

p_atm Atmospheric pressure

p_i Leaf internal CO₂ partial pressure

rh Relative humidity

vpd Vapor pressure deficit in the air

Γ_star CO₂ compensation point with the absence of dark respiration

Some parameters are not used for the Ball-Berry type model.

get_empirical_gsw_from_model(model::ESMGentine, an::FT, k_ratio::FT, p_atm::FT, p_i::FT)

Steady stage gsw from empirical approach given

model An ESMGentine empirical model type parameter set

an Net photosynthetic rate [μmol m⁻² s⁻¹]

p_atm Atmospheric pressure

p_i Leaf internal CO₂ partial pressure

rh Relative humidity

vpd Vapor pressure deficit in the air

Γ_star CO₂ compensation point with the absence of dark respiration

Some parameters are not used for the Leuning type model.

get_empirical_gsw_from_model(model::ESMLeuning, an::FT, p_atm::FT, p_i::FT, rh::FT, vpd::FT, Γ_star::FT)

Steady stage gsw from empirical approach given

model An ESMLeuning empirical model type parameter set

an Net photosynthetic rate [μmol m⁻² s⁻¹]

p_atm Atmospheric pressure

p_i Leaf internal CO₂ partial pressure

rh Relative humidity

vpd Vapor pressure deficit in the air

Γ_star CO₂ compensation point with the absence of dark respiration

Some parameters are not used for the Leuning type model.

get_empirical_gsw_from_model(model::ESMMedlyn, an::FT, p_atm::FT, p_i::FT, rh::FT, vpd::FT, Γ_star::FT)

Steady stage gsw from empirical approach given

model An ESMMedlyn empirical model type parameter set

an Net photosynthetic rate [μmol m⁻² s⁻¹]

p_atm Atmospheric pressure

p_i Leaf internal CO₂ partial pressure

rh Relative humidity

vpd Vapor pressure deficit in the air

Γ_star CO₂ compensation point with the absence of dark respiration

Some parameters are not used for the Medlyn type model.

Land.Plant.get_empirical_gsw — Function

get_empirical_gsw(curvature::FT, g_ias_c::FT, g_ias_e::FT, g_max::FT, j25::FT, p25::FT, p_a::FT, p_atm::FT, p_H₂O::FT, p_O₂::FT, par::FT, qy::FT, r25::FT, t_leaf::FT, v25::FT, paraset::C3ParaSet, model::AbstractEmpiricalStomatalModel)

Steady state gsw from empirical approach, given

curvature Curvature parameter to calculate J

g_ias_c Mesophyll conductance correction multiplier

g_ias_e Mesophyll conductance correction exponent

g_max Maximal leaf diffusive conductance at 298.15 K

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plant only

p_a Atmospheric CO₂ partial pressure

p_atm Atmospheric pressure

p_H₂O Atmospheric H₂O partial pressure

p_O₂ O₂ partial pressure

par Photosynthetic active radiation

qy Quantum yield of electron

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

t_leaf Leaf temperature

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

paraset An C3ParaSet type parameter set

model An AbstractEmpiricalStomatalModel empirical model type

get_empirical_gsw(curvature::FT, g_ias_c::FT, g_ias_e::FT, g_max::FT, j25::FT, p25::FT, p_a::FT, p_atm::FT, p_H₂O::FT, p_O₂::FT, par::FT, qy::FT, r25::FT, t_leaf::FT, v25::FT, paraset::C3ParaSet, model::AbstractEmpiricalStomatalModel)

Steady state gsw from empirical approach, given

curvature Curvature parameter to calculate J

g_ias_c Mesophyll conductance correction multiplier

g_ias_e Mesophyll conductance correction exponent

g_max Maximal leaf diffusive conductance at 298.15 K

j25 Maximal electron transport rate at 298.15 K (25 Celcius)

p25 Maximal PEP carboxylation rate at 298.15 K, C4 plant only

p_a Atmospheric CO₂ partial pressure

p_atm Atmospheric pressure

p_H₂O Atmospheric H₂O partial pressure

p_O₂ O₂ partial pressure

par Photosynthetic active radiation

qy Quantum yield of electron

r25 Leaf respiration rate at 298.15 K (25 Celcius). If ==Inf, r25 will be computed from v25

t_leaf Leaf temperature

v25 Maixmal carboxylation rate at 298.15 K (25 Celcius)

paraset An C4ParaSet type parameter set

model An AbstractEmpiricalStomatalModel empirical model type

Missing docstring.

Missing docstring for Land.Plant.update_empirical_gsw_ss!. Check Documenter's build log for details.

Photosynthesis Tools

Land.Plant.get_a_par_curve — Function

get_a_par_curve(v25, j25, Γ_star, gsc, p_a, tem, p_O₂, r25)

Lists of phtosynthetic active radiation list_par, net phtosynthetic rate list_an, gross phtosynthetic rate list_ag, and leaf internal CO₂ partial pressure list_pi, given constant

v25 Maximal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K

Γ_star CO₂ compensation point with the absence of dark respiration

gsc Leaf diffusive conductance to CO₂

p_a Atmospheric CO₂ partial pressure

tem Leaf temperature

p_atm Atmospheric pressure

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K. If ==Inf, use v25 to compute r25.

Description

This function is meant to obtain the Anet-PAR curve. This function may also be used to test the phtosynthesis module.

Land.Plant.get_a_pi_curve — Function

get_a_pi_curve(v25, j25, Γ_star, tem, par, p_O₂, r25)

Lists of leaf internal CO₂ partial pressure list_pi, net photosynthetic rate list_an, and gross photosynthetic rate list_ag, given constant

v25 Maximal carboxylation rate at 298.15 K (25 Celcius)

j25 Maximal electron transport rate at 298.15 K

Γ_star CO₂ compensation point with the absence of dark respiration

tem Leaf temperature

par Photosynthetic active radiation

p_O₂ O₂ partial pressure

r25 Leaf respiration rate at 298.15 K. If ==Inf, use v25 to compute r25.

Interfaces

Land.Plant.initialize_rt_module — Function

initialize_rt_module(n_layer)

Initialize the RT module so as to interface with Plant, given

n_layer Number of canopy layers

LAI Leaf area index

Note it here that the structcanopy is used in CanopyRT module. This function initializes `canopyrt,canOptrt,canRadrt,arrayOfLeaves` as local variables rather than global variables as in CanopyRT module. See CanopyRT module for further operations on these variables.

Land.Plant.update_canopy_from_rt_module! — Function

update_canopy_from_rt_module!(tree::Tree, canopy_rt::struct_canopy, canOpt_rt::struct_canopyOptProps, canRad_rt::struct_canopyRadiation)

Updates canopy information from the RT module, given

tree A Tree type

canopy_rt A [struct_canopy] type from the Plant module

cadOpt_rt A [struct_canopyOptProps] type from the CanopyRT module

canRad_rt A [struct_canopyRadiation] type from the CanopyRT module

This interface function is pending for leaf temperature...

Water Physics (need generalization)

Land.Plant.get_relative_surface_tension — Function

get_relative_surface_tension(tem)

Surface tension of water relative to 25 degree C (298.15 K), given

tem Water temperature

The equations used are γ = V^(-2/3) * k * (tem_c-tem) γ/γ25 = (tem_c - tem) / (tem_c - 298.15) The empirical values are k = 2.1E-7 J K^-1 mol^(-2/3) V = 18.0 ml/mol tem_c = 647.0 K

May need to merge with other CLIMA repository to be consistent.

Land.Plant.get_relative_viscosity — Function

get_relative_viscosity(tem)

Viscosity relative to 25 degree C (298.15 K), given

tem Water temperature

Equations used are υ = A * exp( B/tem + C*tem + D*tem^2 ) υ/υ25 = exp( B/tem + C*tem + D*tem^2 - B/tem25 - C*tem25 - D*tem25^2 ) fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K A = 1.856E-11 mPa s B = 4209 K C = 0.04527 K⁻¹ D = -3.376E-5 K⁻²

May need to merge with other CLIMA repository to be consistent.

Land.Plant.get_saturated_vapor_pressure — Function

get_saturated_vapor_pressure(tem)

Saturated vapor pressure in [Pa], given

tem Water temperature

Saturated vapor pressure is computed using 611.0 * exp(17.502 * temc / (temc + 240.97)).

May need to merge with other CLIMA repository to be consistent.

get_saturated_vapor_pressure(tem)

List of saturated vapor pressure in [Pa], given

tem A list of temperatures

Saturated vapor pressure is computed using 611.0 * exp(17.502 * temc / (temc + 240.97)).

May need to merge with other CLIMA repository to be consistent.

Land.Plant.get_specific_latent_heat — Function

get_specific_latent_heat(tem)

Specific latent heat, given

tem Water temperature

Equation used is λ = (2500.8 - 2.36*tem + 0.0016*tem^2 -0.00006*tem^3) when tem in -25 to 40 degree C See Polynomial curve fits to Table 2.1. R. R. Rogers; M. K. Yau (1989). A Short Course in Cloud Physics (3rd ed.). Pergamon Press. p. 16. ISBN 0-7506-3215-1.

May need to merge with other CLIMA repository to be consistent.