Land model

Example Documentation

Canopy RT module

Leaf Bio Structure:

Land.CanopyRT.leafbio — Type

leafbio

Fields

N

Leaf structure parameter (-)

Cab

Chlorophyll a+b content (µg cm-2)

Car

Carotenoid content (µg cm-2)

Ant

Anthocynanin content (µg cm-2)

Cs

Senescent material fraction (-)

Cw

Equivalent water thickness (cm)

Cm

Dry matter content (dry leaf mass per unit area) (g cm-2)

Cx

Fractionation between Zeaxanthin and Violaxanthin in Car (1=all Zeaxanthin) (-)

ρ_LW

Broadband thermal reflectance (-)

τ_LW

Broadband thermal transmission (-)

fqe

Leaf fluorescence efficiency (Fo standard)

ρ_SW

shortwave leaf reflectance (-)

τ_SW

shortwave leaf transmission (-)

kChlrel

relative absorbtion by Chlorophyll+Car (-)

kChlrel_old

relative absorbtion by Chlorophyll (-)

Mb

Fluorescence excitation matrix backwards (-)

Mf

Fluorescence excitation matrix forwards (-)

Canopy Structure

Land.CanopyRT.struct_canopy — Type

struct_canopy

Fields

nlayers

number of canopy layers

LAI

Leaf Area Index

Ω

Clumping factor

clump_a

Structure factor a

clump_b

Structure factor b

leafwidth

Leaf width

hc

Vegetation height

LIDFa

Leaf Inclination

LIDFb

Variation in leaf inclination

hot

HotSpot parameter (still need to check!)

height

Canopy height (m)

z0m

Canopy roughness (m)

z0h

d

Canopy displacement height (m)

Cd

m/sqrt(s) turbulent transfer coefficient

lazitab

lidf

xl

dx

Canopy Radiation Structure

Land.CanopyRT.struct_canopyRadiation — Type

struct_canopyRadiation

Fields

intEout

integrated TOC outgoing flux (W m^-2)

incomingPAR

incident spectrally integrated total PAR (moles m^-2 s^-1)

incomingPAR_direct

incident spectrally integrated direct PAR (moles m^-2 s^-1)

incomingPAR_diffuse

incident spectrally integrated diffuse PAR (moles m^-2 s^-1)

RnSoil_diffuse

net radiation of shaded soil (W m^-2)

RnSoil_direct

net Short-wave radiation of sunlit soil (W m^-2)

RnSoil

net Short-wave radiation of soil (shaded + sunlit) (W m^-2)

RnSoilLW

net long-wave radiation of soil (shaded + sunlit) (W m^-2)

absPAR_shade

net PAR of shaded leaves (moles m^-2 s^-1)

absPAR_shadeCab

net PAR by Cab+Car of shaded leaves (moles m^-2 s^-1)

Lo

Short-wave TOC outgoing radiance in observation direction (mW m^-2 μm^-1 sr^-1)

Eout

Short-wave TOC outgoing radiation (mW m^-2 μm^-1)

alb_obs

Short-wave Albedo in viewing direction

alb_direct

Short-wave Albedo for direct incoming radiation

alb_diffuse

Short-wave Albedo for diffuse incoming radiation

E_up

upwelling diffuse short-wave radiation within canopy (mW m^-2 μm^-1)

E_down

downwelling diffuse short-wave radiation within canopy (mW m^-2 μm^-1)

netSW_sunlit

net absorbed direct radiation in each layer (mW m^-2 μm^-1)

netSW_shade

net absorbed diffuse radiation in each layer (mW m^-2 μm^-1)

intNetSW_sunlit

spectrally integrated net absorbed direct radiation in each layer (W m^-2)

intNetSW_shade

spectrally integrated net absorbed diffuse radiation in each layer (W m^-2)

intNetLW_sunlit

spectrally integrated net absorbed direct radiation in each layer (W m^-2)

intNetLW_shade

spectrally integrated net absorbed diffuse radiation in each layer (W m^-2)

absPAR_sun

net PAR of sunlit leaves moles m^-2 s^-1

absPAR_sunCab

net PAR by Cab+Car of sunlit leaves moles m^-2 s^-1

T_sun3D

Leaf temperature (sunlit) (K)

T_sun

Leaf temperature (sunlit) (K)

φ_sun

Fluorescence yield for sunlit leaves

T_shade

Leaf temperature (shaded) (K)

φ_shade

Fluorescence yield for shaded leaves

GPP_sun

GPP of sunlit leaves μmoles m^-2 s^-1

GPP_shade

GPP of shaded leaves μmoles m^-2 s^-1

SIF_hemi

Hemispheric total outgoing SIF flux (mW m^-2 μm^-1))

SIF_obs

Observer-direction outgoing SIF radiance (mW m^-2 μm^-1 sr^-1))

SIF_obs_sunlit

Observer-direction outgoing SIF radiance, sunlit leaves (mW m^-2 μm^-1 sr^-1))

SIF_obs_shaded

Observer-direction outgoing SIF radiance, shaded leaves (mW m^-2 μm^-1 sr^-1))

SIF_obs_scattered

Observer-direction outgoing SIF radiance, scattered (mW m^-2 μm^-1 sr^-1))

SIF_obs_soil

Observer-direction outgoing SIF radiance, soil-reflected (mW m^-2 μm^-1 sr^-1))

SIF_sum

Total SIF sum of layer sources (mW m^-2 μm^-1))

Canopy Optical Property Structure

Land.CanopyRT.struct_canopyOptProps — Type

struct_canopyRadiation

Fields

sdb

Solar -> Diffuse backscatter weight

sdf

Solar -> Diffuse forward scatter weight

dob

Diffuse -> Directional backscatter weight

dof

Diffuse -> Directional forward scatter weight

ddb

Diffuse -> Diffuse backscatter weight

ddf

Diffuse -> Diffuse forward scatter weight

ks

Solar beam extinction coefficient weight

ko

Outgoing beam extinction coefficient weight

bf

sob

weight of specular2directional backscatter coefficient

sof

weight of specular2directional forward coefficient

fs

conversion factor fs to compute irradiance on inclined leaf

absfs

abs(fs)

absfsfo

abs(fs*fo)

fsfo

fs*fo

fo

conversion factor fo for angle towards observer (not sun like fs)

cosΘ_l

Cosine of leaf azimuths

cos2Θ_l

cos of leaf azimuth sqared

Ps

Probability of directly viewing a leaf in solar direction

Po

Probability of directly viewing a leaf in viewing direction

Pso

Bi-directional probability of directly viewing a leaf (solar->canopy->viewing)

sigb

diffuse backscatter scattering coefficient for diffuse incidence

sigf

diffuse forward scattering coefficient for diffuse incidence

sb

diffuse backscatter scattering coefficient for specular incidence

sf

diffuse forward scattering coefficient for specular incidence

vb

directional backscatter scattering coefficient for diffuse incidence

vf

directional forward scattering coefficient for diffuse incidence

w

bidirectional scattering coefficent (directional-directional)

a

attenuation

Xsd

Effective layer transmittance (direct->diffuse)

Xdd

Effective layer transmittance (diffuse->diffuse)

R_sd

Effective layer reflectance (direct->diffuse)

R_dd

Effective layer reflectance (diffuse->diffuse)

Es_

Solar direct radiation per layer)

Photosynthesis module

Leaf Params:

Land.Photosynthesis.leaf_params — Type

leaf_params
Structure with all parameter temperature dependencies of a Leaf (largely based on Bonan's Photosynthesis model but exported as struct)

Fields

α

ε

LMA

DRY leaf mass area

RWC

leaf relative water content

Cleaf

leaf specific heat (J/kg/K)

Chloroplast_rel_volume

PSII_frac

Fraction of absorbed light used by PSII ETR

Kf

Rate constant for fluorescence (const)

Kd

Rate constant for thermal dissipation

Kp

Rate constant for photochemistry (all reaction centers open)

Kn

NPQ rate constant

Kn_ss

maxPSII

max PSII yield

Fo

dark-adapted fluorescence yield Fo

Fo′

light-adapted fluorescence yield in the dark Fo'

Fm

dark adapted Fm yield

Fm′

light adapted Fm' yield

ϕs

steady-state (light-adapted) yield Ft (aka Fs)

φ

PSII yield

qQ

Photochemical quenching

qE

NPQ

Non-Photochemical quenching

dynamic_state

Dynamic Steady state boolean

effcon

CO2_per_electron

Ja

Actual electron transport rate (μmol m-2 s-1)

C3

C3 or C4 plant boolean

use_colim

gstyp

o₂

O2 (mol/mol)

T

Leaf temperature (K)

T_old

Old Leaf temperature (K)

esat

desat

θ_j

vpd_min

gleaf

total leaf conductance (μmol/m2/s)

gm

Mesophyll conductance (μmol/m2/s/)

gs

Stomatal conductance (μmol/m2/s)

gs_ss

Steady state Stomatal conductance (μmol/m2/s)

ppm_to_Pa

ppm to Pa conversion

Ag

An

Ai

Ac

Aj

Ap

H

LE

Rn

Sap

APAR

APAR

Kc

Michaelis-Menten constant for CO₂ (Pa)

Ko

Michaelis-Menten constant for O₂ (Pa)

Kpep

Michaelis-Menten constant for PEP carboxylase (Pa)

Γstar

CO₂ compensation point (Pa)

Vcmax25

Leaf maximum carboxylation rate at 25C (μmol/m2/s)

Vpmax25

Leaf maximum PEP carboxylation rate at 25C (μmol/m2/s)

Jmax25

Maximum electron transport rate at 25C (μmol/m2/s)

Rd25

Leaf respiration rate at 25C (μmol CO2/m2/s)

Vcmax

Actual Leaf maximum carboxylation rate (μmol/m2/s)

Vpmax

Actual Leaf maximum PEP carboxylase rate (μmol/m2/s)

Jmax

Actual Maximum electron transport rate (μmol/m2/s)

Rd

Actual Leaf respiration rate (μmol CO2/m2/s)

Cs

CO2 concentration at leaf surface (ppm)

Cc

CO2 concentration in chloroplast (ppm)

Cc_Pa

CO2 concentration in chloroplast (Pa)

RH

relative humidity at the surface of the leaf

VPD

Vapor pressure difference across leaf interface (kPa)

Je

Electron transport rate (μmol m-2 s-1), light limited only

Je_pot

Potential ETR

height

z0m

z0h

d

dleaf

LAI

Cd

psi_l

psi_l50

kmax

kx

kleaf

ck

Ctree

Plant module

Plant structure

Land.Plant.Tree — Type

Tree{FT<:AbstractFloat}

A Tree type which, by default, includes

5 root layers

a trunk

20 branches

20 canopy layers

Fields

age

Age [year]

ba

Basal area [m²]

ga

Ground area [m²]

h

Tree height [m]

roots

Root which contains 5 RootLayer by default

trunk

Trunk using type Stem

branch

Branch which contains 20 Stem by default

canopy

Canopy which contains 20 CanopyLayer by default

Land.Plant.Root — Type

Root{FT<:AbstractFloat,n}

A Root Struct which contains, by default, 5 even RootLayer.

Fields

root_frac

List of root fraction

root_zs

List of root z lower boundary

root_list

List of RootLayer

Land.Plant.RootLayer — Type

RootLayer{FT<:AbstractFloat}

A RootLayer type which contains root hydraulics information, including rhizosphere conductance.

Fields

z_hi

Upper root depth [m]

z_lo

Lower root depth [m]

r_layer

Average horizental root length [m]

f_layer

Fraction of biomass and k_max

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance, default at 2700.0 Kg h⁻¹ MPa⁻¹ m⁻² basal area [mol s⁻¹ MPa⁻¹]

k_s

Maximal hydraulic conductivity per cross section basal area per root depth [mol s⁻¹ MPa⁻¹ m⁻²]

k_rhiz

Maximal rhizosphere conductance, default at 9.72e12 Kg h⁻¹ MPa⁻¹ mm⁻² basal area [mol s⁻¹ MPa⁻¹]

p_ups

Soil water potential (upstream) [MPa]

soil_a

Soil texture parameter

soil_m

Soil texture parameter, 1 - 1/n

soil_n

Soil texture parameter

soil_msc

Maximal soil volumatric water content, vary with rock fraction

soil_rwc

Relative soil volumatric water content

soil_vwc

Soil volumatric water content

t_soil

Soil temperature [K]

soil_type

Soil texture

p_rhiz

Water potential at the root-rhizosphere interface [MPa]

p_dos

Xylem water pressure at the tree base (downstream) [MPa]

q

Flow rate in the root layer [mol s⁻¹]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹]

p_element

List of xylem water pressure per element [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of water temperature per element [K]

z_element

List of height per element to account for gravity [m]

Land.Plant.Stem — Type

Stem{FT<:AbstractFloat}

A Stem type which contains hydraulic information, including drought history profile. The Stem type is used for creating a trunk and a list of branches.

Fields

r_layer

Horizental radius of the layer [m]

z_hi

Upper height from the ground (trunk-branch joint for trunk, stem-leaf joint for branch) [m]

z_lo

Lower height from the ground (0 for trunk, trunk-branch joint for branch) [m]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance, = 5400.0 * 8/3 Kg h⁻¹ MPa⁻¹ m⁻² basal area [mol s⁻¹ MPa⁻¹]

k_s

Maximal hydraulic conductivity per cross section area per tree height [mol s⁻¹ MPa⁻¹ m⁻²]

p_ups

Xylem water pressure at the tree basa (upstream) [MPa]

p_dos

Xylem water pressure at the trunk-stem or stem-leaf joint (downstream) [MPa]

q

Flow rate in the xylem [mol s⁻¹]

k_element

List of stem k_max per element [mol s⁻¹ MPa⁻¹]

p_element

List of stem xylem pressure per element [MPa]

p_history

List of stem xylem pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of stem temperature for each element [K]

z_element

List of stem height per element to account for gravity [m]

Land.Plant.Branch — Type

Branch{FT<:AbstractFloat, n}

A Branch type which constains n Stem in response to n CanopyLayer

Fields

branch_list

List of Stem

Land.Plant.Canopy — Type

Canopy{FT<:AbstractFloat,n, n_total}

A Canopy type which contains n CanopyLayer.

Fields

n_layer

Canopy layer number

canopy_list

List of CanopyLayer

Land.Plant.CanopyLayer — Type

CanopyLayer{FT<:AbstractFloat, n_total}

A CanopyLayer type, which constains environemntal conditions and leaf-level fluxes for n_total Leaf. The n_total is the sum of n_Ari * n_Incli sunlit leaves and 1 shaded leaves.

Fields

f_layer

Leaf area fraction in the canopy

f_view

Fraction of the sunlit leaves

la

Leaf area in the layer [m²]

width

Leaf width [m]

p_a

Atmospheric CO₂ partial pressure in the layer [Pa]

p_atm

Atmospheric pressure [Pa]

p_O₂

Atmospheric O₂ partial presssure [Pa]

p_H₂O

Atmospheric H₂O partial pressure [Pa]

t_air

Air temperature [K]

wind

Wind speed [m s⁻¹]

g_ias_c

Mesophyll conductance correction factor: multiplier

g_ias_e

Mesophyll conductance correction factor: exponent

g_max

Maximal leaf diffusive conductance for H₂O at 298.15 K [mol m⁻² s⁻¹]

Γ_star

CO₂ compensation point with the absence of dark respiration [Pa]

gs_nssf

Non-steady state factor (use it by multiplying the ∂A/∂E - ∂Θ/∂E)

j_max

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

r_25

Leaf respiration rate at 298.15 K [μmol m⁻² s⁻¹]

v_max

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

leaf_list

List of Leaf

ag_list

List of gross photosynthetic rate for n_total leaves [μmol m⁻² s⁻¹]

an_list

List of net photosynthetic rate [μmol m⁻² s⁻¹]

d_list

List of leaf-to-air vapor pressure deficit [unitless]

e_list

List of flow rate per leaf area [mol m⁻² s⁻¹]

ec_list

List of critical flow rate (when leaf xylem pressure induces desiccation) per leaf area [mol m⁻² s⁻¹]

gsc_list

List of effective leaf diffusive conductance for CO₂ [mol m⁻² s⁻¹]

gsw_list

List of leaf diffusive conductance for H₂O [mol m⁻² s⁻¹]

la_list

List of leaf area per leaf [m²]

par_list

List of PAR for each leaf

pi_list

List of leaf internal CO₂ partial pressure [Pa]

q_list

List of flow rate [mol s⁻¹]

r_list

List of leaf respiration rate [μmol m⁻² s⁻¹]

t_list

List of leaf temperature [K]

Land.Plant.Leaf — Type

Leaf{FT<:AbstractFloat}

A Leaf type which contains leaf hydraulics information.

Fields

angle_locat

direction from the tree trunk to the leaf, 0 for east, 270 for south, and 180 for west [°]

angle_incli

direction of leave, 0 for flat and 90 for vertical [°]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_sla

Maximal leaf hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

p_ups

Leaf xylem water pressure (different from leaf water potential) at the leaf base (upstream) [MPa]

p_dos

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

k_element

List of leaf k_max per element (mol s⁻¹ MPa⁻¹)

p_element

List of xylem water pressure per element [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) per element [MPa]

t_element

List of xylem water temperature per element [K]

z_element

List of leaf element height change to account for gravity, default = 0 [m]

Plant hydraulics

Land.Plant.get_struct_p_end_from_q — Function

get_struct_from_q(root_layer, flow; p_ini)

End pressure p_end (in MPa) for RootLayer, including impacts from rhizosphere conductance and gravity, given

root_layer One RootLayer in Root in Tree

flow Flow rate (in mol s⁻¹) in the given RootLayer

p_ini Upstream soil water potential if p_ini is given, otherwise (Inf) root_layer.p_ups will be used

get_struct_from_q(stem, flow; p_ini)

End pressure p_end (in MPa) for stem Stem, including impact from gravity, given

stem Stem as Trunk in Tree or Stem as side branch in Branch in Tree

flow Flow rate (in mol s⁻¹) in the given Stem

p_ini Upstream xylem pressure if p_ini is given, otherwise (Inf) stem.p_ups will be used

get_struct_from_q(leaf, flow; p_ini)

End pressure p_end (in MPa) for leaf leaf, including impact from gravity, given

leaf Leaf in CanopyLayer in Canopy in Tree

flow Flow rate per leaf area (mol m⁻² s⁻¹) in the given Leaf

p_ini Upstream xylem pressure if p_ini is given, otherwise (Inf) leaf.p_ups will be used

Land.Plant.get_q_layer_from_p_base — Function

get_q_layer_from_p_base(root_layer, p_base)

Flow rate in a root layer q_layer, given

root_layer One RootLayer in Root in Tree

p_base Xylem water pressure at the tree base

This function use Newton Raphson combined with Bi-section to get q_layer. A warning message will display if total iterations exceed 50 times.

Land.Plant.get_p_base_q_list_from_q — Function

get_p_base_q_list_from_q(tree, flow)

Tree base pressyre p_base and a list of flow rate in each root layer q_list, given

tree A Tree type

flow Flow rate for the whole tree [mol s⁻¹]

This function returns the p_base from total transpiration rate using Newton Raphson. A warning will display if total iterations is beyond 50 times.

Land.Plant.update_struct_from_q! — Function

update_struct_from_q!(root_layer, flow; p_ini)

Update root hydraulic information considering the impact from rhizosphere conductance and gravity, given

root_layer One RootLayer in Root in Tree

flow Flow rate in the given root layer

p_ini Upstream soil water potential if p_ini is given, otherwise (Inf) root_layer.p_ups will be used

update_struct_from_q!(stem, flow; p_ini)

Update the stem hydraulic information considering the impact from gravity, given

stem Stem as trunk in a Tree or as side branch in Branch in Tree

flow::FT Flow rate in the given stem

p_ini::FT Upstream xylem pressure if p_ini is given, otherwise (Inf) stem.p_ups will be used

update_struct_from_q!(leaf, flow; p_ini)

Update the leaf hydraulic information considering the impact from gravity, given

leaf Leaf type in CanopyLayer in a Tree

flow Flow rate per leaf area in the given leaf

p_ini Upstream xylem pressure if p_ini is given, otherwise (Inf) leaf.p_ups will be used

Land.Plant.update_tree_e_crit! — Function

update_tree_e_crit!(tree; displaying)

Update the critical transpiration rate information, given

tree A Tree type

displaying If true, messages of how e_crit evolves step-by-step will display

This function update e_crit for each layer (sunlit and shaded). The e_crit for each leaf is either increased or decreased so that p_leaf is in the range -Inf to -20.0 MPa. The e_crit increases by 1e-6 mol m⁻² s⁻¹ if p_leaf is less negative than -20 MPa. The e_crit decreases by 1e-7 mol m⁻² s⁻¹ if p_leaf is -Inf.

Stoma dynamics

Land.Plant.get_marginal_gain — Function

get_marginal_gain(canopyi, indx)

Marginal water use efficiency ∂A/∂E for a leaf in a canopy layer, given

canopyi A CanopyLayer in Canopy in a Tree

indx indx'th Leaf in the canopyi.leaf_list

get_marginal_gain(canopyi, indx)

A list of marginal water use efficiency for all the leaves, given

canopyi A CanopyLayer in Canopy in a Tree

Land.Plant.get_marginal_penalty_wang — Function

get_marginal_penalty_wang(canopyi, indx)

Marginal water penalty (∂Θ/∂E) for opening the stomata for a leaf in canopy layer, given

canopyi A CanopyLayer in Canopy in a Tree

indx indx'th Leaf in the canopyi.leaf_list

This function uses the Wang 2020 model. ∂Θ/∂E = A / (e_crit - e_leaf), unit [mol mol⁻¹]. Should be updated to Abstract Type once more models are added

get_marginal_penalty_wang(canopyi, indx)

List of marginal water penalty for all the leaves, given

canopyi A CanopyLayer in Canopy in a Tree

This function uses the Wang 2020 model. ∂Θ/∂E = A / (e_crit - e_leaf), unit [mol mol⁻¹]. Should be updated to Abstract Type once more models are added

Land.Plant.update_tree_with_time! — Function

update_tree_with_time!(tree, Δt; scheme="Wang2020", updating=false)

Update the tree flux information and hydraulic parameters, given

tree One Tree type

Δt Δt

scheme Optimization model option, default is "Wang2020

updating If true, plant hydraulic parameters will be updated for the Tree

The function updates the stomatal conductance via a Δgsw = factor * (∂A/∂E - ∂Θ/∂E). The ∂A/∂E is the same for every stomatal control model, but ∂Θ/∂E differs among models.

Interfaces

Land.Plant.update_canopy_from_rt_module! — Function

update_canopy_from_rt_module!(tree)

Updates canopy information from the RT module, given

tree A Tree type

This interface function is pending...