CanopyRT module for the LAND model

Canopy RT structures

Canopy

Land.CanopyRT.Canopy4RT — Type

struct Canopy4RT

A canopy struct for the radiation transfer module

Fields

nlayers

Number of canopy layers

ndub

Doubling adding later

LAI

Leaf Area Index

Ω

Clumping factor

clump_a

Structure factor a

clump_b

Structure factor b

leaf_width

Leaf width

hc

Vegetation height

LIDFa

Leaf Inclination

LIDFb

Variation in leaf inclination

hot

HotSpot parameter (still need to check!)

height

Canopy height [m]

z0m

Canopy roughness [m]

z0h

Tree roughtnes [m]

d

Canopy displacement height [m]

Cd

m/sqrt(s) turbulent transfer coefficient

litab

litab_bnd

lazitab

lidf

xl

dx

Canopy optical parameters

Land.CanopyRT.AbstractCanopyOpti — Type

AbstractCanopyOpti

A abstract type for canopy optical properties

Land.CanopyRT.CanopyOptiArray — Type

struct CanopyOptiArray{FT, nWL, nLayer, nLevel, nAzi, nIncl}

A struct for canopy optical properties using Array

Fields

sdb

Solar -> Diffuse backscatter weight

sdf

Solar -> Diffuse forward scatter weight

dob

Diffuse -> Directional backscatter weight

dof

Diffuse -> Directional forward scatter weight

ddb

Diffuse -> Diffuse backscatter weight

ddf

Diffuse -> Diffuse forward scatter weight

ks

Solar beam extinction coefficient weight

ko

Outgoing beam extinction coefficient weight

bf

?

sob

Weight of specular2directional backscatter coefficient

sof

Weight of specular2directional forward coefficient

Ps

Probability of directly viewing a leaf in solar direction

Po

Probability of directly viewing a leaf in viewing direction

Pso

Bi-directional probability of directly viewing a leaf (solar->canopy->viewing)

fs

conversion factor fs to compute irradiance on inclined leaf

absfs

abs(fs)

absfsfo

abs(fs*fo)

fsfo

fs*fo

fo

conversion factor fo for angle towards observer (not sun like fs)

cosΘ_l

Cosine of leaf azimuths

cos2Θ_l

cos of leaf azimuth sqared

sigb

diffuse backscatter scattering coefficient for diffuse incidence

sigf

diffuse forward scattering coefficient for diffuse incidence

sb

diffuse backscatter scattering coefficient for specular incidence

sf

diffuse forward scattering coefficient for specular incidence

vb

directional backscatter scattering coefficient for diffuse incidence

vf

directional forward scattering coefficient for diffuse incidence

w

bidirectional scattering coefficent (directional-directional)

a

attenuation

Xsd

Effective layer transmittance (direct->diffuse)

Xdd

Effective layer transmittance (diffuse->diffuse)

R_sd

Effective layer reflectance (direct->diffuse)

R_dd

Effective layer reflectance (diffuse->diffuse)

Es_

Solar direct radiation per layer)

Land.CanopyRT.CanopyOptiMArray — Type

struct CanopyOptiMArray{FT, nWL, nLayer, nLevel, nAzi, nIncl, nLayer_nWL, nLev_nWL, nIncl_nAzi}

A struct for canopy optical properties using MArray

Fields

sdb

Solar -> Diffuse backscatter weight

sdf

Solar -> Diffuse forward scatter weight

dob

Diffuse -> Directional backscatter weight

dof

Diffuse -> Directional forward scatter weight

ddb

Diffuse -> Diffuse backscatter weight

ddf

Diffuse -> Diffuse forward scatter weight

ks

Solar beam extinction coefficient weight

ko

Outgoing beam extinction coefficient weight

bf

?

sob

Weight of specular2directional backscatter coefficient

sof

Weight of specular2directional forward coefficient

Ps

Probability of directly viewing a leaf in solar direction

Po

Probability of directly viewing a leaf in viewing direction

Pso

Bi-directional probability of directly viewing a leaf (solar->canopy->viewing)

fs

conversion factor fs to compute irradiance on inclined leaf

absfs

abs(fs)

absfsfo

abs(fs*fo)

fsfo

fs*fo

fo

conversion factor fo for angle towards observer (not sun like fs)

cosΘ_l

Cosine of leaf azimuths

cos2Θ_l

cos of leaf azimuth sqared

sigb

diffuse backscatter scattering coefficient for diffuse incidence

sigf

diffuse forward scattering coefficient for diffuse incidence

sb

diffuse backscatter scattering coefficient for specular incidence

sf

diffuse forward scattering coefficient for specular incidence

vb

directional backscatter scattering coefficient for diffuse incidence

vf

directional forward scattering coefficient for diffuse incidence

w

bidirectional scattering coefficent (directional-directional)

a

attenuation

Xsd

Effective layer transmittance (direct->diffuse)

Xdd

Effective layer transmittance (diffuse->diffuse)

R_sd

Effective layer reflectance (direct->diffuse)

R_dd

Effective layer reflectance (diffuse->diffuse)

Es_

Solar direct radiation per layer)

Land.CanopyRT.create_canopy_optical — Function

create_canopy_optical(FType, nWL::Int, nLayers::Int, nAzi::Int, nIncl::Int; using_marray=false)

Create a canopy optical properties struct with an using_marray option, given

FType Floating number type

nWL Number of wave length

nLayers Number of canopy layers

nAzi Number of arimuth angles

AIncl Number of inclination angles

using_marray If true, using MArray; else, using Array

Canopy radiation

Land.CanopyRT.CanopyRadiation — Type

struct CanopyRadiation{FT, nWl, nWlF, nIncl, nAzi, nLayers}

A struct for canopy radiation information

Fields

intEout

Integrated TOC outgoing flux [W m⁻²]

incomingPAR

Incident spectrally integrated total PAR [mol m⁻² s⁻¹]

incomingPAR_direct

Incident spectrally integrated direct PAR [mol m⁻² s⁻¹]

incomingPAR_diffuse

Incident spectrally integrated diffuse PAR [mol m⁻² s⁻¹]

RnSoil_diffuse

Net radiation of shaded soil [W m⁻²]

RnSoil_direct

Net Short-wave radiation of sunlit soil [W m⁻²]

RnSoil

Net Short-wave radiation of soil (shaded + sunlit) [W m⁻²]

RnSoilLW

Net long-wave radiation of soil (shaded + sunlit) [W m⁻²]

absPAR_shade

Net PAR of shaded leaves [mol m⁻² s⁻¹]

absPAR_shadeCab

Net PAR by Cab+Car of shaded leaves [moles m⁻² s⁻¹]

intNetSW_sunlit

Spectrally integrated net absorbed direct radiation in each layer [W m⁻²)]

intNetSW_shade

Spectrally integrated net absorbed diffuse radiation in each layer [W m⁻²)]

intNetLW_sunlit

Spectrally integrated net absorbed direct radiation in each layer [W m⁻²)]

intNetLW_shade

Spectrally integrated net absorbed diffuse radiation in each layer [W m⁻²)]

T_sun

Leaf temperature (sunlit) [K]

T_shade

Leaf temperature (shaded) [K]

ϕ_shade

Fluorescence yield for shaded leaves

H_shade

Sensible Heat flux H of shaded leaves [W m⁻²]

LE_shade

Latent Heat flux LE of shaded leaves [W m⁻²]

NPQ_shade

NPQ of shaded leaves

GPP_shade

GPP of shaded leaves [μmol m⁻² s⁻¹]

gs_shade

gs of shaded leaves [mol m⁻² s⁻¹]

ψl_shade

Leaf water potential of shaded leaves [MPa]

Cc_shade

Cc of shaded leaves [µmol/mol]

Pi_shade

internal CO₂ concentration of shaded leaves [µmol/mol]

Lo

Short-wave TOC outgoing radiance in observation direction [mW m⁻² μm⁻¹ sr⁻¹]

Eout

Short-wave TOC outgoing radiation [mW m⁻² μm⁻¹]

alb_obs

Short-wave Albedo in viewing direction

alb_direct

Short-wave Albedo for direct incoming radiation

alb_diffuse

Short-wave Albedo for diffuse incoming radiation

E_up

Upwelling diffuse short-wave radiation within canopy [mW m⁻² μm⁻¹]

E_down

Downwelling diffuse short-wave radiation within canopy [mW m⁻² μm⁻¹]

netSW_sunlit

Net absorbed direct radiation in each layer [mW m⁻² μm⁻¹]

netSW_shade

net absorbed diffuse radiation in each layer [mW m⁻² μm⁻¹]

absPAR_sun

net PAR of sunlit leaves [mol m⁻² s⁻¹]

absPAR_sunCab

net PAR by Cab+Car of sunlit leaves [mol m⁻² s⁻¹]

T_sun3D

Leaf temperature (sunlit) [K]

ϕ_sun

Fluorescence yield for sunlit leaves

H_sun

Sensible Heat flux H of sunlit leaves [W m⁻²]

LE_sun

Latent Heat flux LE of sunlit leaves [W m⁻²]

NPQ_sun

NPQ of sunlit leaves

GPP_sun

GPP of sunlit leaves [μmol m⁻² s⁻¹]

gs_sun

gs of sunlit leaves [mol m⁻² s⁻¹]

ψl_sun

Leaf water potential of sunlit leaves [MPa]

Cc_sun

Cc of sunlit leaves [µmol/mol]

Pi_sun

Internal CO₂ concentration of sunlit leaves [µmol/mol]

SIF_hemi

Hemispheric total outgoing SIF flux [mW m⁻² μm⁻¹])

SIF_obs

Observer-direction outgoing SIF radiance (mW m⁻² μm⁻¹ sr⁻¹))

SIF_obs_sunlit

Observer-direction outgoing SIF radiance, sunlit leaves (mW m⁻² μm⁻¹ sr⁻¹))

SIF_obs_shaded

Observer-direction outgoing SIF radiance, shaded leaves (mW m⁻² μm⁻¹ sr⁻¹))

SIF_obs_scattered

Observer-direction outgoing SIF radiance, scattered (mW m⁻² μm⁻¹ sr⁻¹))

SIF_obs_soil

Observer-direction outgoing SIF radiance, soil-reflected (mW m⁻² μm⁻¹ sr⁻¹))

SIF_sum

Total SIF sum of layer sources [mW m⁻² μm⁻¹])

Land.CanopyRT.AbstractIncomingRadiation — Type

AbstractIncomingRadiation

Abstract leaf biological parameters type

Land.CanopyRT.IncomingRadiationArray — Type

struct IncomingRadiationArray{FT}

Incoming radiation information.

Fields

wl

Wavelength [nm]

E_direct

Direct incoming radiation [mW m⁻² μm⁻¹]

E_diffuse

Diffuse incoming radiation [mW m⁻² μm⁻¹]

Land.CanopyRT.IncomingRadiationMArray — Type

struct IncomingRadiationMArray{FT, N}

Incoming radiation information.

Fields

wl

Wavelength [nm]

E_direct

Direct incoming radiation [mW m⁻² μm⁻¹]

E_diffuse

Diffuse incoming radiation [mW m⁻² μm⁻¹]

Land.CanopyRT.create_incoming_radiation — Function

create_incoming_radiation(FType, swl:Array, file; using_marray=false)

Create an AbstractIncomingRadiation struct, given

FType Floating number type

swl Standard wave length

file Input file name

using_marray If true, use MArray; else, using Array

Leaf biological parameters

Land.CanopyRT.AbstractLeafBio — Type

AbstractLeafBio

Abstract leaf biological parameters type

Land.CanopyRT.LeafBioArray — Type

struct LeafBioArray{FT, nWl, nWle, nWlf}

A struct of leaf biological parameters using Array

Fields

N

Leaf structure parameter

Cab

Chlorophyll a+b content [µg cm⁻²]

Car

Carotenoid content [µg cm⁻²]

Ant

Anthocynanin content [µg cm⁻²]

Cs

Senescent material fraction

Cw

Equivalent water thickness [cm]

Cm

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

Cx

Fractionation between Zeaxanthin and Violaxanthin in Car (1=all Zeaxanthin) (-)

ρ_LW

Broadband thermal reflectance (-)

τ_LW

Broadband thermal transmission (-)

fqe

Leaf fluorescence efficiency (Fo standard)

ρ_SW

Shortwave leaf reflectance

τ_SW

Shortwave leaf transmission

kChlrel

Relative absorbtion by Chlorophyll+Car

kChlrel_old

Relative absorbtion by Chlorophyll

Mb

Fluorescence excitation matrix backwards

Mf

Fluorescence excitation matrix forwards

Land.CanopyRT.LeafBioMArray — Type

struct LeafBioMArray{FT, nWl, nWle, nWlf, nWLe_nWLf}

A struct of leaf biological parameters using Array

Fields

N

Leaf structure parameter

Cab

Chlorophyll a+b content [µg cm⁻²]

Car

Carotenoid content [µg cm⁻²]

Ant

Anthocynanin content [µg cm⁻²]

Cs

Senescent material fraction

Cw

Equivalent water thickness [cm]

Cm

Dry matter content (dry leaf mass per unit area) [g cm⁻²]

Cx

Fractionation between Zeaxanthin and Violaxanthin in Car (1=all Zeaxanthin)

ρ_LW

Broadband thermal reflectance

τ_LW

Broadband thermal transmission

fqe

Leaf fluorescence efficiency (Fo standard)

ρ_SW

Shortwave leaf reflectance

τ_SW

Shortwave leaf transmission

kChlrel

Relative absorbtion by Chlorophyll+Car

kChlrel_old

Relative absorbtion by Chlorophyll

Mb

Fluorescence excitation matrix backwards

Mf

Fluorescence excitation matrix forwards

Land.CanopyRT.create_leaf_bio — Function

create_leaf_bio(FT, nWl::Int, nWle::Int, nWlf::Int; using_marray=false)

Create a leaf biological parameters struct with an using_marray option, given

FType Floating number type

nWl Number of wave length

nWle Number of excitation wave length

nWlf Number of fluorescence wave length

using_marray If true, using MArray; else, using Array

Leaf optical parameters

Land.CanopyRT.AbstractLeafOptiPara — Type

AbstractLeafOptiPara

Abstract leaf optical parameter

Land.CanopyRT.LeafOptiParaArray — Type

struct LeafOptiParaArray{FT, N}

Struct for leaf optical properties using Array

Fields

nr

Km

Kab

Kant

Kcar

Kw

KBrown

phi

KcaV

KcaZ

lambda

Land.CanopyRT.LeafOptiParaMArray — Type

struct LeafOptiParaMArray{FT, N}

Struct for leaf optical properties using MArray

Fields

nr

Km

Kab

Kant

Kcar

Kw

KBrown

phi

KcaV

KcaZ

lambda

Land.CanopyRT.create_opti_par — Function

create_opti_par(FType, swl::Array, file; using_marray=false)

Create an AbstractLeafOptiPara struct, given

FType Floating number type

swl Standard wave length

file Input file name

using_marray If true, use MArray; else, using Array

Soil

Land.CanopyRT.SoilOpti — Type

struct SoilOpti{FT}

A struct of soil optical parameters

Fields

wl

Wavelength [nm]

albedo_SW

Shortwave albedo

albedo_LW

Longwave albedo

soil_skinT

Soil surface temperature [K]

Solar angles

Land.CanopyRT.SolarAngles — Type

struct SolarAngles{FT}

Struct for observation and solar angles

Fields

tts

Solar Zenith Angle [degree]

tto

Viewing Zenith Angle in [degree]

psi

relative azimuth in [degree]

Pre-set wave length settings

Land.CanopyRT.AbstractWLParaSet — Type

AbstractWLParaSet

An abstract type for wave length parameter set, which corresponds to the AbstractLeafOptiPara

Land.CanopyRT.WLParaSetArray — Type

struct WLParaSetArray{FT}

Struct for pre-set wave length parameters.

Land.CanopyRT.WLParaSetMArray — Type

struct WLParaSetMArray{FT}

Struct for pre-set wave length parameters.

TODO add MArray version?

Land.CanopyRT.create_wl_para_set — Function

create_wl_para_set(FType; using_marray::Bool=false)

Create a pre-set struct of wave length settings, given

Ftype Floating number type

using_marray If true, use MArray; else, use Array.

TODO add MArray version?