Photosynthesis module for the LAND model

Temperature Dependency

Types

Land.Photosynthesis.AbstractTDParameterSet — Type

type AbstractTDParameterSet{FT}

Hierarchy of the AbstractTDParameterSet:

ArrheniusTD

ArrheniusPeakTD

Land.Photosynthesis.ArrheniusTD — Type

struct ArrheniusTD{FT}

An AbstractTDParameterSet type struct using

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right)\]

Fields

VAL_25

Uncorrected value at 298.15 K

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

Land.Photosynthesis.ArrheniusPeakTD — Type

struct ArrheniusPeakTD{FT}

An AbstractTDParameterSet type struct using

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Parameter sets

Land.Photosynthesis.JmaxTDBernacchi — Function

ArrheniusPeakTD type Jmax TD

Land.Photosynthesis.JmaxTDCLM — Function

ArrheniusPeakTD type Jmax TD

Land.Photosynthesis.JmaxTDLeuning — Function

ArrheniusPeakTD type Jmax TD from Leuning's data

Land.Photosynthesis.KcTDBernacchi — Function

ArrheniusTD type Kc TD from Bernacchi's data

Land.Photosynthesis.KcTDCLM — Function

ArrheniusTD type Kc TD

Land.Photosynthesis.KoTDBernacchi — Function

ArrheniusTD type Ko TD from Bernacchi's data

Land.Photosynthesis.KoTDCLM — Function

ArrheniusTD type Ko TD

Land.Photosynthesis.KpepTDBoyd — Function

ArrheniusTD type Kpep TD from Boyd's data

Land.Photosynthesis.KpepTDCLM — Function

ArrheniusTD type Kpep TD

Land.Photosynthesis.RespirationTDBernacchi — Function

ArrheniusTD type Respiration TD from Bernacchi's data

Land.Photosynthesis.RespirationTDCLM — Function

ArrheniusPeakTD type Respiration TD

Land.Photosynthesis.VcmaxTDBernacchi — Function

ArrheniusTD type Vcmax TD from Bernacchi's data

Land.Photosynthesis.VcmaxTDCLM — Function

ArrheniusPeakTD type Vcmax TD

Land.Photosynthesis.VcmaxTDLeuning — Function

ArrheniusPeakTD type Vcmax TD from Leuning's data

Land.Photosynthesis.VomaxTDBernacchi — Function

ArrheniusTD type Vomax TD from Bernacchi's data

Land.Photosynthesis.VpmaxTDBoyd — Function

ArrheniusPeakTD type Vpmax TD from Boyd's data

Land.Photosynthesis.ΓStarTDBernacchi — Function

ArrheniusTD type $Γ^{*}$ TD from Bernacchi's data

Land.Photosynthesis.ΓStarTDCLM — Function

ArrheniusTD type Γ* TD

Photosynthesis Colimitation

Land.MathTools.AbstractColimitation — Type

type AbstractColimitation

Hierarchy of the AbstractColimitation:

MinColimit

CurvedColimit

Land.MathTools.MinColimit — Type

struct MinColimit

Takes the minimum of all computed photosynthetic rates.

Land.MathTools.CurvedColimit — Type

struct CurvedColimit{FT}

Takes curved limitation for computed photosynthetic rates.

Fields

curvature

Curveature ratio

Fluorescence Model

Types

Land.Photosynthesis.AbstractFluoModelParaSet — Type

type AbstractFluoModelParaSet

Hierarchy of the AbstractFluoModelParaSet:

FluoParaSet

Land.Photosynthesis.FluoParaSet — Type

struct FluoParaSet{FT}

A AbstractFluoModelParaSet type paramter set.

Fields

Kn1

Fluorescence model coefficient

Kn2

Fluorescence model coefficient

Kn3

Fluorescence model coefficient

Parameter sets

Land.Photosynthesis.FluorescenceFlexas — Function

FluoParaSet type parameter set using Flexas's data

Respiration Quantification

Land.Photosynthesis.VtoRCollatz — Function

A constant of 0.01

Land.Photosynthesis.VtoRDefault — Function

A constant of 0.015

Stomatal Model Scheme

Land.Photosynthesis.AbstractStomatalModel — Type

type AbstractStomatalModel

Hierarchy of the AbstractStomatalModel:

EmpiricalStomatalModel

OptimizationStomatalModel

Land.Photosynthesis.EmpiricalStomatalModel — Type

type EmpiricalStomatalModel

Hierarchy of the EmpiricalStomatalModel:

ESMBallBerry

ESMGentine

ESMLeuning

ESMMedlyn

Land.Photosynthesis.OptimizationStomatalModel — Type

type OptimizationStomatalModel

Hierarchy of the OptimizationStomatalModel:

OSMEller

OSMSperry

OSMWang

OSMWAP

OSMWAPMod

Empirical Stomatal Model

Land.Photosynthesis.ESMBallBerry — Type

struct ESMBallBerry{FT}

An empirical model parameter set type for Ball-Berry type model. The equation used for Ball-Berry type model is

\[gs = g0 + g1 ⋅ RH ⋅ \dfrac{A}{Cs}\]

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [unitless]

Land.Photosynthesis.ESMGentine — Type

struct ESMGentine{FT}

An empirical model parameter set type for Gentine type model. The equation used for Gentine type model is

\[gs = g0 + g1 ⋅ \dfrac{k_{leaf}}{k_{max}} ⋅ \dfrac{A}{Ca}.\]

Note it that the Gentine model does not require for a β function to tune the soil drought response, but the use of k_leaf also does not permit post-drought stomatal response unless k_leaf can be recovered.

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [unitless]

Land.Photosynthesis.ESMLeuning — Type

struct ESMLeuning{FT}

An empirical model parameter set type for Leuning type model. The equation used for Leuning type model is

\[gs = g0 + g1 ⋅ \dfrac{A}{Cs - Γ^{*}} ⋅ \dfrac{1}{1 + \dfrac{VPD}{d0}}\]

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [unitless]

d0

fitting parameter of d/d0 below the fraction, same unit as vpd [Pa]

Land.Photosynthesis.ESMMedlyn — Type

struct ESMMedlyn{FT}

An empirical model parameter set type for Medlyn type model. The equation used in Medlyn type model is

gs = g0 + (1 + \dfrac{g1}{\sqrt{VPD}} ⋅ \dfrac{A}{Ca}

Fields

g0

minimal stomatal conductance g0 [mol m⁻² s⁻¹]

g1

slope of conductance-photosynthesis correlation [Pa⁽⁵⁾]

Optimization Stomtal Model

Land.Photosynthesis.OSMEller — Type

struct OSMEller

An optimization model parameter set type for Eller model. The equation used for Eller model is

\[\dfrac{∂Θ}{∂E} = -\dfrac{∂K}{∂E} ⋅ \dfrac{A}{K}\]

where K is $\dfrac{∂E}{∂P}$.

Fields

Land.Photosynthesis.OSMSperry — Type

struct OSMSperry

An optimization model parameter set type for Sperry model. The equation used for Sperry model is

\[\dfrac{∂Θ}{∂E} = -\dfrac{∂K}{∂E} ⋅ \dfrac{A_{max}}{K_{max}}\]

where K is $\dfrac{∂E}{∂P}$.

Fields

Land.Photosynthesis.OSMWang — Type

struct OSMWang

An optimization model parameter set type for Eller type model. The equation used for Wang model is

\[\dfrac{∂Θ}{∂E} = \dfrac{A}{E_{crit} - E}\]

Fields

Land.Photosynthesis.OSMWAP — Type

struct OSMWAP{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model. The equation used for Wolf-Anderegg-Pacala model is

\[\dfrac{∂Θ}{∂E} = \dfrac{2aP + b}{K}\]

where K is ∂P/∂E.

Fields

a

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻²]

b

Quadratic equation parameter [μmol m⁻² s⁻¹ MPa⁻¹]

Land.Photosynthesis.OSMWAPMod — Type

struct OSMWAPMod{FT}

An optimization model parameter set type for Wolf-Anderegg-Pacala type model, modified by adding a photosynthesis component while set b and c = 0. The equation used for modified Wolf-Anderegg-Pacala model is

\[\dfrac{∂Θ}{∂E} = \dfrac{aAP}{K}\]

where P is absolute value of leaf xylem pressure.

Fields

a

Quadratic equation parameter [mol mol⁻¹ MPa⁻¹]

Parameter Sets Collection

Types

Land.Photosynthesis.AbstractPhotoModelParaSet — Type

type AbstractPhotoModelParaSet

Hierarchy of the AbstractPhotoModelParaSet:

C3ParaSet

C4ParaSet

Land.Photosynthesis.C3ParaSet — Type

struct C3Paraset{FT}

Parameter sets for C3 photosynthesis.

Fields

JT

Jmax temperature dependency

KcT

Kc temperature dependency

KoT

Ko temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

ΓsT

Γ_star temperature dependency

Col

Colimitation mode

Flu

Fluorescence model

Sto

Stomatal model scheme

VR

Vcmax25 and respiration correlation

Eff_1

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

Eff_2

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

Land.Photosynthesis.C4ParaSet — Type

struct C4ParaSet{FT}

Parameter sets for C3 photosynthesis.

Fields

KpT

Kpep temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

VpT

Vpmax temperature dependency

Col

Colimitation mode

Flu

Fluorescence model

Sto

Stomatal model scheme

VR

Vcmax25 and respiration correlation

Parameter sets

Land.Photosynthesis.C3Bernacchi — Function

C3ParaSet type C3 photosynthesis using Bernacchi's data

Land.Photosynthesis.C3CLM — Function

C3ParaSet type C3 photosynthesis using CLM5's data

Land.Photosynthesis.C4CLM — Function

C4ParaSet type C4 photosynthesis using CLM5's data

Leaf and Environment

Land.Photosynthesis.Leaf — Type

struct Leaf{FT}

Struct to store leaf information.

Fields

H

Sensible Heat Flux [W m⁻²]

LE

Latent Heat Flux [W m⁻²]

LV

Latent Heat of evaporation [J mol⁻¹]

Rn

Net Radiation Balance [W m⁻²]

T

Temperature [K]

T_old

Old temperature [K]

width

Leaf width [m]

Kd

Rate constant for thermal dissipation

Kf

Rate constant for fluorescence (const)

Kn

NPQ rate constant (initially zero)

Kp

Rate constant for photochemistry (all reaction centers open)

maxPSII

max PSII yield (Kn=0, all RC open)

PSII_frac

Fraction of absorbed light used by PSII ETR

g_bc

Boundary layer conductance to CO₂ [mol m⁻² s⁻¹]

g_bh

Boundary layer conductance to heat [mol m⁻² s⁻¹]

g_bw

Boundary layer conductance to H₂O [mol m⁻² s⁻¹]

g_ias_c

Gias correction constant

g_ias_e

Gias correction exponent

g_lc

Leaf diffusive conductance to water CO₂ [mol m⁻² s⁻¹]

g_lw

Leaf diffusive conductance to water H₂O [mol m⁻² s⁻¹]

g_m

Mesophyll conductance for CO₂ [mol m⁻² s⁻¹]

g_max

Maximal leaf diffusive conductance [mol m⁻² s⁻¹]

g_max25

Maximal leaf diffusive conductance at 298.15 K [mol m⁻² s⁻¹]

g_min

Minimal leaf diffusive conductance [mol m⁻² s⁻¹]

g_min25

Minimal leaf diffusive conductance at 298.15 K [mol m⁻² s⁻¹]

g_sc

Stomatal conductance to water CO₂ [mol m⁻² s⁻¹]

g_sw

Stomatal conductance to water H₂O [mol m⁻² s⁻¹]

p_i

Leaf internal CO₂ partial pressure [Pa]

p_s

Leaf surface CO₂ partial pressure [Pa]

p_sat

Leaf saturation vapor pressure [Pa]

Ac

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

Aj

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

Ag

Gross photosynthetic rate [μmol m⁻² s⁻¹]

An

Net photosynthetic rate [μmol m⁻² s⁻¹]

Ap

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

J

Electron transport [μmol m⁻² s⁻¹]

J_pot

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

Jmax

Maximal electron transport rate [μmol m⁻² s⁻¹]

Jmax25

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

Kc

RubisCO coefficient Kc [Pa]

Ko

RubisCO coefficient Ko [Pa]

Kpep

PEP coefficient Ko [Pa]

Km

Michaelis-Menten's coefficient [Pa]

Rd

Respiration rate [μmol m⁻² s⁻¹]

Rd25

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

Vcmax

Maximal carboxylation rate [μmol m⁻² s⁻¹]

Vcmax25

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

Vpmax

Maximal PEP carboxylation rate [μmol m⁻² s⁻¹]

Vpmax25

Maximal PEP carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

Γ_star

CO₂ compensation point with the absence of Rd [Pa]

CO₂_per_electron

Total efficiency, incl. photorespiration [mol CO₂ mol⁻¹ e-]

Fm

dark adapted yield (Kp=0)

Fm′

light adapted yield (Kp=0)

Fo

dark-adapted fluorescence yield (Kp=max)

Fo′

light-adapted fluorescence yield in the dark (Kp=max)

Ja

Actual electron transport rate [μmol m⁻² s⁻¹]

NPQ

Non-Photochemical quenching

qQ

Photochemical quenching

qE

energy quenching

φ

PSII yield

ϕs

Steady-state (light-adapted) yield (aka Fs)

APAR

Absorbed photosynthetic active radiation [μmol m⁻² s⁻¹]

hs

Leaf hydraulic system

a_max

Maximal photosynthetic rate [μmol m⁻² s⁻¹]

e

Flow rate [mol m⁻² s⁻¹]

ec

Critical flow rate [mol m⁻² s⁻¹]

kr_max

Maximal hydraulic conductance ratio

p_ups

Base xylem pressre [MPa]

p_old

Base xylem pressre memory [MPa]

Land.Photosynthesis.AirLayer — Type

struct AirLayer{FT}

Struct to store environmental conditions in each air layer corresponds to one canopy layer.

Fields

t_air

Air temperature [K]

p_a

Atmospheric CO₂ partial pressure [Pa]

p_atm

Atmospheric pressure [Pa]

p_H₂O

Atmospheric vapor pressure [Pa]

p_O₂

Atmospheric O₂ partial pressure [Pa]

p_sat

Saturation vapor pressure [Pa]

RH

Relatiev humidity

Photosynthesis Model

Temperature corrections

Land.Photosynthesis.arrhenius_correction — Function

arrhenius_correction(td_set::AbstractTDParameterSet{FT}, T::FT)

A correction factor based on arrhenius's fitting procedure, given

td_set ArrheniusTD or ArrheniusPeakTD type struct

T Leaf temperature in [K]

The equation used for ArrheniusTD is

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right)\]

The equations used for ArrheniusPeakTD are

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Land.Photosynthesis.photo_TD_from_set — Function

photo_TD_from_set(td_set::ArrheniusTD{FT}, T::FT)

Make temperature correction from parameter set, given

td_set ArrheniusTD type parameter set, which has a VAL_25 field

T Leaf temperature

Useful for Kc, Ko, Kpep, and $Γ^{*}$.

Land.Photosynthesis.photo_TD_from_val — Function

photo_TD_from_val(td_set::AbstractTDParameterSet, val::FT, T::FT)

Make temperature correction from a given value, given

td_set ArrheniusTD or ArrheniusPeakTD type struct

val Uncorrected value at 298.15 K

T Leaf temperature

Useful for Vcmax, Vomac, Vpmax, Jmax, and Respiration.

Land.Photosynthesis.leaf_jmax! — Function

leaf_jmax!(td_set::AbstractTDParameterSet, leaf::Leaf{FT})

Update maximal electron transport rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_kc! — Function

leaf_kc!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update Kc at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_km! — Function

leaf_km!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT})

Update Ko at leaf temperature, given

photo_set C3ParaSet type photosynthesis parameter set

leaf Leaf type struct

envir AirLayer type struct

Land.Photosynthesis.leaf_ko! — Function

leaf_ko!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update Ko at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_kpep! — Function

leaf_kpep!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update Kpep at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_rd! — Function

leaf_rd!(td_set::AbstractTDParameterSet, leaf::Leaf)

Update leaf dark respiration rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_vcmax! — Function

leaf_vcmax!(td_set::AbstractTDParameterSet, leaf::Leaf)

Update leaf maximal carboxylation rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_vpmax! — Function

leaf_vpmax!(td_set::AbstractTDParameterSet, leaf::Leaf)

Update leaf maximal PEP carboxylation rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Land.Photosynthesis.leaf_Γstar! — Function

leaf_Γstar!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update $Γ^{*}$ at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthetic rates (individual)

Land.Photosynthesis.rubisco_limited_rate! — Function

rubisco_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
rubisco_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Calculate the RubisCO limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Land.Photosynthesis.rubisco_limited_rate_glc! — Function

rubisco_limited_rate_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Calculate the RubisCO limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Land.Photosynthesis.leaf_ETR! — Function

leaf_ETR!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
leaf_ETR!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Update the electron transport variables in the leaf struct, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Land.Photosynthesis.light_limited_rate! — Function

light_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
light_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Calculate the Light limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Land.Photosynthesis.light_limited_rate_glc! — Function

light_limited_rate_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Calculate the Light limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Land.Photosynthesis.product_limited_rate! — Function

product_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
product_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Calculate the Product limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Land.Photosynthesis.product_limited_rate_glc! — Function

product_limited_rate_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Calculate the Product limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Missing docstring.

Missing docstring for Land.Photosynthesis.photosynthesis_colimit. Check Documenter's build log for details.

Land.Photosynthesis.leaf_fluorescence! — Function

leaf_fluorescence!(fluo_set::FluoParaSet{FT}, leaf::Leaf{FT})

Compute fluorescence yield, Kn, and Kp for leaf, given

fluo_set FluoParaSet type parameter set

leaf Leaf struct

Photosynthetic rates

Land.Photosynthesis.leaf_temperature_dependence! — Function

leaf_temperature_dependence!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::AbstractLeaf,
        envir::AirLayer{FT})

Update the temperature dependent photosynthesis only, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

Land.Photosynthesis.leaf_photo_from_pi! — Function

leaf_photo_from_pi!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::AbstractLeaf,
        envir::AirLayer{FT})

Compute leaf photosynthetic rates, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

The C3 photosynthesis model is from Farquhar et al. (1980) "A biochemical model of photosynthetic CO₂ assimilation in leaves of C3 species."

The C4 photosynthesis model is adapted from Collatz et al. (1992) "Coupled photosynthesis-stomatal conductance model for leaves of C4 plants."

Land.Photosynthesis.leaf_photo_from_glc! — Function

leaf_photo_from_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Update leaf photosynthetic rates from a known leaf diffusive conductance, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

Stomatal Response to Environment

Land.Photosynthesis.empirical_gsw_from_model — Function

empirical_gsw_from_model(
        model::EmpiricalStomatalModel,
        leaf::Leaf{FT},
        envir::AirLayer{FT},
        β::FT)

Steady state gsw from empirical approach given

model EmpiricalStomatalModel type empirical model parameter set

leaf Leaf type struct

envir AirLayer type struct

β Correction factor over the g1 part of an empirical model

Land.Photosynthesis.envir_diff! — Function

envir_diff!(
        x::FT,
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT},
        sm::AbstractStomatalModel)

Calculate the difference to be minimized for a given

x Assumed leaf diffusive conductance

photo_setC3ParaSet or C4ParaSet type parameter set

leafLeaf type struct

envirAirLayer type struct

sm Stomatal model option (photo_set.Sto)

Land.Photosynthesis.leaf_photo_from_envir! — Function

leaf_photo_from_envir!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT},
        sm::AbstractStomatalModel)

Calculate steady state gsw and photosynthesis from empirical approach, given

photo_set C3ParaSet or C4ParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

sm EmpiricalStomatalModel or OptimizationStomatalModel

Note that sm is stored in photo_set.Sto

Land.Photosynthesis.leaf_gsw_control! — Function

leaf_gsw_control!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

make sure g_sw is in its physiological range limited by diffusion, given

photo_set C3ParaSet or C4ParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

Land.Photosynthesis.leaf_heat_flux! — Function

leaf_heat_flux!(leaf::Leaf{FT}, envir::AirLayer{FT})

Update leaf sensible and latent fluexes, given

leaf Leaf type struct

envir AirLayer type struct

The sensible heat flux is computed using

\[H = 2 ⋅ ρ ⋅ c_p ⋅ g_b ⋅ (T - T_{air})\]

where 2 describes the two-sided leaf, ρ is the density of dry air ($ρ = \dfrac{p_{atm} ⋅ M_{dryair}}{R ⋅ T_{air}}$, $M_{dryair}$ is the molar mass of dry air and $R$ is the gas constant), $c_p$ is heat capacity of dry air at constant pressure, $g_b$ is the leaf boundary layer conductance for heat transfer in [m s⁻¹] ([mol m⁻² s⁻¹] times $\dfrac{RT_{air}}{P_{atm}}$ gives [m s⁻¹] ), $T$ is leaf temperature, and $T_{air}$ is air temperature. Thus, the formulation is simplified to

\[H = 2 ⋅ c_p ⋅ g_{bw} ⋅ (T - T_{air}) ⋅ M_{dryair}\]

The latent heat flux is computed using

\[LE = LV ⋅ E\]

where $LV$ is specific latent heat of evporation in [J mol⁻¹], and $E$ is evaporation rate in [mol m⁻² s⁻¹].