Utils for the LAND model

LandParameters module

Land.LandParameters.EarthParameterSet — Type

Define a local struct inherited from AbstractEarthParameterSet

Land.LandParameters.EARTH — Constant

Earth as EarthParameterSet

Land.LandParameters.AVOGADRO — Constant

Avogadro's number

Land.LandParameters.CP_D — Constant

Isobaric specific heat of dry air [J kg⁻¹ K⁻¹]

Land.LandParameters.CP_I — Constant

Isobaric specific heat of ice [J kg⁻¹ K⁻¹]

Land.LandParameters.CP_L — Constant

Isobaric specific heat of liquid water [J kg⁻¹ K⁻¹]

Land.LandParameters.GAS_R — Constant

Universal gas constant [J K⁻¹ mol⁻¹]

Land.LandParameters.GFT — Type

Global Floating Type for Land module

Land.LandParameters.GRAVITY — Constant

Gravitational acceleration [m s⁻²]

Land.LandParameters.H_PLANCK — Constant

Planck constant [m² kg s⁻¹]

Land.LandParameters.K_0 — Constant

Temperature at 0 Celcius [K]

Land.LandParameters.K_25 — Constant

Temperature at 25 Celcius [K]

Land.LandParameters.K_BOLTZMANN — Constant

Stefan-Boltzmann constant (W m⁻² K⁻⁴)

Land.LandParameters.LH_S0 — Constant

Latent heat sublimation at $T_{triple}$ [J kg⁻¹]

Land.LandParameters.LH_V0 — Constant

Latent heat vaporization at $T_{triple}$ [J kg⁻¹]

Land.LandParameters.LIGHT_SPEED — Constant

Light speed [m s⁻¹]

Land.LandParameters.PRESS_TRIPLE — Constant

Triple point vapor pressure [Pa]

Land.LandParameters.R_D — Constant

Ideal gas constant for dry air [J kg⁻¹ K⁻¹]

Land.LandParameters.R_V — Constant

Ideal gas constant for water vapor [J kg⁻¹ K⁻¹]

Land.LandParameters.T_TRIPLE — Constant

Triple point temperature [K]

Land.LandParameters.VON_KARMAN_CONST — Constant

von Karman constant

Land.LandParameters.WATER_AIR_MRATIO — Constant

Ratio of molecular mass of water and dry air

Land.LandParameters.ρ_H₂O — Constant

Density of water [kg m⁻³]

MathTools module

Land.MathTools.e2phot — Function

e2phot(λ::Array,E::Array)

Calculates the number of moles of photons, given

λ An array of wave length in [nm], converted to [m] by _FAC

E Joules of energy

Land.MathTools.fast∫ — Function

fast∫(dx::Vector, f::Vector)

A fast way of integrating functions, given

dx Delta x for each x

f f(x) for each x

Land.MathTools.lower_quadratic — Function

lower_quadratic

Return the lower quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

Land.MathTools.quadratic — Function

quadratic(a::FT, b::FT, c::FT)

Return a list of quadratic function solutions or an error message, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

Land.MathTools.volscatt — Function

volscatt(tts::FT, tto::FT, psi::FT, ttl::FT)

Calculate interception parameters (chi_s and chi_s) and leaf reflectance multiplier (frho) and transmittance multiplier (ftau), given

tts Solar zenith angle

tto Viewing zenith angle

psi Azimuth angle

ttl Leaf inclination angle

WaterPhysics module

Land.WaterPhysics.latent_heat_vapor — Function

latent_heat_vapor(T::FT)
latent_heat_vapor(T::Array)

The specific latent heat of vaporization, given

T (Array of) temperature

Land.WaterPhysics.saturation_vapor_pressure — Function

saturation_vapor_pressure(T::FT)
saturation_vapor_pressure(T::Array)

Return the saturation vapor pressure over a plane liquid surface given

T (Array of) temperature

Land.WaterPhysics.saturation_vapor_pressure_slope — Function

saturation_vapor_pressure_slope(T::FT)
saturation_vapor_pressure_slope(T::Array)

Return the 1st order derivative of saturation vapor pressure over a plane liquid surface given

T (Array of) temperature

Compute the the 1st order derivative of saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The re-arranged Clausius-Clapeyron relation: dp_v_sat/dT = [LH_0 + Δcp * (T-T_triple)]/(R_v*T^2) * p_v_sat.

Land.WaterPhysics.surface_tension — Function

surface_tension(T::FT)
surface_tension(T::Array)

Surface tension [N m⁻¹] of water against air, given

T (Array of) temperature

The equation used is γ = 0.2358 * (1 - T/T_c)^1.256 * [1 - 0.625 * (1 - T/T_c)] See http://www.iapws.org/relguide/Surf-H2O.html

Land.WaterPhysics.relative_surface_tension — Function

relative_surface_tension(T::FT)
relative_surface_tension(T::Array)

Relative surface tension of water against air relative to 298.15 K, given

T (Array of) temperature

The equation used is γ = 0.2358 * (1 - T/T_c)^1.256 * [1 - 0.625 * (1 - T/T_c)]. See http://www.iapws.org/relguide/Surf-H2O.html

A reference temperature at 298.15 K is used because the hydraulic conductances and vulnerability curves of plants are described at 298.15 K.

Land.WaterPhysics.viscosity — Function

viscosity(T::FT)
viscosity(T::Array)

Viscosity of water [Pa s], given

T (Array of) temperature

Equation used is υ = A * exp( B/T + C*T + D*T^2 ), and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

A = 1.856E-14 # Pa s
B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²

Land.WaterPhysics.relative_viscosity — Function

relative_viscosity(T::FT)
relative_viscosity(T::Array)

Viscosity relative to 25 degree C (298.15 K), given

T Water temperature

Equation used is υ/υ25 = exp( B/T + C*T + D*T^2 - B/tem25 - C*tem25 - D*tem25^2 ), and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²